chore: tweaked the settings a bit
