+++ docs/assets/index-8165e77e.js
@@ -1,0 +1,630 @@
+(function(){const e=document.createElement("link").relList;if(e&&e.supports&&e.supports("modulepreload"))return;for(const o of document.querySelectorAll('link[rel="modulepreload"]'))s(o);new MutationObserver(o=>{for(const l of o)if(l.type==="childList")for(const i of l.addedNodes)i.tagName==="LINK"&&i.rel==="modulepreload"&&s(i)}).observe(document,{childList:!0,subtree:!0});function n(o){const l={};return o.integrity&&(l.integrity=o.integrity),o.referrerPolicy&&(l.referrerPolicy=o.referrerPolicy),o.crossOrigin==="use-credentials"?l.credentials="include":o.crossOrigin==="anonymous"?l.credentials="omit":l.credentials="same-origin",l}function s(o){if(o.ep)return;o.ep=!0;const l=n(o);fetch(o.href,l)}})();function es(t,e){const n=Object.create(null),s=t.split(",");for(let o=0;o<s.length;o++)n[s[o]]=!0;return e?o=>!!n[o.toLowerCase()]:o=>!!n[o]}const nt={},ye=[],It=()=>{},Rl=()=>!1,Pl=/^on[^a-z]/,bn=t=>Pl.test(t),ns=t=>t.startsWith("onUpdate:"),it=Object.assign,ss=(t,e)=>{const n=t.indexOf(e);n>-1&&t.splice(n,1)},Sl=Object.prototype.hasOwnProperty,z=(t,e)=>Sl.call(t,e),N=Array.isArray,we=t=>vn(t)==="[object Map]",Eo=t=>vn(t)==="[object Set]",D=t=>typeof t=="function",ct=t=>typeof t=="string",os=t=>typeof t=="symbol",st=t=>t!==null&&typeof t=="object",Ao=t=>st(t)&&D(t.then)&&D(t.catch),Co=Object.prototype.toString,vn=t=>Co.call(t),Ol=t=>vn(t).slice(8,-1),Io=t=>vn(t)==="[object Object]",ls=t=>ct(t)&&t!=="NaN"&&t[0]!=="-"&&""+parseInt(t,10)===t,on=es(",key,ref,ref_for,ref_key,onVnodeBeforeMount,onVnodeMounted,onVnodeBeforeUpdate,onVnodeUpdated,onVnodeBeforeUnmount,onVnodeUnmounted"),yn=t=>{const e=Object.create(null);return n=>e[n]||(e[n]=t(n))},$l=/-(\w)/g,Lt=yn(t=>t.replace($l,(e,n)=>n?n.toUpperCase():"")),ql=/\B([A-Z])/g,Be=yn(t=>t.replace(ql,"-$1").toLowerCase()),wn=yn(t=>t.charAt(0).toUpperCase()+t.slice(1)),Rn=yn(t=>t?`on${wn(t)}`:""),De=(t,e)=>!Object.is(t,e),Pn=(t,e)=>{for(let n=0;n<t.length;n++)t[n](e)},un=(t,e,n)=>{Object.defineProperty(t,e,{configurable:!0,enumerable:!1,value:n})},Ml=t=>{const e=parseFloat(t);return isNaN(e)?t:e};let Ts;const Ln=()=>Ts||(Ts=typeof globalThis<"u"?globalThis:typeof self<"u"?self:typeof window<"u"?window:typeof global<"u"?global:{});function as(t){if(N(t)){const e={};for(let n=0;n<t.length;n++){const s=t[n],o=ct(s)?Nl(s):as(s);if(o)for(const l in o)e[l]=o[l]}return e}else{if(ct(t))return t;if(st(t))return t}}const Hl=/;(?![^(]*\))/g,Fl=/:([^]+)/,Ll=new RegExp("\\/\\*.*?\\*\\/","gs");function Nl(t){const e={};return t.replace(Ll,"").split(Hl).forEach(n=>{if(n){const s=n.split(Fl);s.length>1&&(e[s[0].trim()]=s[1].trim())}}),e}function is(t){let e="";if(ct(t))e=t;else if(N(t))for(let n=0;n<t.length;n++){const s=is(t[n]);s&&(e+=s+" ")}else if(st(t))for(const n in t)t[n]&&(e+=n+" ");return e.trim()}const jl="itemscope,allowfullscreen,formnovalidate,ismap,nomodule,novalidate,readonly",Dl=es(jl);function Bo(t){return!!t||t===""}const U=t=>ct(t)?t:t==null?"":N(t)||st(t)&&(t.toString===Co||!D(t.toString))?JSON.stringify(t,To,2):String(t),To=(t,e)=>e&&e.__v_isRef?To(t,e.value):we(e)?{[`Map(${e.size})`]:[...e.entries()].reduce((n,[s,o])=>(n[`${s} =>`]=o,n),{})}:Eo(e)?{[`Set(${e.size})`]:[...e.values()]}:st(e)&&!N(e)&&!Io(e)?String(e):e;let kt;class Ul{constructor(e=!1){this.detached=e,this._active=!0,this.effects=[],this.cleanups=[],this.parent=kt,!e&&kt&&(this.index=(kt.scopes||(kt.scopes=[])).push(this)-1)}get active(){return this._active}run(e){if(this._active){const n=kt;try{return kt=this,e()}finally{kt=n}}}on(){kt=this}off(){kt=this.parent}stop(e){if(this._active){let n,s;for(n=0,s=this.effects.length;n<s;n++)this.effects[n].stop();for(n=0,s=this.cleanups.length;n<s;n++)this.cleanups[n]();if(this.scopes)for(n=0,s=this.scopes.length;n<s;n++)this.scopes[n].stop(!0);if(!this.detached&&this.parent&&!e){const o=this.parent.scopes.pop();o&&o!==this&&(this.parent.scopes[this.index]=o,o.index=this.index)}this.parent=void 0,this._active=!1}}}function Wl(t,e=kt){e&&e.active&&e.effects.push(t)}function Kl(){return kt}const cs=t=>{const e=new Set(t);return e.w=0,e.n=0,e},Ro=t=>(t.w&ne)>0,Po=t=>(t.n&ne)>0,zl=({deps:t})=>{if(t.length)for(let e=0;e<t.length;e++)t[e].w|=ne},Yl=t=>{const{deps:e}=t;if(e.length){let n=0;for(let s=0;s<e.length;s++){const o=e[s];Ro(o)&&!Po(o)?o.delete(t):e[n++]=o,o.w&=~ne,o.n&=~ne}e.length=n}},Nn=new WeakMap;let Me=0,ne=1;const jn=30;let Et;const ce=Symbol(""),Dn=Symbol("");class rs{constructor(e,n=null,s){this.fn=e,this.scheduler=n,this.active=!0,this.deps=[],this.parent=void 0,Wl(this,s)}run(){if(!this.active)return this.fn();let e=Et,n=te;for(;e;){if(e===this)return;e=e.parent}try{return this.parent=Et,Et=this,te=!0,ne=1<<++Me,Me<=jn?zl(this):Rs(this),this.fn()}finally{Me<=jn&&Yl(this),ne=1<<--Me,Et=this.parent,te=n,this.parent=void 0,this.deferStop&&this.stop()}}stop(){Et===this?this.deferStop=!0:this.active&&(Rs(this),this.onStop&&this.onStop(),this.active=!1)}}function Rs(t){const{deps:e}=t;if(e.length){for(let n=0;n<e.length;n++)e[n].delete(t);e.length=0}}let te=!0;const So=[];function Te(){So.push(te),te=!1}function Re(){const t=So.pop();te=t===void 0?!0:t}function vt(t,e,n){if(te&&Et){let s=Nn.get(t);s||Nn.set(t,s=new Map);let o=s.get(n);o||s.set(n,o=cs()),Oo(o)}}function Oo(t,e){let n=!1;Me<=jn?Po(t)||(t.n|=ne,n=!Ro(t)):n=!t.has(Et),n&&(t.add(Et),Et.deps.push(t))}function Wt(t,e,n,s,o,l){const i=Nn.get(t);if(!i)return;let d=[];if(e==="clear")d=[...i.values()];else if(n==="length"&&N(t)){const c=Number(s);i.forEach((h,f)=>{(f==="length"||f>=c)&&d.push(h)})}else switch(n!==void 0&&d.push(i.get(n)),e){case"add":N(t)?ls(n)&&d.push(i.get("length")):(d.push(i.get(ce)),we(t)&&d.push(i.get(Dn)));break;case"delete":N(t)||(d.push(i.get(ce)),we(t)&&d.push(i.get(Dn)));break;case"set":we(t)&&d.push(i.get(ce));break}if(d.length===1)d[0]&&Un(d[0]);else{const c=[];for(const h of d)h&&c.push(...h);Un(cs(c))}}function Un(t,e){const n=N(t)?t:[...t];for(const s of n)s.computed&&Ps(s);for(const s of n)s.computed||Ps(s)}function Ps(t,e){(t!==Et||t.allowRecurse)&&(t.scheduler?t.scheduler():t.run())}const Xl=es("__proto__,__v_isRef,__isVue"),$o=new Set(Object.getOwnPropertyNames(Symbol).filter(t=>t!=="arguments"&&t!=="caller").map(t=>Symbol[t]).filter(os)),Jl=ds(),Vl=ds(!1,!0),Ql=ds(!0),Ss=Gl();function Gl(){const t={};return["includes","indexOf","lastIndexOf"].forEach(e=>{t[e]=function(...n){const s=J(this);for(let l=0,i=this.length;l<i;l++)vt(s,"get",l+"");const o=s[e](...n);return o===-1||o===!1?s[e](...n.map(J)):o}}),["push","pop","shift","unshift","splice"].forEach(e=>{t[e]=function(...n){Te();const s=J(this)[e].apply(this,n);return Re(),s}}),t}function Zl(t){const e=J(this);return vt(e,"has",t),e.hasOwnProperty(t)}function ds(t=!1,e=!1){return function(s,o,l){if(o==="__v_isReactive")return!t;if(o==="__v_isReadonly")return t;if(o==="__v_isShallow")return e;if(o==="__v_raw"&&l===(t?e?ma:Lo:e?Fo:Ho).get(s))return s;const i=N(s);if(!t){if(i&&z(Ss,o))return Reflect.get(Ss,o,l);if(o==="hasOwnProperty")return Zl}const d=Reflect.get(s,o,l);return(os(o)?$o.has(o):Xl(o))||(t||vt(s,"get",o),e)?d:ft(d)?i&&ls(o)?d:d.value:st(d)?t?No(d):Ve(d):d}}const ta=qo(),ea=qo(!0);function qo(t=!1){return function(n,s,o,l){let i=n[s];if(ke(i)&&ft(i)&&!ft(o))return!1;if(!t&&(!hn(o)&&!ke(o)&&(i=J(i),o=J(o)),!N(n)&&ft(i)&&!ft(o)))return i.value=o,!0;const d=N(n)&&ls(s)?Number(s)<n.length:z(n,s),c=Reflect.set(n,s,o,l);return n===J(l)&&(d?De(o,i)&&Wt(n,"set",s,o):Wt(n,"add",s,o)),c}}function na(t,e){const n=z(t,e);t[e];const s=Reflect.deleteProperty(t,e);return s&&n&&Wt(t,"delete",e,void 0),s}function sa(t,e){const n=Reflect.has(t,e);return(!os(e)||!$o.has(e))&&vt(t,"has",e),n}function oa(t){return vt(t,"iterate",N(t)?"length":ce),Reflect.ownKeys(t)}const Mo={get:Jl,set:ta,deleteProperty:na,has:sa,ownKeys:oa},la={get:Ql,set(t,e){return!0},deleteProperty(t,e){return!0}},aa=it({},Mo,{get:Vl,set:ea}),us=t=>t,_n=t=>Reflect.getPrototypeOf(t);function Ge(t,e,n=!1,s=!1){t=t.__v_raw;const o=J(t),l=J(e);n||(e!==l&&vt(o,"get",e),vt(o,"get",l));const{has:i}=_n(o),d=s?us:n?ps:Ue;if(i.call(o,e))return d(t.get(e));if(i.call(o,l))return d(t.get(l));t!==o&&t.get(e)}function Ze(t,e=!1){const n=this.__v_raw,s=J(n),o=J(t);return e||(t!==o&&vt(s,"has",t),vt(s,"has",o)),t===o?n.has(t):n.has(t)||n.has(o)}function tn(t,e=!1){return t=t.__v_raw,!e&&vt(J(t),"iterate",ce),Reflect.get(t,"size",t)}function Os(t){t=J(t);const e=J(this);return _n(e).has.call(e,t)||(e.add(t),Wt(e,"add",t,t)),this}function $s(t,e){e=J(e);const n=J(this),{has:s,get:o}=_n(n);let l=s.call(n,t);l||(t=J(t),l=s.call(n,t));const i=o.call(n,t);return n.set(t,e),l?De(e,i)&&Wt(n,"set",t,e):Wt(n,"add",t,e),this}function qs(t){const e=J(this),{has:n,get:s}=_n(e);let o=n.call(e,t);o||(t=J(t),o=n.call(e,t)),s&&s.call(e,t);const l=e.delete(t);return o&&Wt(e,"delete",t,void 0),l}function Ms(){const t=J(this),e=t.size!==0,n=t.clear();return e&&Wt(t,"clear",void 0,void 0),n}function en(t,e){return function(s,o){const l=this,i=l.__v_raw,d=J(i),c=e?us:t?ps:Ue;return!t&&vt(d,"iterate",ce),i.forEach((h,f)=>s.call(o,c(h),c(f),l))}}function nn(t,e,n){return function(...s){const o=this.__v_raw,l=J(o),i=we(l),d=t==="entries"||t===Symbol.iterator&&i,c=t==="keys"&&i,h=o[t](...s),f=n?us:e?ps:Ue;return!e&&vt(l,"iterate",c?Dn:ce),{next(){const{value:m,done:g}=h.next();return g?{value:m,done:g}:{value:d?[f(m[0]),f(m[1])]:f(m),done:g}},[Symbol.iterator](){return this}}}}function Vt(t){return function(...e){return t==="delete"?!1:this}}function ia(){const t={get(l){return Ge(this,l)},get size(){return tn(this)},has:Ze,add:Os,set:$s,delete:qs,clear:Ms,forEach:en(!1,!1)},e={get(l){return Ge(this,l,!1,!0)},get size(){return tn(this)},has:Ze,add:Os,set:$s,delete:qs,clear:Ms,forEach:en(!1,!0)},n={get(l){return Ge(this,l,!0)},get size(){return tn(this,!0)},has(l){return Ze.call(this,l,!0)},add:Vt("add"),set:Vt("set"),delete:Vt("delete"),clear:Vt("clear"),forEach:en(!0,!1)},s={get(l){return Ge(this,l,!0,!0)},get size(){return tn(this,!0)},has(l){return Ze.call(this,l,!0)},add:Vt("add"),set:Vt("set"),delete:Vt("delete"),clear:Vt("clear"),forEach:en(!0,!0)};return["keys","values","entries",Symbol.iterator].forEach(l=>{t[l]=nn(l,!1,!1),n[l]=nn(l,!0,!1),e[l]=nn(l,!1,!0),s[l]=nn(l,!0,!0)}),[t,n,e,s]}const[ca,ra,da,ua]=ia();function hs(t,e){const n=e?t?ua:da:t?ra:ca;return(s,o,l)=>o==="__v_isReactive"?!t:o==="__v_isReadonly"?t:o==="__v_raw"?s:Reflect.get(z(n,o)&&o in s?n:s,o,l)}const ha={get:hs(!1,!1)},fa={get:hs(!1,!0)},pa={get:hs(!0,!1)},Ho=new WeakMap,Fo=new WeakMap,Lo=new WeakMap,ma=new WeakMap;function ga(t){switch(t){case"Object":case"Array":return 1;case"Map":case"Set":case"WeakMap":case"WeakSet":return 2;default:return 0}}function ba(t){return t.__v_skip||!Object.isExtensible(t)?0:ga(Ol(t))}function Ve(t){return ke(t)?t:fs(t,!1,Mo,ha,Ho)}function va(t){return fs(t,!1,aa,fa,Fo)}function No(t){return fs(t,!0,la,pa,Lo)}function fs(t,e,n,s,o){if(!st(t)||t.__v_raw&&!(e&&t.__v_isReactive))return t;const l=o.get(t);if(l)return l;const i=ba(t);if(i===0)return t;const d=new Proxy(t,i===2?s:n);return o.set(t,d),d}function _e(t){return ke(t)?_e(t.__v_raw):!!(t&&t.__v_isReactive)}function ke(t){return!!(t&&t.__v_isReadonly)}function hn(t){return!!(t&&t.__v_isShallow)}function jo(t){return _e(t)||ke(t)}function J(t){const e=t&&t.__v_raw;return e?J(e):t}function Do(t){return un(t,"__v_skip",!0),t}const Ue=t=>st(t)?Ve(t):t,ps=t=>st(t)?No(t):t;function Uo(t){te&&Et&&(t=J(t),Oo(t.dep||(t.dep=cs())))}function Wo(t,e){t=J(t);const n=t.dep;n&&Un(n)}function ft(t){return!!(t&&t.__v_isRef===!0)}function ya(t){return Ko(t,!1)}function wa(t){return Ko(t,!0)}function Ko(t,e){return ft(t)?t:new _a(t,e)}class _a{constructor(e,n){this.__v_isShallow=n,this.dep=void 0,this.__v_isRef=!0,this._rawValue=n?e:J(e),this._value=n?e:Ue(e)}get value(){return Uo(this),this._value}set value(e){const n=this.__v_isShallow||hn(e)||ke(e);e=n?e:J(e),De(e,this._rawValue)&&(this._rawValue=e,this._value=n?e:Ue(e),Wo(this))}}function L(t){return ft(t)?t.value:t}const xa={get:(t,e,n)=>L(Reflect.get(t,e,n)),set:(t,e,n,s)=>{const o=t[e];return ft(o)&&!ft(n)?(o.value=n,!0):Reflect.set(t,e,n,s)}};function zo(t){return _e(t)?t:new Proxy(t,xa)}class ka{constructor(e,n,s,o){this._setter=n,this.dep=void 0,this.__v_isRef=!0,this.__v_isReadonly=!1,this._dirty=!0,this.effect=new rs(e,()=>{this._dirty||(this._dirty=!0,Wo(this))}),this.effect.computed=this,this.effect.active=this._cacheable=!o,this.__v_isReadonly=s}get value(){const e=J(this);return Uo(e),(e._dirty||!e._cacheable)&&(e._dirty=!1,e._value=e.effect.run()),e._value}set value(e){this._setter(e)}}function Ea(t,e,n=!1){let s,o;const l=D(t);return l?(s=t,o=It):(s=t.get,o=t.set),new ka(s,o,l||!o,n)}function ee(t,e,n,s){let o;try{o=s?t(...s):t()}catch(l){xn(l,e,n)}return o}function Bt(t,e,n,s){if(D(t)){const l=ee(t,e,n,s);return l&&Ao(l)&&l.catch(i=>{xn(i,e,n)}),l}const o=[];for(let l=0;l<t.length;l++)o.push(Bt(t[l],e,n,s));return o}function xn(t,e,n,s=!0){const o=e?e.vnode:null;if(e){let l=e.parent;const i=e.proxy,d=n;for(;l;){const h=l.ec;if(h){for(let f=0;f<h.length;f++)if(h[f](t,i,d)===!1)return}l=l.parent}const c=e.appContext.config.errorHandler;if(c){ee(c,null,10,[t,i,d]);return}}Aa(t,n,o,s)}function Aa(t,e,n,s=!0){console.error(t)}let We=!1,Wn=!1;const ht=[];let Ft=0;const xe=[];let Ut=null,ae=0;const Yo=Promise.resolve();let ms=null;function Xo(t){const e=ms||Yo;return t?e.then(this?t.bind(this):t):e}function Ca(t){let e=Ft+1,n=ht.length;for(;e<n;){const s=e+n>>>1;Ke(ht[s])<t?e=s+1:n=s}return e}function gs(t){(!ht.length||!ht.includes(t,We&&t.allowRecurse?Ft+1:Ft))&&(t.id==null?ht.push(t):ht.splice(Ca(t.id),0,t),Jo())}function Jo(){!We&&!Wn&&(Wn=!0,ms=Yo.then(Qo))}function Ia(t){const e=ht.indexOf(t);e>Ft&&ht.splice(e,1)}function Ba(t){N(t)?xe.push(...t):(!Ut||!Ut.includes(t,t.allowRecurse?ae+1:ae))&&xe.push(t),Jo()}function Hs(t,e=We?Ft+1:0){for(;e<ht.length;e++){const n=ht[e];n&&n.pre&&(ht.splice(e,1),e--,n())}}function Vo(t){if(xe.length){const e=[...new Set(xe)];if(xe.length=0,Ut){Ut.push(...e);return}for(Ut=e,Ut.sort((n,s)=>Ke(n)-Ke(s)),ae=0;ae<Ut.length;ae++)Ut[ae]();Ut=null,ae=0}}const Ke=t=>t.id==null?1/0:t.id,Ta=(t,e)=>{const n=Ke(t)-Ke(e);if(n===0){if(t.pre&&!e.pre)return-1;if(e.pre&&!t.pre)return 1}return n};function Qo(t){Wn=!1,We=!0,ht.sort(Ta);const e=It;try{for(Ft=0;Ft<ht.length;Ft++){const n=ht[Ft];n&&n.active!==!1&&ee(n,null,14)}}finally{Ft=0,ht.length=0,Vo(),We=!1,ms=null,(ht.length||xe.length)&&Qo()}}function Ra(t,e,...n){if(t.isUnmounted)return;const s=t.vnode.props||nt;let o=n;const l=e.startsWith("update:"),i=l&&e.slice(7);if(i&&i in s){const f=`${i==="modelValue"?"model":i}Modifiers`,{number:m,trim:g}=s[f]||nt;g&&(o=n.map(x=>ct(x)?x.trim():x)),m&&(o=n.map(Ml))}let d,c=s[d=Rn(e)]||s[d=Rn(Lt(e))];!c&&l&&(c=s[d=Rn(Be(e))]),c&&Bt(c,t,6,o);const h=s[d+"Once"];if(h){if(!t.emitted)t.emitted={};else if(t.emitted[d])return;t.emitted[d]=!0,Bt(h,t,6,o)}}function Go(t,e,n=!1){const s=e.emitsCache,o=s.get(t);if(o!==void 0)return o;const l=t.emits;let i={},d=!1;if(!D(t)){const c=h=>{const f=Go(h,e,!0);f&&(d=!0,it(i,f))};!n&&e.mixins.length&&e.mixins.forEach(c),t.extends&&c(t.extends),t.mixins&&t.mixins.forEach(c)}return!l&&!d?(st(t)&&s.set(t,null),null):(N(l)?l.forEach(c=>i[c]=null):it(i,l),st(t)&&s.set(t,i),i)}function kn(t,e){return!t||!bn(e)?!1:(e=e.slice(2).replace(/Once$/,""),z(t,e[0].toLowerCase()+e.slice(1))||z(t,Be(e))||z(t,e))}let At=null,En=null;function fn(t){const e=At;return At=t,En=t&&t.type.__scopeId||null,e}function Pa(t){En=t}function Sa(){En=null}function at(t,e=At,n){if(!e||t._n)return t;const s=(...o)=>{s._d&&Xs(-1);const l=fn(e);let i;try{i=t(...o)}finally{fn(l),s._d&&Xs(1)}return i};return s._n=!0,s._c=!0,s._d=!0,s}function Sn(t){const{type:e,vnode:n,proxy:s,withProxy:o,props:l,propsOptions:[i],slots:d,attrs:c,emit:h,render:f,renderCache:m,data:g,setupState:x,ctx:T,inheritAttrs:S}=t;let j,q;const M=fn(t);try{if(n.shapeFlag&4){const H=o||s;j=Ht(f.call(H,H,m,l,x,g,T)),q=c}else{const H=e;j=Ht(H.length>1?H(l,{attrs:c,slots:d,emit:h}):H(l,null)),q=e.props?c:Oa(c)}}catch(H){Le.length=0,xn(H,t,1),j=Y(ze)}let X=j;if(q&&S!==!1){const H=Object.keys(q),{shapeFlag:dt}=X;H.length&&dt&7&&(i&&H.some(ns)&&(q=$a(q,i)),X=Ee(X,q))}return n.dirs&&(X=Ee(X),X.dirs=X.dirs?X.dirs.concat(n.dirs):n.dirs),n.transition&&(X.transition=n.transition),j=X,fn(M),j}const Oa=t=>{let e;for(const n in t)(n==="class"||n==="style"||bn(n))&&((e||(e={}))[n]=t[n]);return e},$a=(t,e)=>{const n={};for(const s in t)(!ns(s)||!(s.slice(9)in e))&&(n[s]=t[s]);return n};function qa(t,e,n){const{props:s,children:o,component:l}=t,{props:i,children:d,patchFlag:c}=e,h=l.emitsOptions;if(e.dirs||e.transition)return!0;if(n&&c>=0){if(c&1024)return!0;if(c&16)return s?Fs(s,i,h):!!i;if(c&8){const f=e.dynamicProps;for(let m=0;m<f.length;m++){const g=f[m];if(i[g]!==s[g]&&!kn(h,g))return!0}}}else return(o||d)&&(!d||!d.$stable)?!0:s===i?!1:s?i?Fs(s,i,h):!0:!!i;return!1}function Fs(t,e,n){const s=Object.keys(e);if(s.length!==Object.keys(t).length)return!0;for(let o=0;o<s.length;o++){const l=s[o];if(e[l]!==t[l]&&!kn(n,l))return!0}return!1}function Ma({vnode:t,parent:e},n){for(;e&&e.subTree===t;)(t=e.vnode).el=n,e=e.parent}const Ha=t=>t.__isSuspense;function Fa(t,e){e&&e.pendingBranch?N(t)?e.effects.push(...t):e.effects.push(t):Ba(t)}const sn={};function ln(t,e,n){return Zo(t,e,n)}function Zo(t,e,{immediate:n,deep:s,flush:o,onTrack:l,onTrigger:i}=nt){var d;const c=Kl()===((d=rt)==null?void 0:d.scope)?rt:null;let h,f=!1,m=!1;if(ft(t)?(h=()=>t.value,f=hn(t)):_e(t)?(h=()=>t,s=!0):N(t)?(m=!0,f=t.some(H=>_e(H)||hn(H)),h=()=>t.map(H=>{if(ft(H))return H.value;if(_e(H))return ve(H);if(D(H))return ee(H,c,2)})):D(t)?e?h=()=>ee(t,c,2):h=()=>{if(!(c&&c.isUnmounted))return g&&g(),Bt(t,c,3,[x])}:h=It,e&&s){const H=h;h=()=>ve(H())}let g,x=H=>{g=M.onStop=()=>{ee(H,c,4)}},T;if(Xe)if(x=It,e?n&&Bt(e,c,3,[h(),m?[]:void 0,x]):h(),o==="sync"){const H=Pi();T=H.__watcherHandles||(H.__watcherHandles=[])}else return It;let S=m?new Array(t.length).fill(sn):sn;const j=()=>{if(M.active)if(e){const H=M.run();(s||f||(m?H.some((dt,pt)=>De(dt,S[pt])):De(H,S)))&&(g&&g(),Bt(e,c,3,[H,S===sn?void 0:m&&S[0]===sn?[]:S,x]),S=H)}else M.run()};j.allowRecurse=!!e;let q;o==="sync"?q=j:o==="post"?q=()=>bt(j,c&&c.suspense):(j.pre=!0,c&&(j.id=c.uid),q=()=>gs(j));const M=new rs(h,q);e?n?j():S=M.run():o==="post"?bt(M.run.bind(M),c&&c.suspense):M.run();const X=()=>{M.stop(),c&&c.scope&&ss(c.scope.effects,M)};return T&&T.push(X),X}function La(t,e,n){const s=this.proxy,o=ct(t)?t.includes(".")?tl(s,t):()=>s[t]:t.bind(s,s);let l;D(e)?l=e:(l=e.handler,n=e);const i=rt;Ae(this);const d=Zo(o,l.bind(s),n);return i?Ae(i):re(),d}function tl(t,e){const n=e.split(".");return()=>{let s=t;for(let o=0;o<n.length&&s;o++)s=s[n[o]];return s}}function ve(t,e){if(!st(t)||t.__v_skip||(e=e||new Set,e.has(t)))return t;if(e.add(t),ft(t))ve(t.value,e);else if(N(t))for(let n=0;n<t.length;n++)ve(t[n],e);else if(Eo(t)||we(t))t.forEach(n=>{ve(n,e)});else if(Io(t))for(const n in t)ve(t[n],e);return t}function oe(t,e,n,s){const o=t.dirs,l=e&&e.dirs;for(let i=0;i<o.length;i++){const d=o[i];l&&(d.oldValue=l[i].value);let c=d.dir[s];c&&(Te(),Bt(c,n,8,[t.el,d,t,e]),Re())}}function el(t,e){return D(t)?(()=>it({name:t.name},e,{setup:t}))():t}const an=t=>!!t.type.__asyncLoader,nl=t=>t.type.__isKeepAlive;function Na(t,e){sl(t,"a",e)}function ja(t,e){sl(t,"da",e)}function sl(t,e,n=rt){const s=t.__wdc||(t.__wdc=()=>{let o=n;for(;o;){if(o.isDeactivated)return;o=o.parent}return t()});if(An(e,s,n),n){let o=n.parent;for(;o&&o.parent;)nl(o.parent.vnode)&&Da(s,e,n,o),o=o.parent}}function Da(t,e,n,s){const o=An(e,t,s,!0);ol(()=>{ss(s[e],o)},n)}function An(t,e,n=rt,s=!1){if(n){const o=n[t]||(n[t]=[]),l=e.__weh||(e.__weh=(...i)=>{if(n.isUnmounted)return;Te(),Ae(n);const d=Bt(e,n,t,i);return re(),Re(),d});return s?o.unshift(l):o.push(l),l}}const Kt=t=>(e,n=rt)=>(!Xe||t==="sp")&&An(t,(...s)=>e(...s),n),Ua=Kt("bm"),Pe=Kt("m"),Wa=Kt("bu"),Ka=Kt("u"),za=Kt("bum"),ol=Kt("um"),Ya=Kt("sp"),Xa=Kt("rtg"),Ja=Kt("rtc");function Va(t,e=rt){An("ec",t,e)}const ll="components";function bs(t,e){return Ga(ll,t,!0,e)||t}const Qa=Symbol.for("v-ndc");function Ga(t,e,n=!0,s=!1){const o=At||rt;if(o){const l=o.type;if(t===ll){const d=Bi(l,!1);if(d&&(d===e||d===Lt(e)||d===wn(Lt(e))))return l}const i=Ls(o[t]||l[t],e)||Ls(o.appContext[t],e);return!i&&s?l:i}}function Ls(t,e){return t&&(t[e]||t[Lt(e)]||t[wn(Lt(e))])}const Kn=t=>t?ml(t)?xs(t)||t.proxy:Kn(t.parent):null,Fe=it(Object.create(null),{$:t=>t,$el:t=>t.vnode.el,$data:t=>t.data,$props:t=>t.props,$attrs:t=>t.attrs,$slots:t=>t.slots,$refs:t=>t.refs,$parent:t=>Kn(t.parent),$root:t=>Kn(t.root),$emit:t=>t.emit,$options:t=>vs(t),$forceUpdate:t=>t.f||(t.f=()=>gs(t.update)),$nextTick:t=>t.n||(t.n=Xo.bind(t.proxy)),$watch:t=>La.bind(t)}),On=(t,e)=>t!==nt&&!t.__isScriptSetup&&z(t,e),Za={get({_:t},e){const{ctx:n,setupState:s,data:o,props:l,accessCache:i,type:d,appContext:c}=t;let h;if(e[0]!=="$"){const x=i[e];if(x!==void 0)switch(x){case 1:return s[e];case 2:return o[e];case 4:return n[e];case 3:return l[e]}else{if(On(s,e))return i[e]=1,s[e];if(o!==nt&&z(o,e))return i[e]=2,o[e];if((h=t.propsOptions[0])&&z(h,e))return i[e]=3,l[e];if(n!==nt&&z(n,e))return i[e]=4,n[e];zn&&(i[e]=0)}}const f=Fe[e];let m,g;if(f)return e==="$attrs"&&vt(t,"get",e),f(t);if((m=d.__cssModules)&&(m=m[e]))return m;if(n!==nt&&z(n,e))return i[e]=4,n[e];if(g=c.config.globalProperties,z(g,e))return g[e]},set({_:t},e,n){const{data:s,setupState:o,ctx:l}=t;return On(o,e)?(o[e]=n,!0):s!==nt&&z(s,e)?(s[e]=n,!0):z(t.props,e)||e[0]==="$"&&e.slice(1)in t?!1:(l[e]=n,!0)},has({_:{data:t,setupState:e,accessCache:n,ctx:s,appContext:o,propsOptions:l}},i){let d;return!!n[i]||t!==nt&&z(t,i)||On(e,i)||(d=l[0])&&z(d,i)||z(s,i)||z(Fe,i)||z(o.config.globalProperties,i)},defineProperty(t,e,n){return n.get!=null?t._.accessCache[e]=0:z(n,"value")&&this.set(t,e,n.value,null),Reflect.defineProperty(t,e,n)}};function Ns(t){return N(t)?t.reduce((e,n)=>(e[n]=null,e),{}):t}let zn=!0;function ti(t){const e=vs(t),n=t.proxy,s=t.ctx;zn=!1,e.beforeCreate&&js(e.beforeCreate,t,"bc");const{data:o,computed:l,methods:i,watch:d,provide:c,inject:h,created:f,beforeMount:m,mounted:g,beforeUpdate:x,updated:T,activated:S,deactivated:j,beforeDestroy:q,beforeUnmount:M,destroyed:X,unmounted:H,render:dt,renderTracked:pt,renderTriggered:Rt,errorCaptured:Nt,serverPrefetch:de,expose:Pt,inheritAttrs:Xt,components:se,directives:St,filters:Se}=e;if(h&&ei(h,s,null),i)for(const tt in i){const V=i[tt];D(V)&&(s[tt]=V.bind(n))}if(o){const tt=o.call(n,n);st(tt)&&(t.data=Ve(tt))}if(zn=!0,l)for(const tt in l){const V=l[tt],jt=D(V)?V.bind(n,n):D(V.get)?V.get.bind(n,n):It,Jt=!D(V)&&D(V.set)?V.set.bind(n):It,Ot=wt({get:jt,set:Jt});Object.defineProperty(s,tt,{enumerable:!0,configurable:!0,get:()=>Ot.value,set:gt=>Ot.value=gt})}if(d)for(const tt in d)al(d[tt],s,n,tt);if(c){const tt=D(c)?c.call(n):c;Reflect.ownKeys(tt).forEach(V=>{cn(V,tt[V])})}f&&js(f,t,"c");function lt(tt,V){N(V)?V.forEach(jt=>tt(jt.bind(n))):V&&tt(V.bind(n))}if(lt(Ua,m),lt(Pe,g),lt(Wa,x),lt(Ka,T),lt(Na,S),lt(ja,j),lt(Va,Nt),lt(Ja,pt),lt(Xa,Rt),lt(za,M),lt(ol,H),lt(Ya,de),N(Pt))if(Pt.length){const tt=t.exposed||(t.exposed={});Pt.forEach(V=>{Object.defineProperty(tt,V,{get:()=>n[V],set:jt=>n[V]=jt})})}else t.exposed||(t.exposed={});dt&&t.render===It&&(t.render=dt),Xt!=null&&(t.inheritAttrs=Xt),se&&(t.components=se),St&&(t.directives=St)}function ei(t,e,n=It){N(t)&&(t=Yn(t));for(const s in t){const o=t[s];let l;st(o)?"default"in o?l=_t(o.from||s,o.default,!0):l=_t(o.from||s):l=_t(o),ft(l)?Object.defineProperty(e,s,{enumerable:!0,configurable:!0,get:()=>l.value,set:i=>l.value=i}):e[s]=l}}function js(t,e,n){Bt(N(t)?t.map(s=>s.bind(e.proxy)):t.bind(e.proxy),e,n)}function al(t,e,n,s){const o=s.includes(".")?tl(n,s):()=>n[s];if(ct(t)){const l=e[t];D(l)&&ln(o,l)}else if(D(t))ln(o,t.bind(n));else if(st(t))if(N(t))t.forEach(l=>al(l,e,n,s));else{const l=D(t.handler)?t.handler.bind(n):e[t.handler];D(l)&&ln(o,l,t)}}function vs(t){const e=t.type,{mixins:n,extends:s}=e,{mixins:o,optionsCache:l,config:{optionMergeStrategies:i}}=t.appContext,d=l.get(e);let c;return d?c=d:!o.length&&!n&&!s?c=e:(c={},o.length&&o.forEach(h=>pn(c,h,i,!0)),pn(c,e,i)),st(e)&&l.set(e,c),c}function pn(t,e,n,s=!1){const{mixins:o,extends:l}=e;l&&pn(t,l,n,!0),o&&o.forEach(i=>pn(t,i,n,!0));for(const i in e)if(!(s&&i==="expose")){const d=ni[i]||n&&n[i];t[i]=d?d(t[i],e[i]):e[i]}return t}const ni={data:Ds,props:Us,emits:Us,methods:He,computed:He,beforeCreate:mt,created:mt,beforeMount:mt,mounted:mt,beforeUpdate:mt,updated:mt,beforeDestroy:mt,beforeUnmount:mt,destroyed:mt,unmounted:mt,activated:mt,deactivated:mt,errorCaptured:mt,serverPrefetch:mt,components:He,directives:He,watch:oi,provide:Ds,inject:si};function Ds(t,e){return e?t?function(){return it(D(t)?t.call(this,this):t,D(e)?e.call(this,this):e)}:e:t}function si(t,e){return He(Yn(t),Yn(e))}function Yn(t){if(N(t)){const e={};for(let n=0;n<t.length;n++)e[t[n]]=t[n];return e}return t}function mt(t,e){return t?[...new Set([].concat(t,e))]:e}function He(t,e){return t?it(Object.create(null),t,e):e}function Us(t,e){return t?N(t)&&N(e)?[...new Set([...t,...e])]:it(Object.create(null),Ns(t),Ns(e??{})):e}function oi(t,e){if(!t)return e;if(!e)return t;const n=it(Object.create(null),t);for(const s in e)n[s]=mt(t[s],e[s]);return n}function il(){return{app:null,config:{isNativeTag:Rl,performance:!1,globalProperties:{},optionMergeStrategies:{},errorHandler:void 0,warnHandler:void 0,compilerOptions:{}},mixins:[],components:{},directives:{},provides:Object.create(null),optionsCache:new WeakMap,propsCache:new WeakMap,emitsCache:new WeakMap}}let li=0;function ai(t,e){return function(s,o=null){D(s)||(s=it({},s)),o!=null&&!st(o)&&(o=null);const l=il(),i=new Set;let d=!1;const c=l.app={_uid:li++,_component:s,_props:o,_container:null,_context:l,_instance:null,version:Si,get config(){return l.config},set config(h){},use(h,...f){return i.has(h)||(h&&D(h.install)?(i.add(h),h.install(c,...f)):D(h)&&(i.add(h),h(c,...f))),c},mixin(h){return l.mixins.includes(h)||l.mixins.push(h),c},component(h,f){return f?(l.components[h]=f,c):l.components[h]},directive(h,f){return f?(l.directives[h]=f,c):l.directives[h]},mount(h,f,m){if(!d){const g=Y(s,o);return g.appContext=l,f&&e?e(g,h):t(g,h,m),d=!0,c._container=h,h.__vue_app__=c,xs(g.component)||g.component.proxy}},unmount(){d&&(t(null,c._container),delete c._container.__vue_app__)},provide(h,f){return l.provides[h]=f,c},runWithContext(h){mn=c;try{return h()}finally{mn=null}}};return c}}let mn=null;function cn(t,e){if(rt){let n=rt.provides;const s=rt.parent&&rt.parent.provides;s===n&&(n=rt.provides=Object.create(s)),n[t]=e}}function _t(t,e,n=!1){const s=rt||At;if(s||mn){const o=s?s.parent==null?s.vnode.appContext&&s.vnode.appContext.provides:s.parent.provides:mn._context.provides;if(o&&t in o)return o[t];if(arguments.length>1)return n&&D(e)?e.call(s&&s.proxy):e}}function ii(t,e,n,s=!1){const o={},l={};un(l,In,1),t.propsDefaults=Object.create(null),cl(t,e,o,l);for(const i in t.propsOptions[0])i in o||(o[i]=void 0);n?t.props=s?o:va(o):t.type.props?t.props=o:t.props=l,t.attrs=l}function ci(t,e,n,s){const{props:o,attrs:l,vnode:{patchFlag:i}}=t,d=J(o),[c]=t.propsOptions;let h=!1;if((s||i>0)&&!(i&16)){if(i&8){const f=t.vnode.dynamicProps;for(let m=0;m<f.length;m++){let g=f[m];if(kn(t.emitsOptions,g))continue;const x=e[g];if(c)if(z(l,g))x!==l[g]&&(l[g]=x,h=!0);else{const T=Lt(g);o[T]=Xn(c,d,T,x,t,!1)}else x!==l[g]&&(l[g]=x,h=!0)}}}else{cl(t,e,o,l)&&(h=!0);let f;for(const m in d)(!e||!z(e,m)&&((f=Be(m))===m||!z(e,f)))&&(c?n&&(n[m]!==void 0||n[f]!==void 0)&&(o[m]=Xn(c,d,m,void 0,t,!0)):delete o[m]);if(l!==d)for(const m in l)(!e||!z(e,m))&&(delete l[m],h=!0)}h&&Wt(t,"set","$attrs")}function cl(t,e,n,s){const[o,l]=t.propsOptions;let i=!1,d;if(e)for(let c in e){if(on(c))continue;const h=e[c];let f;o&&z(o,f=Lt(c))?!l||!l.includes(f)?n[f]=h:(d||(d={}))[f]=h:kn(t.emitsOptions,c)||(!(c in s)||h!==s[c])&&(s[c]=h,i=!0)}if(l){const c=J(n),h=d||nt;for(let f=0;f<l.length;f++){const m=l[f];n[m]=Xn(o,c,m,h[m],t,!z(h,m))}}return i}function Xn(t,e,n,s,o,l){const i=t[n];if(i!=null){const d=z(i,"default");if(d&&s===void 0){const c=i.default;if(i.type!==Function&&!i.skipFactory&&D(c)){const{propsDefaults:h}=o;n in h?s=h[n]:(Ae(o),s=h[n]=c.call(null,e),re())}else s=c}i[0]&&(l&&!d?s=!1:i[1]&&(s===""||s===Be(n))&&(s=!0))}return s}function rl(t,e,n=!1){const s=e.propsCache,o=s.get(t);if(o)return o;const l=t.props,i={},d=[];let c=!1;if(!D(t)){const f=m=>{c=!0;const[g,x]=rl(m,e,!0);it(i,g),x&&d.push(...x)};!n&&e.mixins.length&&e.mixins.forEach(f),t.extends&&f(t.extends),t.mixins&&t.mixins.forEach(f)}if(!l&&!c)return st(t)&&s.set(t,ye),ye;if(N(l))for(let f=0;f<l.length;f++){const m=Lt(l[f]);Ws(m)&&(i[m]=nt)}else if(l)for(const f in l){const m=Lt(f);if(Ws(m)){const g=l[f],x=i[m]=N(g)||D(g)?{type:g}:it({},g);if(x){const T=Ys(Boolean,x.type),S=Ys(String,x.type);x[0]=T>-1,x[1]=S<0||T<S,(T>-1||z(x,"default"))&&d.push(m)}}}const h=[i,d];return st(t)&&s.set(t,h),h}function Ws(t){return t[0]!=="$"}function Ks(t){const e=t&&t.toString().match(/^\s*(function|class) (\w+)/);return e?e[2]:t===null?"null":""}function zs(t,e){return Ks(t)===Ks(e)}function Ys(t,e){return N(e)?e.findIndex(n=>zs(n,t)):D(e)&&zs(e,t)?0:-1}const dl=t=>t[0]==="_"||t==="$stable",ys=t=>N(t)?t.map(Ht):[Ht(t)],ri=(t,e,n)=>{if(e._n)return e;const s=at((...o)=>ys(e(...o)),n);return s._c=!1,s},ul=(t,e,n)=>{const s=t._ctx;for(const o in t){if(dl(o))continue;const l=t[o];if(D(l))e[o]=ri(o,l,s);else if(l!=null){const i=ys(l);e[o]=()=>i}}},hl=(t,e)=>{const n=ys(e);t.slots.default=()=>n},di=(t,e)=>{if(t.vnode.shapeFlag&32){const n=e._;n?(t.slots=J(e),un(e,"_",n)):ul(e,t.slots={})}else t.slots={},e&&hl(t,e);un(t.slots,In,1)},ui=(t,e,n)=>{const{vnode:s,slots:o}=t;let l=!0,i=nt;if(s.shapeFlag&32){const d=e._;d?n&&d===1?l=!1:(it(o,e),!n&&d===1&&delete o._):(l=!e.$stable,ul(e,o)),i=e}else e&&(hl(t,e),i={default:1});if(l)for(const d in o)!dl(d)&&!(d in i)&&delete o[d]};function Jn(t,e,n,s,o=!1){if(N(t)){t.forEach((g,x)=>Jn(g,e&&(N(e)?e[x]:e),n,s,o));return}if(an(s)&&!o)return;const l=s.shapeFlag&4?xs(s.component)||s.component.proxy:s.el,i=o?null:l,{i:d,r:c}=t,h=e&&e.r,f=d.refs===nt?d.refs={}:d.refs,m=d.setupState;if(h!=null&&h!==c&&(ct(h)?(f[h]=null,z(m,h)&&(m[h]=null)):ft(h)&&(h.value=null)),D(c))ee(c,d,12,[i,f]);else{const g=ct(c),x=ft(c);if(g||x){const T=()=>{if(t.f){const S=g?z(m,c)?m[c]:f[c]:c.value;o?N(S)&&ss(S,l):N(S)?S.includes(l)||S.push(l):g?(f[c]=[l],z(m,c)&&(m[c]=f[c])):(c.value=[l],t.k&&(f[t.k]=c.value))}else g?(f[c]=i,z(m,c)&&(m[c]=i)):x&&(c.value=i,t.k&&(f[t.k]=i))};i?(T.id=-1,bt(T,n)):T()}}}const bt=Fa;function hi(t){return fi(t)}function fi(t,e){const n=Ln();n.__VUE__=!0;const{insert:s,remove:o,patchProp:l,createElement:i,createText:d,createComment:c,setText:h,setElementText:f,parentNode:m,nextSibling:g,setScopeId:x=It,insertStaticContent:T}=t,S=(r,u,p,b=null,y=null,w=null,C=!1,k=null,E=!!u.dynamicChildren)=>{if(r===u)return;r&&!$e(r,u)&&(b=v(r),gt(r,y,w,!0),r=null),u.patchFlag===-2&&(E=!1,u.dynamicChildren=null);const{type:_,ref:O,shapeFlag:B}=u;switch(_){case Cn:j(r,u,p,b);break;case ze:q(r,u,p,b);break;case rn:r==null&&M(u,p,b,C);break;case Mt:se(r,u,p,b,y,w,C,k,E);break;default:B&1?dt(r,u,p,b,y,w,C,k,E):B&6?St(r,u,p,b,y,w,C,k,E):(B&64||B&128)&&_.process(r,u,p,b,y,w,C,k,E,A)}O!=null&&y&&Jn(O,r&&r.ref,w,u||r,!u)},j=(r,u,p,b)=>{if(r==null)s(u.el=d(u.children),p,b);else{const y=u.el=r.el;u.children!==r.children&&h(y,u.children)}},q=(r,u,p,b)=>{r==null?s(u.el=c(u.children||""),p,b):u.el=r.el},M=(r,u,p,b)=>{[r.el,r.anchor]=T(r.children,u,p,b,r.el,r.anchor)},X=({el:r,anchor:u},p,b)=>{let y;for(;r&&r!==u;)y=g(r),s(r,p,b),r=y;s(u,p,b)},H=({el:r,anchor:u})=>{let p;for(;r&&r!==u;)p=g(r),o(r),r=p;o(u)},dt=(r,u,p,b,y,w,C,k,E)=>{C=C||u.type==="svg",r==null?pt(u,p,b,y,w,C,k,E):de(r,u,y,w,C,k,E)},pt=(r,u,p,b,y,w,C,k)=>{let E,_;const{type:O,props:B,shapeFlag:$,transition:F,dirs:W}=r;if(E=r.el=i(r.type,w,B&&B.is,B),$&8?f(E,r.children):$&16&&Nt(r.children,E,null,b,y,w&&O!=="foreignObject",C,k),W&&oe(r,null,b,"created"),Rt(E,r,r.scopeId,C,b),B){for(const Z in B)Z!=="value"&&!on(Z)&&l(E,Z,null,B[Z],w,r.children,b,y,ut);"value"in B&&l(E,"value",null,B.value),(_=B.onVnodeBeforeMount)&&qt(_,b,r)}W&&oe(r,null,b,"beforeMount");const et=(!y||y&&!y.pendingBranch)&&F&&!F.persisted;et&&F.beforeEnter(E),s(E,u,p),((_=B&&B.onVnodeMounted)||et||W)&&bt(()=>{_&&qt(_,b,r),et&&F.enter(E),W&&oe(r,null,b,"mounted")},y)},Rt=(r,u,p,b,y)=>{if(p&&x(r,p),b)for(let w=0;w<b.length;w++)x(r,b[w]);if(y){let w=y.subTree;if(u===w){const C=y.vnode;Rt(r,C,C.scopeId,C.slotScopeIds,y.parent)}}},Nt=(r,u,p,b,y,w,C,k,E=0)=>{for(let _=E;_<r.length;_++){const O=r[_]=k?Gt(r[_]):Ht(r[_]);S(null,O,u,p,b,y,w,C,k)}},de=(r,u,p,b,y,w,C)=>{const k=u.el=r.el;let{patchFlag:E,dynamicChildren:_,dirs:O}=u;E|=r.patchFlag&16;const B=r.props||nt,$=u.props||nt;let F;p&&le(p,!1),(F=$.onVnodeBeforeUpdate)&&qt(F,p,u,r),O&&oe(u,r,p,"beforeUpdate"),p&&le(p,!0);const W=y&&u.type!=="foreignObject";if(_?Pt(r.dynamicChildren,_,k,p,b,W,w):C||V(r,u,k,null,p,b,W,w,!1),E>0){if(E&16)Xt(k,u,B,$,p,b,y);else if(E&2&&B.class!==$.class&&l(k,"class",null,$.class,y),E&4&&l(k,"style",B.style,$.style,y),E&8){const et=u.dynamicProps;for(let Z=0;Z<et.length;Z++){const ot=et[Z],xt=B[ot],pe=$[ot];(pe!==xt||ot==="value")&&l(k,ot,xt,pe,y,r.children,p,b,ut)}}E&1&&r.children!==u.children&&f(k,u.children)}else!C&&_==null&&Xt(k,u,B,$,p,b,y);((F=$.onVnodeUpdated)||O)&&bt(()=>{F&&qt(F,p,u,r),O&&oe(u,r,p,"updated")},b)},Pt=(r,u,p,b,y,w,C)=>{for(let k=0;k<u.length;k++){const E=r[k],_=u[k],O=E.el&&(E.type===Mt||!$e(E,_)||E.shapeFlag&70)?m(E.el):p;S(E,_,O,null,b,y,w,C,!0)}},Xt=(r,u,p,b,y,w,C)=>{if(p!==b){if(p!==nt)for(const k in p)!on(k)&&!(k in b)&&l(r,k,p[k],null,C,u.children,y,w,ut);for(const k in b){if(on(k))continue;const E=b[k],_=p[k];E!==_&&k!=="value"&&l(r,k,_,E,C,u.children,y,w,ut)}"value"in b&&l(r,"value",p.value,b.value)}},se=(r,u,p,b,y,w,C,k,E)=>{const _=u.el=r?r.el:d(""),O=u.anchor=r?r.anchor:d("");let{patchFlag:B,dynamicChildren:$,slotScopeIds:F}=u;F&&(k=k?k.concat(F):F),r==null?(s(_,p,b),s(O,p,b),Nt(u.children,p,O,y,w,C,k,E)):B>0&&B&64&&$&&r.dynamicChildren?(Pt(r.dynamicChildren,$,p,y,w,C,k),(u.key!=null||y&&u===y.subTree)&&fl(r,u,!0)):V(r,u,p,O,y,w,C,k,E)},St=(r,u,p,b,y,w,C,k,E)=>{u.slotScopeIds=k,r==null?u.shapeFlag&512?y.ctx.activate(u,p,b,C,E):Se(u,p,b,y,w,C,E):ue(r,u,E)},Se=(r,u,p,b,y,w,C)=>{const k=r.component=ki(r,b,y);if(nl(r)&&(k.ctx.renderer=A),Ei(k),k.asyncDep){if(y&&y.registerDep(k,lt),!r.el){const E=k.subTree=Y(ze);q(null,E,u,p)}return}lt(k,r,u,p,y,w,C)},ue=(r,u,p)=>{const b=u.component=r.component;if(qa(r,u,p))if(b.asyncDep&&!b.asyncResolved){tt(b,u,p);return}else b.next=u,Ia(b.update),b.update();else u.el=r.el,b.vnode=u},lt=(r,u,p,b,y,w,C)=>{const k=()=>{if(r.isMounted){let{next:O,bu:B,u:$,parent:F,vnode:W}=r,et=O,Z;le(r,!1),O?(O.el=W.el,tt(r,O,C)):O=W,B&&Pn(B),(Z=O.props&&O.props.onVnodeBeforeUpdate)&&qt(Z,F,O,W),le(r,!0);const ot=Sn(r),xt=r.subTree;r.subTree=ot,S(xt,ot,m(xt.el),v(xt),r,y,w),O.el=ot.el,et===null&&Ma(r,ot.el),$&&bt($,y),(Z=O.props&&O.props.onVnodeUpdated)&&bt(()=>qt(Z,F,O,W),y)}else{let O;const{el:B,props:$}=u,{bm:F,m:W,parent:et}=r,Z=an(u);if(le(r,!1),F&&Pn(F),!Z&&(O=$&&$.onVnodeBeforeMount)&&qt(O,et,u),le(r,!0),B&&Q){const ot=()=>{r.subTree=Sn(r),Q(B,r.subTree,r,y,null)};Z?u.type.__asyncLoader().then(()=>!r.isUnmounted&&ot()):ot()}else{const ot=r.subTree=Sn(r);S(null,ot,p,b,r,y,w),u.el=ot.el}if(W&&bt(W,y),!Z&&(O=$&&$.onVnodeMounted)){const ot=u;bt(()=>qt(O,et,ot),y)}(u.shapeFlag&256||et&&an(et.vnode)&&et.vnode.shapeFlag&256)&&r.a&&bt(r.a,y),r.isMounted=!0,u=p=b=null}},E=r.effect=new rs(k,()=>gs(_),r.scope),_=r.update=()=>E.run();_.id=r.uid,le(r,!0),_()},tt=(r,u,p)=>{u.component=r;const b=r.vnode.props;r.vnode=u,r.next=null,ci(r,u.props,b,p),ui(r,u.children,p),Te(),Hs(),Re()},V=(r,u,p,b,y,w,C,k,E=!1)=>{const _=r&&r.children,O=r?r.shapeFlag:0,B=u.children,{patchFlag:$,shapeFlag:F}=u;if($>0){if($&128){Jt(_,B,p,b,y,w,C,k,E);return}else if($&256){jt(_,B,p,b,y,w,C,k,E);return}}F&8?(O&16&&ut(_,y,w),B!==_&&f(p,B)):O&16?F&16?Jt(_,B,p,b,y,w,C,k,E):ut(_,y,w,!0):(O&8&&f(p,""),F&16&&Nt(B,p,b,y,w,C,k,E))},jt=(r,u,p,b,y,w,C,k,E)=>{r=r||ye,u=u||ye;const _=r.length,O=u.length,B=Math.min(_,O);let $;for($=0;$<B;$++){const F=u[$]=E?Gt(u[$]):Ht(u[$]);S(r[$],F,p,null,y,w,C,k,E)}_>O?ut(r,y,w,!0,!1,B):Nt(u,p,b,y,w,C,k,E,B)},Jt=(r,u,p,b,y,w,C,k,E)=>{let _=0;const O=u.length;let B=r.length-1,$=O-1;for(;_<=B&&_<=$;){const F=r[_],W=u[_]=E?Gt(u[_]):Ht(u[_]);if($e(F,W))S(F,W,p,null,y,w,C,k,E);else break;_++}for(;_<=B&&_<=$;){const F=r[B],W=u[$]=E?Gt(u[$]):Ht(u[$]);if($e(F,W))S(F,W,p,null,y,w,C,k,E);else break;B--,$--}if(_>B){if(_<=$){const F=$+1,W=F<O?u[F].el:b;for(;_<=$;)S(null,u[_]=E?Gt(u[_]):Ht(u[_]),p,W,y,w,C,k,E),_++}}else if(_>$)for(;_<=B;)gt(r[_],y,w,!0),_++;else{const F=_,W=_,et=new Map;for(_=W;_<=$;_++){const yt=u[_]=E?Gt(u[_]):Ht(u[_]);yt.key!=null&&et.set(yt.key,_)}let Z,ot=0;const xt=$-W+1;let pe=!1,Cs=0;const Oe=new Array(xt);for(_=0;_<xt;_++)Oe[_]=0;for(_=F;_<=B;_++){const yt=r[_];if(ot>=xt){gt(yt,y,w,!0);continue}let $t;if(yt.key!=null)$t=et.get(yt.key);else for(Z=W;Z<=$;Z++)if(Oe[Z-W]===0&&$e(yt,u[Z])){$t=Z;break}$t===void 0?gt(yt,y,w,!0):(Oe[$t-W]=_+1,$t>=Cs?Cs=$t:pe=!0,S(yt,u[$t],p,null,y,w,C,k,E),ot++)}const Is=pe?pi(Oe):ye;for(Z=Is.length-1,_=xt-1;_>=0;_--){const yt=W+_,$t=u[yt],Bs=yt+1<O?u[yt+1].el:b;Oe[_]===0?S(null,$t,p,Bs,y,w,C,k,E):pe&&(Z<0||_!==Is[Z]?Ot($t,p,Bs,2):Z--)}}},Ot=(r,u,p,b,y=null)=>{const{el:w,type:C,transition:k,children:E,shapeFlag:_}=r;if(_&6){Ot(r.component.subTree,u,p,b);return}if(_&128){r.suspense.move(u,p,b);return}if(_&64){C.move(r,u,p,A);return}if(C===Mt){s(w,u,p);for(let B=0;B<E.length;B++)Ot(E[B],u,p,b);s(r.anchor,u,p);return}if(C===rn){X(r,u,p);return}if(b!==2&&_&1&&k)if(b===0)k.beforeEnter(w),s(w,u,p),bt(()=>k.enter(w),y);else{const{leave:B,delayLeave:$,afterLeave:F}=k,W=()=>s(w,u,p),et=()=>{B(w,()=>{W(),F&&F()})};$?$(w,W,et):et()}else s(w,u,p)},gt=(r,u,p,b=!1,y=!1)=>{const{type:w,props:C,ref:k,children:E,dynamicChildren:_,shapeFlag:O,patchFlag:B,dirs:$}=r;if(k!=null&&Jn(k,null,p,r,!0),O&256){u.ctx.deactivate(r);return}const F=O&1&&$,W=!an(r);let et;if(W&&(et=C&&C.onVnodeBeforeUnmount)&&qt(et,u,r),O&6)Qe(r.component,p,b);else{if(O&128){r.suspense.unmount(p,b);return}F&&oe(r,null,u,"beforeUnmount"),O&64?r.type.remove(r,u,p,y,A,b):_&&(w!==Mt||B>0&&B&64)?ut(_,u,p,!1,!0):(w===Mt&&B&384||!y&&O&16)&&ut(E,u,p),b&&he(r)}(W&&(et=C&&C.onVnodeUnmounted)||F)&&bt(()=>{et&&qt(et,u,r),F&&oe(r,null,u,"unmounted")},p)},he=r=>{const{type:u,el:p,anchor:b,transition:y}=r;if(u===Mt){fe(p,b);return}if(u===rn){H(r);return}const w=()=>{o(p),y&&!y.persisted&&y.afterLeave&&y.afterLeave()};if(r.shapeFlag&1&&y&&!y.persisted){const{leave:C,delayLeave:k}=y,E=()=>C(p,w);k?k(r.el,w,E):E()}else w()},fe=(r,u)=>{let p;for(;r!==u;)p=g(r),o(r),r=p;o(u)},Qe=(r,u,p)=>{const{bum:b,scope:y,update:w,subTree:C,um:k}=r;b&&Pn(b),y.stop(),w&&(w.active=!1,gt(C,r,u,p)),k&&bt(k,u),bt(()=>{r.isUnmounted=!0},u),u&&u.pendingBranch&&!u.isUnmounted&&r.asyncDep&&!r.asyncResolved&&r.suspenseId===u.pendingId&&(u.deps--,u.deps===0&&u.resolve())},ut=(r,u,p,b=!1,y=!1,w=0)=>{for(let C=w;C<r.length;C++)gt(r[C],u,p,b,y)},v=r=>r.shapeFlag&6?v(r.component.subTree):r.shapeFlag&128?r.suspense.next():g(r.anchor||r.el),I=(r,u,p)=>{r==null?u._vnode&&gt(u._vnode,null,null,!0):S(u._vnode||null,r,u,null,null,null,p),Hs(),Vo(),u._vnode=r},A={p:S,um:gt,m:Ot,r:he,mt:Se,mc:Nt,pc:V,pbc:Pt,n:v,o:t};let R,Q;return e&&([R,Q]=e(A)),{render:I,hydrate:R,createApp:ai(I,R)}}function le({effect:t,update:e},n){t.allowRecurse=e.allowRecurse=n}function fl(t,e,n=!1){const s=t.children,o=e.children;if(N(s)&&N(o))for(let l=0;l<s.length;l++){const i=s[l];let d=o[l];d.shapeFlag&1&&!d.dynamicChildren&&((d.patchFlag<=0||d.patchFlag===32)&&(d=o[l]=Gt(o[l]),d.el=i.el),n||fl(i,d)),d.type===Cn&&(d.el=i.el)}}function pi(t){const e=t.slice(),n=[0];let s,o,l,i,d;const c=t.length;for(s=0;s<c;s++){const h=t[s];if(h!==0){if(o=n[n.length-1],t[o]<h){e[s]=o,n.push(s);continue}for(l=0,i=n.length-1;l<i;)d=l+i>>1,t[n[d]]<h?l=d+1:i=d;h<t[n[l]]&&(l>0&&(e[s]=n[l-1]),n[l]=s)}}for(l=n.length,i=n[l-1];l-- >0;)n[l]=i,i=e[i];return n}const mi=t=>t.__isTeleport,Mt=Symbol.for("v-fgt"),Cn=Symbol.for("v-txt"),ze=Symbol.for("v-cmt"),rn=Symbol.for("v-stc"),Le=[];let Ct=null;function zt(t=!1){Le.push(Ct=t?null:[])}function gi(){Le.pop(),Ct=Le[Le.length-1]||null}let Ye=1;function Xs(t){Ye+=t}function bi(t){return t.dynamicChildren=Ye>0?Ct||ye:null,gi(),Ye>0&&Ct&&Ct.push(t),t}function Yt(t,e,n,s,o,l){return bi(a(t,e,n,s,o,l,!0))}function Vn(t){return t?t.__v_isVNode===!0:!1}function $e(t,e){return t.type===e.type&&t.key===e.key}const In="__vInternal",pl=({key:t})=>t??null,dn=({ref:t,ref_key:e,ref_for:n})=>(typeof t=="number"&&(t=""+t),t!=null?ct(t)||ft(t)||D(t)?{i:At,r:t,k:e,f:!!n}:t:null);function a(t,e=null,n=null,s=0,o=null,l=t===Mt?0:1,i=!1,d=!1){const c={__v_isVNode:!0,__v_skip:!0,type:t,props:e,key:e&&pl(e),ref:e&&dn(e),scopeId:En,slotScopeIds:null,children:n,component:null,suspense:null,ssContent:null,ssFallback:null,dirs:null,transition:null,el:null,anchor:null,target:null,targetAnchor:null,staticCount:0,shapeFlag:l,patchFlag:s,dynamicProps:o,dynamicChildren:null,appContext:null,ctx:At};return d?(ws(c,n),l&128&&t.normalize(c)):n&&(c.shapeFlag|=ct(n)?8:16),Ye>0&&!i&&Ct&&(c.patchFlag>0||l&6)&&c.patchFlag!==32&&Ct.push(c),c}const Y=vi;function vi(t,e=null,n=null,s=0,o=null,l=!1){if((!t||t===Qa)&&(t=ze),Vn(t)){const d=Ee(t,e,!0);return n&&ws(d,n),Ye>0&&!l&&Ct&&(d.shapeFlag&6?Ct[Ct.indexOf(t)]=d:Ct.push(d)),d.patchFlag|=-2,d}if(Ti(t)&&(t=t.__vccOpts),e){e=yi(e);let{class:d,style:c}=e;d&&!ct(d)&&(e.class=is(d)),st(c)&&(jo(c)&&!N(c)&&(c=it({},c)),e.style=as(c))}const i=ct(t)?1:Ha(t)?128:mi(t)?64:st(t)?4:D(t)?2:0;return a(t,e,n,s,o,i,l,!0)}function yi(t){return t?jo(t)||In in t?it({},t):t:null}function Ee(t,e,n=!1){const{props:s,ref:o,patchFlag:l,children:i}=t,d=e?wi(s||{},e):s;return{__v_isVNode:!0,__v_skip:!0,type:t.type,props:d,key:d&&pl(d),ref:e&&e.ref?n&&o?N(o)?o.concat(dn(e)):[o,dn(e)]:dn(e):o,scopeId:t.scopeId,slotScopeIds:t.slotScopeIds,children:i,target:t.target,targetAnchor:t.targetAnchor,staticCount:t.staticCount,shapeFlag:t.shapeFlag,patchFlag:e&&t.type!==Mt?l===-1?16:l|16:l,dynamicProps:t.dynamicProps,dynamicChildren:t.dynamicChildren,appContext:t.appContext,dirs:t.dirs,transition:t.transition,component:t.component,suspense:t.suspense,ssContent:t.ssContent&&Ee(t.ssContent),ssFallback:t.ssFallback&&Ee(t.ssFallback),el:t.el,anchor:t.anchor,ctx:t.ctx,ce:t.ce}}function P(t=" ",e=0){return Y(Cn,null,t,e)}function K(t,e){const n=Y(rn,null,t);return n.staticCount=e,n}function Ht(t){return t==null||typeof t=="boolean"?Y(ze):N(t)?Y(Mt,null,t.slice()):typeof t=="object"?Gt(t):Y(Cn,null,String(t))}function Gt(t){return t.el===null&&t.patchFlag!==-1||t.memo?t:Ee(t)}function ws(t,e){let n=0;const{shapeFlag:s}=t;if(e==null)e=null;else if(N(e))n=16;else if(typeof e=="object")if(s&65){const o=e.default;o&&(o._c&&(o._d=!1),ws(t,o()),o._c&&(o._d=!0));return}else{n=32;const o=e._;!o&&!(In in e)?e._ctx=At:o===3&&At&&(At.slots._===1?e._=1:(e._=2,t.patchFlag|=1024))}else D(e)?(e={default:e,_ctx:At},n=32):(e=String(e),s&64?(n=16,e=[P(e)]):n=8);t.children=e,t.shapeFlag|=n}function wi(...t){const e={};for(let n=0;n<t.length;n++){const s=t[n];for(const o in s)if(o==="class")e.class!==s.class&&(e.class=is([e.class,s.class]));else if(o==="style")e.style=as([e.style,s.style]);else if(bn(o)){const l=e[o],i=s[o];i&&l!==i&&!(N(l)&&l.includes(i))&&(e[o]=l?[].concat(l,i):i)}else o!==""&&(e[o]=s[o])}return e}function qt(t,e,n,s=null){Bt(t,e,7,[n,s])}const _i=il();let xi=0;function ki(t,e,n){const s=t.type,o=(e?e.appContext:t.appContext)||_i,l={uid:xi++,vnode:t,type:s,parent:e,appContext:o,root:null,next:null,subTree:null,effect:null,update:null,scope:new Ul(!0),render:null,proxy:null,exposed:null,exposeProxy:null,withProxy:null,provides:e?e.provides:Object.create(o.provides),accessCache:null,renderCache:[],components:null,directives:null,propsOptions:rl(s,o),emitsOptions:Go(s,o),emit:null,emitted:null,propsDefaults:nt,inheritAttrs:s.inheritAttrs,ctx:nt,data:nt,props:nt,attrs:nt,slots:nt,refs:nt,setupState:nt,setupContext:null,attrsProxy:null,slotsProxy:null,suspense:n,suspenseId:n?n.pendingId:0,asyncDep:null,asyncResolved:!1,isMounted:!1,isUnmounted:!1,isDeactivated:!1,bc:null,c:null,bm:null,m:null,bu:null,u:null,um:null,bum:null,da:null,a:null,rtg:null,rtc:null,ec:null,sp:null};return l.ctx={_:l},l.root=e?e.root:l,l.emit=Ra.bind(null,l),t.ce&&t.ce(l),l}let rt=null,_s,me,Js="__VUE_INSTANCE_SETTERS__";(me=Ln()[Js])||(me=Ln()[Js]=[]),me.push(t=>rt=t),_s=t=>{me.length>1?me.forEach(e=>e(t)):me[0](t)};const Ae=t=>{_s(t),t.scope.on()},re=()=>{rt&&rt.scope.off(),_s(null)};function ml(t){return t.vnode.shapeFlag&4}let Xe=!1;function Ei(t,e=!1){Xe=e;const{props:n,children:s}=t.vnode,o=ml(t);ii(t,n,o,e),di(t,s);const l=o?Ai(t,e):void 0;return Xe=!1,l}function Ai(t,e){const n=t.type;t.accessCache=Object.create(null),t.proxy=Do(new Proxy(t.ctx,Za));const{setup:s}=n;if(s){const o=t.setupContext=s.length>1?Ii(t):null;Ae(t),Te();const l=ee(s,t,0,[t.props,o]);if(Re(),re(),Ao(l)){if(l.then(re,re),e)return l.then(i=>{Vs(t,i,e)}).catch(i=>{xn(i,t,0)});t.asyncDep=l}else Vs(t,l,e)}else gl(t,e)}function Vs(t,e,n){D(e)?t.type.__ssrInlineRender?t.ssrRender=e:t.render=e:st(e)&&(t.setupState=zo(e)),gl(t,n)}let Qs;function gl(t,e,n){const s=t.type;if(!t.render){if(!e&&Qs&&!s.render){const o=s.template||vs(t).template;if(o){const{isCustomElement:l,compilerOptions:i}=t.appContext.config,{delimiters:d,compilerOptions:c}=s,h=it(it({isCustomElement:l,delimiters:d},i),c);s.render=Qs(o,h)}}t.render=s.render||It}Ae(t),Te(),ti(t),Re(),re()}function Ci(t){return t.attrsProxy||(t.attrsProxy=new Proxy(t.attrs,{get(e,n){return vt(t,"get","$attrs"),e[n]}}))}function Ii(t){const e=n=>{t.exposed=n||{}};return{get attrs(){return Ci(t)},slots:t.slots,emit:t.emit,expose:e}}function xs(t){if(t.exposed)return t.exposeProxy||(t.exposeProxy=new Proxy(zo(Do(t.exposed)),{get(e,n){if(n in e)return e[n];if(n in Fe)return Fe[n](t)},has(e,n){return n in e||n in Fe}}))}function Bi(t,e=!0){return D(t)?t.displayName||t.name:t.name||e&&t.__name}function Ti(t){return D(t)&&"__vccOpts"in t}const wt=(t,e)=>Ea(t,e,Xe);function bl(t,e,n){const s=arguments.length;return s===2?st(e)&&!N(e)?Vn(e)?Y(t,null,[e]):Y(t,e):Y(t,null,e):(s>3?n=Array.prototype.slice.call(arguments,2):s===3&&Vn(n)&&(n=[n]),Y(t,e,n))}const Ri=Symbol.for("v-scx"),Pi=()=>_t(Ri),Si="3.3.1",Oi="http://www.w3.org/2000/svg",ie=typeof document<"u"?document:null,Gs=ie&&ie.createElement("template"),$i={insert:(t,e,n)=>{e.insertBefore(t,n||null)},remove:t=>{const e=t.parentNode;e&&e.removeChild(t)},createElement:(t,e,n,s)=>{const o=e?ie.createElementNS(Oi,t):ie.createElement(t,n?{is:n}:void 0);return t==="select"&&s&&s.multiple!=null&&o.setAttribute("multiple",s.multiple),o},createText:t=>ie.createTextNode(t),createComment:t=>ie.createComment(t),setText:(t,e)=>{t.nodeValue=e},setElementText:(t,e)=>{t.textContent=e},parentNode:t=>t.parentNode,nextSibling:t=>t.nextSibling,querySelector:t=>ie.querySelector(t),setScopeId(t,e){t.setAttribute(e,"")},insertStaticContent(t,e,n,s,o,l){const i=n?n.previousSibling:e.lastChild;if(o&&(o===l||o.nextSibling))for(;e.insertBefore(o.cloneNode(!0),n),!(o===l||!(o=o.nextSibling)););else{Gs.innerHTML=s?`<svg>${t}</svg>`:t;const d=Gs.content;if(s){const c=d.firstChild;for(;c.firstChild;)d.appendChild(c.firstChild);d.removeChild(c)}e.insertBefore(d,n)}return[i?i.nextSibling:e.firstChild,n?n.previousSibling:e.lastChild]}};function qi(t,e,n){const s=t._vtc;s&&(e=(e?[e,...s]:[...s]).join(" ")),e==null?t.removeAttribute("class"):n?t.setAttribute("class",e):t.className=e}function Mi(t,e,n){const s=t.style,o=ct(n);if(n&&!o){if(e&&!ct(e))for(const l in e)n[l]==null&&Qn(s,l,"");for(const l in n)Qn(s,l,n[l])}else{const l=s.display;o?e!==n&&(s.cssText=n):e&&t.removeAttribute("style"),"_vod"in t&&(s.display=l)}}const Zs=/\s*!important$/;function Qn(t,e,n){if(N(n))n.forEach(s=>Qn(t,e,s));else if(n==null&&(n=""),e.startsWith("--"))t.setProperty(e,n);else{const s=Hi(t,e);Zs.test(n)?t.setProperty(Be(s),n.replace(Zs,""),"important"):t[s]=n}}const to=["Webkit","Moz","ms"],$n={};function Hi(t,e){const n=$n[e];if(n)return n;let s=Lt(e);if(s!=="filter"&&s in t)return $n[e]=s;s=wn(s);for(let o=0;o<to.length;o++){const l=to[o]+s;if(l in t)return $n[e]=l}return e}const eo="http://www.w3.org/1999/xlink";function Fi(t,e,n,s,o){if(s&&e.startsWith("xlink:"))n==null?t.removeAttributeNS(eo,e.slice(6,e.length)):t.setAttributeNS(eo,e,n);else{const l=Dl(e);n==null||l&&!Bo(n)?t.removeAttribute(e):t.setAttribute(e,l?"":n)}}function Li(t,e,n,s,o,l,i){if(e==="innerHTML"||e==="textContent"){s&&i(s,o,l),t[e]=n??"";return}const d=t.tagName;if(e==="value"&&d!=="PROGRESS"&&!d.includes("-")){t._value=n;const h=d==="OPTION"?t.getAttribute("value"):t.value,f=n??"";h!==f&&(t.value=f),n==null&&t.removeAttribute(e);return}let c=!1;if(n===""||n==null){const h=typeof t[e];h==="boolean"?n=Bo(n):n==null&&h==="string"?(n="",c=!0):h==="number"&&(n=0,c=!0)}try{t[e]=n}catch{}c&&t.removeAttribute(e)}function Ni(t,e,n,s){t.addEventListener(e,n,s)}function ji(t,e,n,s){t.removeEventListener(e,n,s)}function Di(t,e,n,s,o=null){const l=t._vei||(t._vei={}),i=l[e];if(s&&i)i.value=s;else{const[d,c]=Ui(e);if(s){const h=l[e]=zi(s,o);Ni(t,d,h,c)}else i&&(ji(t,d,i,c),l[e]=void 0)}}const no=/(?:Once|Passive|Capture)$/;function Ui(t){let e;if(no.test(t)){e={};let s;for(;s=t.match(no);)t=t.slice(0,t.length-s[0].length),e[s[0].toLowerCase()]=!0}return[t[2]===":"?t.slice(3):Be(t.slice(2)),e]}let qn=0;const Wi=Promise.resolve(),Ki=()=>qn||(Wi.then(()=>qn=0),qn=Date.now());function zi(t,e){const n=s=>{if(!s._vts)s._vts=Date.now();else if(s._vts<=n.attached)return;Bt(Yi(s,n.value),e,5,[s])};return n.value=t,n.attached=Ki(),n}function Yi(t,e){if(N(e)){const n=t.stopImmediatePropagation;return t.stopImmediatePropagation=()=>{n.call(t),t._stopped=!0},e.map(s=>o=>!o._stopped&&s&&s(o))}else return e}const so=/^on[a-z]/,Xi=(t,e,n,s,o=!1,l,i,d,c)=>{e==="class"?qi(t,s,o):e==="style"?Mi(t,n,s):bn(e)?ns(e)||Di(t,e,n,s,i):(e[0]==="."?(e=e.slice(1),!0):e[0]==="^"?(e=e.slice(1),!1):Ji(t,e,s,o))?Li(t,e,s,l,i,d,c):(e==="true-value"?t._trueValue=s:e==="false-value"&&(t._falseValue=s),Fi(t,e,s,o))};function Ji(t,e,n,s){return s?!!(e==="innerHTML"||e==="textContent"||e in t&&so.test(e)&&D(n)):e==="spellcheck"||e==="draggable"||e==="translate"||e==="form"||e==="list"&&t.tagName==="INPUT"||e==="type"&&t.tagName==="TEXTAREA"||so.test(e)&&ct(n)?!1:e in t}const Vi=it({patchProp:Xi},$i);let oo;function Qi(){return oo||(oo=hi(Vi))}const Gi=(...t)=>{const e=Qi().createApp(...t),{mount:n}=e;return e.mount=s=>{const o=Zi(s);if(!o)return;const l=e._component;!D(l)&&!l.render&&!l.template&&(l.template=o.innerHTML),o.innerHTML="";const i=n(o,!1,o instanceof SVGElement);return o instanceof Element&&(o.removeAttribute("v-cloak"),o.setAttribute("data-v-app","")),i},e};function Zi(t){return ct(t)?document.querySelector(t):t}/*!
+  * vue-router v4.2.0
+  * (c) 2023 Eduardo San Martin Morote
+  * @license MIT
+  */const be=typeof window<"u";function tc(t){return t.__esModule||t[Symbol.toStringTag]==="Module"}const G=Object.assign;function Mn(t,e){const n={};for(const s in e){const o=e[s];n[s]=Tt(o)?o.map(t):t(o)}return n}const Ne=()=>{},Tt=Array.isArray,ec=/\/$/,nc=t=>t.replace(ec,"");function Hn(t,e,n="/"){let s,o={},l="",i="";const d=e.indexOf("#");let c=e.indexOf("?");return d<c&&d>=0&&(c=-1),c>-1&&(s=e.slice(0,c),l=e.slice(c+1,d>-1?d:e.length),o=t(l)),d>-1&&(s=s||e.slice(0,d),i=e.slice(d,e.length)),s=ac(s??e,n),{fullPath:s+(l&&"?")+l+i,path:s,query:o,hash:i}}function sc(t,e){const n=e.query?t(e.query):"";return e.path+(n&&"?")+n+(e.hash||"")}function lo(t,e){return!e||!t.toLowerCase().startsWith(e.toLowerCase())?t:t.slice(e.length)||"/"}function oc(t,e,n){const s=e.matched.length-1,o=n.matched.length-1;return s>-1&&s===o&&Ce(e.matched[s],n.matched[o])&&vl(e.params,n.params)&&t(e.query)===t(n.query)&&e.hash===n.hash}function Ce(t,e){return(t.aliasOf||t)===(e.aliasOf||e)}function vl(t,e){if(Object.keys(t).length!==Object.keys(e).length)return!1;for(const n in t)if(!lc(t[n],e[n]))return!1;return!0}function lc(t,e){return Tt(t)?ao(t,e):Tt(e)?ao(e,t):t===e}function ao(t,e){return Tt(e)?t.length===e.length&&t.every((n,s)=>n===e[s]):t.length===1&&t[0]===e}function ac(t,e){if(t.startsWith("/"))return t;if(!t)return e;const n=e.split("/"),s=t.split("/"),o=s[s.length-1];(o===".."||o===".")&&s.push("");let l=n.length-1,i,d;for(i=0;i<s.length;i++)if(d=s[i],d!==".")if(d==="..")l>1&&l--;else break;return n.slice(0,l).join("/")+"/"+s.slice(i-(i===s.length?1:0)).join("/")}var Je;(function(t){t.pop="pop",t.push="push"})(Je||(Je={}));var je;(function(t){t.back="back",t.forward="forward",t.unknown=""})(je||(je={}));function ic(t){if(!t)if(be){const e=document.querySelector("base");t=e&&e.getAttribute("href")||"/",t=t.replace(/^\w+:\/\/[^\/]+/,"")}else t="/";return t[0]!=="/"&&t[0]!=="#"&&(t="/"+t),nc(t)}const cc=/^[^#]+#/;function rc(t,e){return t.replace(cc,"#")+e}function dc(t,e){const n=document.documentElement.getBoundingClientRect(),s=t.getBoundingClientRect();return{behavior:e.behavior,left:s.left-n.left-(e.left||0),top:s.top-n.top-(e.top||0)}}const Bn=()=>({left:window.pageXOffset,top:window.pageYOffset});function uc(t){let e;if("el"in t){const n=t.el,s=typeof n=="string"&&n.startsWith("#"),o=typeof n=="string"?s?document.getElementById(n.slice(1)):document.querySelector(n):n;if(!o)return;e=dc(o,t)}else e=t;"scrollBehavior"in document.documentElement.style?window.scrollTo(e):window.scrollTo(e.left!=null?e.left:window.pageXOffset,e.top!=null?e.top:window.pageYOffset)}function io(t,e){return(history.state?history.state.position-e:-1)+t}const Gn=new Map;function hc(t,e){Gn.set(t,e)}function fc(t){const e=Gn.get(t);return Gn.delete(t),e}let pc=()=>location.protocol+"//"+location.host;function yl(t,e){const{pathname:n,search:s,hash:o}=e,l=t.indexOf("#");if(l>-1){let d=o.includes(t.slice(l))?t.slice(l).length:1,c=o.slice(d);return c[0]!=="/"&&(c="/"+c),lo(c,"")}return lo(n,t)+s+o}function mc(t,e,n,s){let o=[],l=[],i=null;const d=({state:g})=>{const x=yl(t,location),T=n.value,S=e.value;let j=0;if(g){if(n.value=x,e.value=g,i&&i===T){i=null;return}j=S?g.position-S.position:0}else s(x);o.forEach(q=>{q(n.value,T,{delta:j,type:Je.pop,direction:j?j>0?je.forward:je.back:je.unknown})})};function c(){i=n.value}function h(g){o.push(g);const x=()=>{const T=o.indexOf(g);T>-1&&o.splice(T,1)};return l.push(x),x}function f(){const{history:g}=window;g.state&&g.replaceState(G({},g.state,{scroll:Bn()}),"")}function m(){for(const g of l)g();l=[],window.removeEventListener("popstate",d),window.removeEventListener("beforeunload",f)}return window.addEventListener("popstate",d),window.addEventListener("beforeunload",f,{passive:!0}),{pauseListeners:c,listen:h,destroy:m}}function co(t,e,n,s=!1,o=!1){return{back:t,current:e,forward:n,replaced:s,position:window.history.length,scroll:o?Bn():null}}function gc(t){const{history:e,location:n}=window,s={value:yl(t,n)},o={value:e.state};o.value||l(s.value,{back:null,current:s.value,forward:null,position:e.length-1,replaced:!0,scroll:null},!0);function l(c,h,f){const m=t.indexOf("#"),g=m>-1?(n.host&&document.querySelector("base")?t:t.slice(m))+c:pc()+t+c;try{e[f?"replaceState":"pushState"](h,"",g),o.value=h}catch(x){console.error(x),n[f?"replace":"assign"](g)}}function i(c,h){const f=G({},e.state,co(o.value.back,c,o.value.forward,!0),h,{position:o.value.position});l(c,f,!0),s.value=c}function d(c,h){const f=G({},o.value,e.state,{forward:c,scroll:Bn()});l(f.current,f,!0);const m=G({},co(s.value,c,null),{position:f.position+1},h);l(c,m,!1),s.value=c}return{location:s,state:o,push:d,replace:i}}function bc(t){t=ic(t);const e=gc(t),n=mc(t,e.state,e.location,e.replace);function s(l,i=!0){i||n.pauseListeners(),history.go(l)}const o=G({location:"",base:t,go:s,createHref:rc.bind(null,t)},e,n);return Object.defineProperty(o,"location",{enumerable:!0,get:()=>e.location.value}),Object.defineProperty(o,"state",{enumerable:!0,get:()=>e.state.value}),o}function vc(t){return t=location.host?t||location.pathname+location.search:"",t.includes("#")||(t+="#"),bc(t)}function yc(t){return typeof t=="string"||t&&typeof t=="object"}function wl(t){return typeof t=="string"||typeof t=="symbol"}const Qt={path:"/",name:void 0,params:{},query:{},hash:"",fullPath:"/",matched:[],meta:{},redirectedFrom:void 0},_l=Symbol("");var ro;(function(t){t[t.aborted=4]="aborted",t[t.cancelled=8]="cancelled",t[t.duplicated=16]="duplicated"})(ro||(ro={}));function Ie(t,e){return G(new Error,{type:t,[_l]:!0},e)}function Dt(t,e){return t instanceof Error&&_l in t&&(e==null||!!(t.type&e))}const uo="[^/]+?",wc={sensitive:!1,strict:!1,start:!0,end:!0},_c=/[.+*?^${}()[\]/\\]/g;function xc(t,e){const n=G({},wc,e),s=[];let o=n.start?"^":"";const l=[];for(const h of t){const f=h.length?[]:[90];n.strict&&!h.length&&(o+="/");for(let m=0;m<h.length;m++){const g=h[m];let x=40+(n.sensitive?.25:0);if(g.type===0)m||(o+="/"),o+=g.value.replace(_c,"\\$&"),x+=40;else if(g.type===1){const{value:T,repeatable:S,optional:j,regexp:q}=g;l.push({name:T,repeatable:S,optional:j});const M=q||uo;if(M!==uo){x+=10;try{new RegExp(`(${M})`)}catch(H){throw new Error(`Invalid custom RegExp for param "${T}" (${M}): `+H.message)}}let X=S?`((?:${M})(?:/(?:${M}))*)`:`(${M})`;m||(X=j&&h.length<2?`(?:/${X})`:"/"+X),j&&(X+="?"),o+=X,x+=20,j&&(x+=-8),S&&(x+=-20),M===".*"&&(x+=-50)}f.push(x)}s.push(f)}if(n.strict&&n.end){const h=s.length-1;s[h][s[h].length-1]+=.7000000000000001}n.strict||(o+="/?"),n.end?o+="$":n.strict&&(o+="(?:/|$)");const i=new RegExp(o,n.sensitive?"":"i");function d(h){const f=h.match(i),m={};if(!f)return null;for(let g=1;g<f.length;g++){const x=f[g]||"",T=l[g-1];m[T.name]=x&&T.repeatable?x.split("/"):x}return m}function c(h){let f="",m=!1;for(const g of t){(!m||!f.endsWith("/"))&&(f+="/"),m=!1;for(const x of g)if(x.type===0)f+=x.value;else if(x.type===1){const{value:T,repeatable:S,optional:j}=x,q=T in h?h[T]:"";if(Tt(q)&&!S)throw new Error(`Provided param "${T}" is an array but it is not repeatable (* or + modifiers)`);const M=Tt(q)?q.join("/"):q;if(!M)if(j)g.length<2&&(f.endsWith("/")?f=f.slice(0,-1):m=!0);else throw new Error(`Missing required param "${T}"`);f+=M}}return f||"/"}return{re:i,score:s,keys:l,parse:d,stringify:c}}function kc(t,e){let n=0;for(;n<t.length&&n<e.length;){const s=e[n]-t[n];if(s)return s;n++}return t.length<e.length?t.length===1&&t[0]===40+40?-1:1:t.length>e.length?e.length===1&&e[0]===40+40?1:-1:0}function Ec(t,e){let n=0;const s=t.score,o=e.score;for(;n<s.length&&n<o.length;){const l=kc(s[n],o[n]);if(l)return l;n++}if(Math.abs(o.length-s.length)===1){if(ho(s))return 1;if(ho(o))return-1}return o.length-s.length}function ho(t){const e=t[t.length-1];return t.length>0&&e[e.length-1]<0}const Ac={type:0,value:""},Cc=/[a-zA-Z0-9_]/;function Ic(t){if(!t)return[[]];if(t==="/")return[[Ac]];if(!t.startsWith("/"))throw new Error(`Invalid path "${t}"`);function e(x){throw new Error(`ERR (${n})/"${h}": ${x}`)}let n=0,s=n;const o=[];let l;function i(){l&&o.push(l),l=[]}let d=0,c,h="",f="";function m(){h&&(n===0?l.push({type:0,value:h}):n===1||n===2||n===3?(l.length>1&&(c==="*"||c==="+")&&e(`A repeatable param (${h}) must be alone in its segment. eg: '/:ids+.`),l.push({type:1,value:h,regexp:f,repeatable:c==="*"||c==="+",optional:c==="*"||c==="?"})):e("Invalid state to consume buffer"),h="")}function g(){h+=c}for(;d<t.length;){if(c=t[d++],c==="\\"&&n!==2){s=n,n=4;continue}switch(n){case 0:c==="/"?(h&&m(),i()):c===":"?(m(),n=1):g();break;case 4:g(),n=s;break;case 1:c==="("?n=2:Cc.test(c)?g():(m(),n=0,c!=="*"&&c!=="?"&&c!=="+"&&d--);break;case 2:c===")"?f[f.length-1]=="\\"?f=f.slice(0,-1)+c:n=3:f+=c;break;case 3:m(),n=0,c!=="*"&&c!=="?"&&c!=="+"&&d--,f="";break;default:e("Unknown state");break}}return n===2&&e(`Unfinished custom RegExp for param "${h}"`),m(),i(),o}function Bc(t,e,n){const s=xc(Ic(t.path),n),o=G(s,{record:t,parent:e,children:[],alias:[]});return e&&!o.record.aliasOf==!e.record.aliasOf&&e.children.push(o),o}function Tc(t,e){const n=[],s=new Map;e=mo({strict:!1,end:!0,sensitive:!1},e);function o(f){return s.get(f)}function l(f,m,g){const x=!g,T=Rc(f);T.aliasOf=g&&g.record;const S=mo(e,f),j=[T];if("alias"in f){const X=typeof f.alias=="string"?[f.alias]:f.alias;for(const H of X)j.push(G({},T,{components:g?g.record.components:T.components,path:H,aliasOf:g?g.record:T}))}let q,M;for(const X of j){const{path:H}=X;if(m&&H[0]!=="/"){const dt=m.record.path,pt=dt[dt.length-1]==="/"?"":"/";X.path=m.record.path+(H&&pt+H)}if(q=Bc(X,m,S),g?g.alias.push(q):(M=M||q,M!==q&&M.alias.push(q),x&&f.name&&!po(q)&&i(f.name)),T.children){const dt=T.children;for(let pt=0;pt<dt.length;pt++)l(dt[pt],q,g&&g.children[pt])}g=g||q,(q.record.components&&Object.keys(q.record.components).length||q.record.name||q.record.redirect)&&c(q)}return M?()=>{i(M)}:Ne}function i(f){if(wl(f)){const m=s.get(f);m&&(s.delete(f),n.splice(n.indexOf(m),1),m.children.forEach(i),m.alias.forEach(i))}else{const m=n.indexOf(f);m>-1&&(n.splice(m,1),f.record.name&&s.delete(f.record.name),f.children.forEach(i),f.alias.forEach(i))}}function d(){return n}function c(f){let m=0;for(;m<n.length&&Ec(f,n[m])>=0&&(f.record.path!==n[m].record.path||!xl(f,n[m]));)m++;n.splice(m,0,f),f.record.name&&!po(f)&&s.set(f.record.name,f)}function h(f,m){let g,x={},T,S;if("name"in f&&f.name){if(g=s.get(f.name),!g)throw Ie(1,{location:f});S=g.record.name,x=G(fo(m.params,g.keys.filter(M=>!M.optional).map(M=>M.name)),f.params&&fo(f.params,g.keys.map(M=>M.name))),T=g.stringify(x)}else if("path"in f)T=f.path,g=n.find(M=>M.re.test(T)),g&&(x=g.parse(T),S=g.record.name);else{if(g=m.name?s.get(m.name):n.find(M=>M.re.test(m.path)),!g)throw Ie(1,{location:f,currentLocation:m});S=g.record.name,x=G({},m.params,f.params),T=g.stringify(x)}const j=[];let q=g;for(;q;)j.unshift(q.record),q=q.parent;return{name:S,path:T,params:x,matched:j,meta:Sc(j)}}return t.forEach(f=>l(f)),{addRoute:l,resolve:h,removeRoute:i,getRoutes:d,getRecordMatcher:o}}function fo(t,e){const n={};for(const s of e)s in t&&(n[s]=t[s]);return n}function Rc(t){return{path:t.path,redirect:t.redirect,name:t.name,meta:t.meta||{},aliasOf:void 0,beforeEnter:t.beforeEnter,props:Pc(t),children:t.children||[],instances:{},leaveGuards:new Set,updateGuards:new Set,enterCallbacks:{},components:"components"in t?t.components||null:t.component&&{default:t.component}}}function Pc(t){const e={},n=t.props||!1;if("component"in t)e.default=n;else for(const s in t.components)e[s]=typeof n=="boolean"?n:n[s];return e}function po(t){for(;t;){if(t.record.aliasOf)return!0;t=t.parent}return!1}function Sc(t){return t.reduce((e,n)=>G(e,n.meta),{})}function mo(t,e){const n={};for(const s in t)n[s]=s in e?e[s]:t[s];return n}function xl(t,e){return e.children.some(n=>n===t||xl(t,n))}const kl=/#/g,Oc=/&/g,$c=/\//g,qc=/=/g,Mc=/\?/g,El=/\+/g,Hc=/%5B/g,Fc=/%5D/g,Al=/%5E/g,Lc=/%60/g,Cl=/%7B/g,Nc=/%7C/g,Il=/%7D/g,jc=/%20/g;function ks(t){return encodeURI(""+t).replace(Nc,"|").replace(Hc,"[").replace(Fc,"]")}function Dc(t){return ks(t).replace(Cl,"{").replace(Il,"}").replace(Al,"^")}function Zn(t){return ks(t).replace(El,"%2B").replace(jc,"+").replace(kl,"%23").replace(Oc,"%26").replace(Lc,"`").replace(Cl,"{").replace(Il,"}").replace(Al,"^")}function Uc(t){return Zn(t).replace(qc,"%3D")}function Wc(t){return ks(t).replace(kl,"%23").replace(Mc,"%3F")}function Kc(t){return t==null?"":Wc(t).replace($c,"%2F")}function gn(t){try{return decodeURIComponent(""+t)}catch{}return""+t}function zc(t){const e={};if(t===""||t==="?")return e;const s=(t[0]==="?"?t.slice(1):t).split("&");for(let o=0;o<s.length;++o){const l=s[o].replace(El," "),i=l.indexOf("="),d=gn(i<0?l:l.slice(0,i)),c=i<0?null:gn(l.slice(i+1));if(d in e){let h=e[d];Tt(h)||(h=e[d]=[h]),h.push(c)}else e[d]=c}return e}function go(t){let e="";for(let n in t){const s=t[n];if(n=Uc(n),s==null){s!==void 0&&(e+=(e.length?"&":"")+n);continue}(Tt(s)?s.map(l=>l&&Zn(l)):[s&&Zn(s)]).forEach(l=>{l!==void 0&&(e+=(e.length?"&":"")+n,l!=null&&(e+="="+l))})}return e}function Yc(t){const e={};for(const n in t){const s=t[n];s!==void 0&&(e[n]=Tt(s)?s.map(o=>o==null?null:""+o):s==null?s:""+s)}return e}const Xc=Symbol(""),bo=Symbol(""),Es=Symbol(""),Bl=Symbol(""),ts=Symbol("");function qe(){let t=[];function e(s){return t.push(s),()=>{const o=t.indexOf(s);o>-1&&t.splice(o,1)}}function n(){t=[]}return{add:e,list:()=>t,reset:n}}function Zt(t,e,n,s,o){const l=s&&(s.enterCallbacks[o]=s.enterCallbacks[o]||[]);return()=>new Promise((i,d)=>{const c=m=>{m===!1?d(Ie(4,{from:n,to:e})):m instanceof Error?d(m):yc(m)?d(Ie(2,{from:e,to:m})):(l&&s.enterCallbacks[o]===l&&typeof m=="function"&&l.push(m),i())},h=t.call(s&&s.instances[o],e,n,c);let f=Promise.resolve(h);t.length<3&&(f=f.then(c)),f.catch(m=>d(m))})}function Fn(t,e,n,s){const o=[];for(const l of t)for(const i in l.components){let d=l.components[i];if(!(e!=="beforeRouteEnter"&&!l.instances[i]))if(Jc(d)){const h=(d.__vccOpts||d)[e];h&&o.push(Zt(h,n,s,l,i))}else{let c=d();o.push(()=>c.then(h=>{if(!h)return Promise.reject(new Error(`Couldn't resolve component "${i}" at "${l.path}"`));const f=tc(h)?h.default:h;l.components[i]=f;const g=(f.__vccOpts||f)[e];return g&&Zt(g,n,s,l,i)()}))}}return o}function Jc(t){return typeof t=="object"||"displayName"in t||"props"in t||"__vccOpts"in t}function vo(t){const e=_t(Es),n=_t(Bl),s=wt(()=>e.resolve(L(t.to))),o=wt(()=>{const{matched:c}=s.value,{length:h}=c,f=c[h-1],m=n.matched;if(!f||!m.length)return-1;const g=m.findIndex(Ce.bind(null,f));if(g>-1)return g;const x=yo(c[h-2]);return h>1&&yo(f)===x&&m[m.length-1].path!==x?m.findIndex(Ce.bind(null,c[h-2])):g}),l=wt(()=>o.value>-1&&Zc(n.params,s.value.params)),i=wt(()=>o.value>-1&&o.value===n.matched.length-1&&vl(n.params,s.value.params));function d(c={}){return Gc(c)?e[L(t.replace)?"replace":"push"](L(t.to)).catch(Ne):Promise.resolve()}return{route:s,href:wt(()=>s.value.href),isActive:l,isExactActive:i,navigate:d}}const Vc=el({name:"RouterLink",compatConfig:{MODE:3},props:{to:{type:[String,Object],required:!0},replace:Boolean,activeClass:String,exactActiveClass:String,custom:Boolean,ariaCurrentValue:{type:String,default:"page"}},useLink:vo,setup(t,{slots:e}){const n=Ve(vo(t)),{options:s}=_t(Es),o=wt(()=>({[wo(t.activeClass,s.linkActiveClass,"router-link-active")]:n.isActive,[wo(t.exactActiveClass,s.linkExactActiveClass,"router-link-exact-active")]:n.isExactActive}));return()=>{const l=e.default&&e.default(n);return t.custom?l:bl("a",{"aria-current":n.isExactActive?t.ariaCurrentValue:null,href:n.href,onClick:n.navigate,class:o.value},l)}}}),Qc=Vc;function Gc(t){if(!(t.metaKey||t.altKey||t.ctrlKey||t.shiftKey)&&!t.defaultPrevented&&!(t.button!==void 0&&t.button!==0)){if(t.currentTarget&&t.currentTarget.getAttribute){const e=t.currentTarget.getAttribute("target");if(/\b_blank\b/i.test(e))return}return t.preventDefault&&t.preventDefault(),!0}}function Zc(t,e){for(const n in e){const s=e[n],o=t[n];if(typeof s=="string"){if(s!==o)return!1}else if(!Tt(o)||o.length!==s.length||s.some((l,i)=>l!==o[i]))return!1}return!0}function yo(t){return t?t.aliasOf?t.aliasOf.path:t.path:""}const wo=(t,e,n)=>t??e??n,tr=el({name:"RouterView",inheritAttrs:!1,props:{name:{type:String,default:"default"},route:Object},compatConfig:{MODE:3},setup(t,{attrs:e,slots:n}){const s=_t(ts),o=wt(()=>t.route||s.value),l=_t(bo,0),i=wt(()=>{let h=L(l);const{matched:f}=o.value;let m;for(;(m=f[h])&&!m.components;)h++;return h}),d=wt(()=>o.value.matched[i.value]);cn(bo,wt(()=>i.value+1)),cn(Xc,d),cn(ts,o);const c=ya();return ln(()=>[c.value,d.value,t.name],([h,f,m],[g,x,T])=>{f&&(f.instances[m]=h,x&&x!==f&&h&&h===g&&(f.leaveGuards.size||(f.leaveGuards=x.leaveGuards),f.updateGuards.size||(f.updateGuards=x.updateGuards))),h&&f&&(!x||!Ce(f,x)||!g)&&(f.enterCallbacks[m]||[]).forEach(S=>S(h))},{flush:"post"}),()=>{const h=o.value,f=t.name,m=d.value,g=m&&m.components[f];if(!g)return _o(n.default,{Component:g,route:h});const x=m.props[f],T=x?x===!0?h.params:typeof x=="function"?x(h):x:null,j=bl(g,G({},T,e,{onVnodeUnmounted:q=>{q.component.isUnmounted&&(m.instances[f]=null)},ref:c}));return _o(n.default,{Component:j,route:h})||j}}});function _o(t,e){if(!t)return null;const n=t(e);return n.length===1?n[0]:n}const er=tr;function nr(t){const e=Tc(t.routes,t),n=t.parseQuery||zc,s=t.stringifyQuery||go,o=t.history,l=qe(),i=qe(),d=qe(),c=wa(Qt);let h=Qt;be&&t.scrollBehavior&&"scrollRestoration"in history&&(history.scrollRestoration="manual");const f=Mn.bind(null,v=>""+v),m=Mn.bind(null,Kc),g=Mn.bind(null,gn);function x(v,I){let A,R;return wl(v)?(A=e.getRecordMatcher(v),R=I):R=v,e.addRoute(R,A)}function T(v){const I=e.getRecordMatcher(v);I&&e.removeRoute(I)}function S(){return e.getRoutes().map(v=>v.record)}function j(v){return!!e.getRecordMatcher(v)}function q(v,I){if(I=G({},I||c.value),typeof v=="string"){const p=Hn(n,v,I.path),b=e.resolve({path:p.path},I),y=o.createHref(p.fullPath);return G(p,b,{params:g(b.params),hash:gn(p.hash),redirectedFrom:void 0,href:y})}let A;if("path"in v)A=G({},v,{path:Hn(n,v.path,I.path).path});else{const p=G({},v.params);for(const b in p)p[b]==null&&delete p[b];A=G({},v,{params:m(p)}),I.params=m(I.params)}const R=e.resolve(A,I),Q=v.hash||"";R.params=f(g(R.params));const r=sc(s,G({},v,{hash:Dc(Q),path:R.path})),u=o.createHref(r);return G({fullPath:r,hash:Q,query:s===go?Yc(v.query):v.query||{}},R,{redirectedFrom:void 0,href:u})}function M(v){return typeof v=="string"?Hn(n,v,c.value.path):G({},v)}function X(v,I){if(h!==v)return Ie(8,{from:I,to:v})}function H(v){return Rt(v)}function dt(v){return H(G(M(v),{replace:!0}))}function pt(v){const I=v.matched[v.matched.length-1];if(I&&I.redirect){const{redirect:A}=I;let R=typeof A=="function"?A(v):A;return typeof R=="string"&&(R=R.includes("?")||R.includes("#")?R=M(R):{path:R},R.params={}),G({query:v.query,hash:v.hash,params:"path"in R?{}:v.params},R)}}function Rt(v,I){const A=h=q(v),R=c.value,Q=v.state,r=v.force,u=v.replace===!0,p=pt(A);if(p)return Rt(G(M(p),{state:typeof p=="object"?G({},Q,p.state):Q,force:r,replace:u}),I||A);const b=A;b.redirectedFrom=I;let y;return!r&&oc(s,R,A)&&(y=Ie(16,{to:b,from:R}),Ot(R,R,!0,!1)),(y?Promise.resolve(y):Pt(b,R)).catch(w=>Dt(w)?Dt(w,2)?w:Jt(w):V(w,b,R)).then(w=>{if(w){if(Dt(w,2))return Rt(G({replace:u},M(w.to),{state:typeof w.to=="object"?G({},Q,w.to.state):Q,force:r}),I||b)}else w=se(b,R,!0,u,Q);return Xt(b,R,w),w})}function Nt(v,I){const A=X(v,I);return A?Promise.reject(A):Promise.resolve()}function de(v){const I=fe.values().next().value;return I&&typeof I.runWithContext=="function"?I.runWithContext(v):v()}function Pt(v,I){let A;const[R,Q,r]=sr(v,I);A=Fn(R.reverse(),"beforeRouteLeave",v,I);for(const p of R)p.leaveGuards.forEach(b=>{A.push(Zt(b,v,I))});const u=Nt.bind(null,v,I);return A.push(u),ut(A).then(()=>{A=[];for(const p of l.list())A.push(Zt(p,v,I));return A.push(u),ut(A)}).then(()=>{A=Fn(Q,"beforeRouteUpdate",v,I);for(const p of Q)p.updateGuards.forEach(b=>{A.push(Zt(b,v,I))});return A.push(u),ut(A)}).then(()=>{A=[];for(const p of v.matched)if(p.beforeEnter&&!I.matched.includes(p))if(Tt(p.beforeEnter))for(const b of p.beforeEnter)A.push(Zt(b,v,I));else A.push(Zt(p.beforeEnter,v,I));return A.push(u),ut(A)}).then(()=>(v.matched.forEach(p=>p.enterCallbacks={}),A=Fn(r,"beforeRouteEnter",v,I),A.push(u),ut(A))).then(()=>{A=[];for(const p of i.list())A.push(Zt(p,v,I));return A.push(u),ut(A)}).catch(p=>Dt(p,8)?p:Promise.reject(p))}function Xt(v,I,A){for(const R of d.list())de(()=>R(v,I,A))}function se(v,I,A,R,Q){const r=X(v,I);if(r)return r;const u=I===Qt,p=be?history.state:{};A&&(R||u?o.replace(v.fullPath,G({scroll:u&&p&&p.scroll},Q)):o.push(v.fullPath,Q)),c.value=v,Ot(v,I,A,u),Jt()}let St;function Se(){St||(St=o.listen((v,I,A)=>{if(!Qe.listening)return;const R=q(v),Q=pt(R);if(Q){Rt(G(Q,{replace:!0}),R).catch(Ne);return}h=R;const r=c.value;be&&hc(io(r.fullPath,A.delta),Bn()),Pt(R,r).catch(u=>Dt(u,12)?u:Dt(u,2)?(Rt(u.to,R).then(p=>{Dt(p,20)&&!A.delta&&A.type===Je.pop&&o.go(-1,!1)}).catch(Ne),Promise.reject()):(A.delta&&o.go(-A.delta,!1),V(u,R,r))).then(u=>{u=u||se(R,r,!1),u&&(A.delta&&!Dt(u,8)?o.go(-A.delta,!1):A.type===Je.pop&&Dt(u,20)&&o.go(-1,!1)),Xt(R,r,u)}).catch(Ne)}))}let ue=qe(),lt=qe(),tt;function V(v,I,A){Jt(v);const R=lt.list();return R.length?R.forEach(Q=>Q(v,I,A)):console.error(v),Promise.reject(v)}function jt(){return tt&&c.value!==Qt?Promise.resolve():new Promise((v,I)=>{ue.add([v,I])})}function Jt(v){return tt||(tt=!v,Se(),ue.list().forEach(([I,A])=>v?A(v):I()),ue.reset()),v}function Ot(v,I,A,R){const{scrollBehavior:Q}=t;if(!be||!Q)return Promise.resolve();const r=!A&&fc(io(v.fullPath,0))||(R||!A)&&history.state&&history.state.scroll||null;return Xo().then(()=>Q(v,I,r)).then(u=>u&&uc(u)).catch(u=>V(u,v,I))}const gt=v=>o.go(v);let he;const fe=new Set,Qe={currentRoute:c,listening:!0,addRoute:x,removeRoute:T,hasRoute:j,getRoutes:S,resolve:q,options:t,push:H,replace:dt,go:gt,back:()=>gt(-1),forward:()=>gt(1),beforeEach:l.add,beforeResolve:i.add,afterEach:d.add,onError:lt.add,isReady:jt,install(v){const I=this;v.component("RouterLink",Qc),v.component("RouterView",er),v.config.globalProperties.$router=I,Object.defineProperty(v.config.globalProperties,"$route",{enumerable:!0,get:()=>L(c)}),be&&!he&&c.value===Qt&&(he=!0,H(o.location).catch(Q=>{}));const A={};for(const Q in Qt)A[Q]=wt(()=>c.value[Q]);v.provide(Es,I),v.provide(Bl,Ve(A)),v.provide(ts,c);const R=v.unmount;fe.add(v),v.unmount=function(){fe.delete(v),fe.size<1&&(h=Qt,St&&St(),St=null,c.value=Qt,he=!1,tt=!1),R()}}};function ut(v){return v.reduce((I,A)=>I.then(()=>de(A)),Promise.resolve())}return Qe}function sr(t,e){const n=[],s=[],o=[],l=Math.max(e.matched.length,t.matched.length);for(let i=0;i<l;i++){const d=e.matched[i];d&&(t.matched.find(h=>Ce(h,d))?s.push(d):n.push(d));const c=t.matched[i];c&&(e.matched.find(h=>Ce(h,c))||o.push(c))}return[n,s,o]}const Tn=(t,e)=>{const n=t.__vccOpts||t;for(const[s,o]of e)n[s]=o;return n},or={},Tl=t=>(Pa("data-v-1849710e"),t=t(),Sa(),t),lr={id:"main-nav",class:"main-navigation"},ar={class:"desktop"},ir=Tl(()=>a("h2",null,"Jelmers Easy Style System",-1)),cr={class:"mobile"},rr=Tl(()=>a("h2",{class:"pb-3"},"Jelmers Easy Style System",-1)),dr={class:"flex-layout three-columns gap-3"};function ur(t,e){const n=bs("router-link");return zt(),Yt("nav",lr,[a("ul",ar,[a("li",null,[Y(n,{to:"/",class:"no-deco"},{default:at(()=>[ir]),_:1})]),a("li",null,[Y(n,{to:"/"},{default:at(()=>[P("Home")]),_:1})]),a("li",null,[Y(n,{to:"/layout"},{default:at(()=>[P("Layout")]),_:1})]),a("li",null,[Y(n,{to:"/buttons-and-inputs"},{default:at(()=>[P("Buttons and Inputs")]),_:1})]),a("li",null,[Y(n,{to:"/components"},{default:at(()=>[P("Components")]),_:1})]),a("li",null,[Y(n,{to:"/typography"},{default:at(()=>[P("Typography")]),_:1})]),a("li",null,[Y(n,{to:"/utility-classes"},{default:at(()=>[P("Utility classes")]),_:1})]),a("li",null,[Y(n,{to:"/icons"},{default:at(()=>[P("Icons")]),_:1})])]),a("div",cr,[a("div",null,[Y(n,{to:"/",class:"no-deco"},{default:at(()=>[rr]),_:1})]),a("section",dr,[Y(n,{to:"/"},{default:at(()=>[P("Home")]),_:1}),Y(n,{to:"/layout"},{default:at(()=>[P("Layout")]),_:1}),Y(n,{to:"/buttons-and-inputs"},{default:at(()=>[P("Buttons and Inputs")]),_:1}),Y(n,{to:"/components"},{default:at(()=>[P("Components")]),_:1}),Y(n,{to:"/typography"},{default:at(()=>[P("Typography")]),_:1}),Y(n,{to:"/utility-classes"},{default:at(()=>[P("Utility classes")]),_:1}),Y(n,{to:"/icons"},{default:at(()=>[P("Icons")]),_:1})])])])}const hr=Tn(or,[["render",ur],["__scopeId","data-v-1849710e"]]),xo="/JESS/images/pennions-logo.svg",fr={},pr=K('<footer class="main-footer desktop mt-auto"><div class="ml-auto row align-center"><img class="inline-block" height="36" src="'+xo+'" alt="The logo of Pennions."></div></footer><footer class="main-footer mt-auto mobile"><section class="column align-center w-100"><div class="row"><div class="ml-auto row align-center"><img class="inline-block" height="36" src="'+xo+'" alt="The logo of Pennions."></div></div></section></footer>',2);function mr(t,e){return pr}const gr=Tn(fr,[["render",mr]]),br={__name:"App",setup(t){return(e,n)=>{const s=bs("RouterView");return zt(),Yt(Mt,null,[Y(hr),Y(s),Y(gr,{class:"mt-auto"})],64)}}},vr={},yr={class:"mx-3"},wr=K('<article class="card column"><h1>Welcome!</h1><p> I have written this Style System so you can quickly make any webpage look consistent and good without having an overkill on features. <br> Keeping its footprint low. And I separated theming. Which means colours, borders, box-shadows, fonts and the like can be adjusted super easy. </p><h2>Features</h2><ul><li>Out-of-the-box responsive</li><li>A lot of layout variations</li><li>Semantic classes and use of HTML5 elements</li><li>Based on Microsoft Fluent Design.</li><li>Allows for easy theming.</li><li>Built with Less, so you can easily modify inside a webbrowser.</li></ul><h2>Goal</h2><p> My goal is to make the life of fullstack webapplication developers easier by providing a lot of things out-of-the-box <br> so they can focus on the business logic and add value instantly. </p></article>',1),_r=[wr];function xr(t,e){return zt(),Yt("main",yr,_r)}const kr=Tn(vr,[["render",xr]]),Er={class:"flex-layout pl-0 gap-3"},Ar={class:"sidebar"},Cr={class:"sticky p-3"},Ir=K('<article class="card column"><header><h1>Layout elements</h1></header><div class="body column"><p> Use the class <code class="language-html">flex-layout</code> as a base for an easy flex layout. It will apply <code class="language-css">display:flex</code> and will unlock more classes that can be used on the child elements. <br> The child elements are all displayed by default at <code class="language-css">width: 100%</code> which results in a single column. <br> All columns will be automatically scaled down and wrapped when the viewport can not fit all the columns properly. <br> You can try it by resizing the browser. <br> If you do not want it to have an automatic row layout where the children have 100% width you can apply the <code class="language-html">auto</code> class. e.g. <code class="language-html">class=&quot;flex-layout auto&quot;</code></p><hr><small><b>N.B.</b> The blue fill texture is for demonstration purpose only. </small></div></article>',1),Br=a("header",null,[a("h2",null,"Single column")],-1),Tr=a("div",{class:"body column"},[a("p",null,[P(" All child elements start out as "),a("code",{class:"language-css"},"width:100%;"),P(" and will take up more space if there is room, up to their specified maximum. ")]),a("section",{class:"flex-layout"},[a("div",{class:"demo-block"},"Block 100% width")])],-1),Rr=K('<header><h2>Two columns</h2></header><p> Add <code class="language-html">class=&quot;flex-layout two-columns&quot;</code> to get a double column layout. </p><ul><li> It will display two columns when the viewport has <code class="language-css">min-width:576px;</code></li></ul><section class="flex-layout two-columns"><div class="demo-block">Block 50% width</div><div class="demo-block">Block 50% width</div></section>',4),Pr=K('<header><h2>Three columns</h2></header><p> Add <code class="language-html">class=&quot;flex-layout three-columns&quot;</code> to get a three column layout. </p><ul><li> It will display three columns when the viewport has <code class="language-css">min-width:768px;</code></li><li> It will display two columns when the viewport has <code class="language-css">min-width:576px;</code></li></ul><section class="flex-layout three-columns"><div class="demo-block">Block 33.336% width</div><div class="demo-block">Block 33.336% width</div><div class="demo-block">Block 33.336% width</div></section>',4),Sr=K('<header><h2>Four columns</h2></header><p> Add <code class="language-html">class=&quot;flex-layout four-columns&quot;</code> to get a four column layout. </p><ul><li> It will display four columns when the viewport has <code class="language-css">min-width:992px;</code></li><li> It will display three columns when the viewport has <code class="language-css">min-width:768px;</code></li><li> It will display two columns when the viewport has <code class="language-css">min-width:576px;</code></li></ul><section class="flex-layout four-columns"><div class="demo-block">Block 25% width</div><div class="demo-block">Block 25% width</div><div class="demo-block">Block 25% width</div><div class="demo-block">Block 25% width</div></section>',4),Or=K('<header><h2>Five columns</h2></header><p> Add <code class="language-html">class=&quot;flex-layout five-columns&quot;</code> to get a five column layout. </p><ul><li> It will display five columns when the viewport has <code class="language-css">min-width:1200px;</code></li><li> It will display four columns when the viewport has <code class="language-css">min-width:992px;</code></li><li> It will display three columns when the viewport has <code class="language-css">min-width:768px;</code></li><li> It will display two columns when the viewport has <code class="language-css">min-width:576px;</code></li></ul><section class="flex-layout five-columns"><div class="demo-block">Block 20% width</div><div class="demo-block">Block 20% width</div><div class="demo-block">Block 20% width</div><div class="demo-block">Block 20% width</div><div class="demo-block">Block 20% width</div></section>',4),$r=K('<header><h2>Six columns</h2></header><p> Add <code class="language-html">class=&quot;flex-layout six-columns&quot;</code> to get a six column layout. </p><ul><li> It will display six columns when the viewport has <code class="language-css">min-width:1400px;</code></li><li> It will display five columns when the viewport has <code class="language-css">min-width:1200px;</code></li><li> It will display four columns when the viewport has <code class="language-css">min-width:992px;</code></li><li> It will display three columns when the viewport has <code class="language-css">min-width:768px;</code></li><li> It will display two columns when the viewport has <code class="language-css">min-width:576px;</code></li></ul><section class="flex-layout six-columns"><div class="demo-block">Block 16.667% width</div><div class="demo-block">Block 16.667% width</div><div class="demo-block">Block 16.667% width</div><div class="demo-block">Block 16.667% width</div><div class="demo-block">Block 16.667% width</div><div class="demo-block">Block 16.667% width</div></section>',4),qr=K('<header><h2>Twelve columns</h2></header><p> Add <code class="language-html">class=&quot;twelve-columns&quot;</code> to get a twelve column layout. </p><ul><li> It will display twelve columns when the viewport has <code class="language-css">min-width:1400px;</code></li><li> It will display six columns when the viewport has <code class="language-css">min-width:1200px;</code></li><li> It will display five columns when the viewport has <code class="language-css">min-width:992px;</code></li><li> It will display four columns when the viewport has <code class="language-css">min-width:768px;</code></li><li> It will display three columns when the viewport has <code class="language-css">min-width:576px;</code></li></ul><section class="flex-layout twelve-columns"><div class="demo-block">Block 8.3333% width</div><div class="demo-block">Block 8.3333% width</div><div class="demo-block">Block 8.3333% width</div><div class="demo-block">Block 8.3333% width</div><div class="demo-block">Block 8.3333% width</div><div class="demo-block">Block 8.3333% width</div><div class="demo-block">Block 8.3333% width</div><div class="demo-block">Block 8.3333% width</div><div class="demo-block">Block 8.3333% width</div><div class="demo-block">Block 8.3333% width</div><div class="demo-block">Block 8.3333% width</div><div class="demo-block">Block 8.3333% width</div></section>',4),Mr=K('<header><h2>Aside layout</h2></header><p> The <code class="language-css">sidebar</code> can be used if you want to have a small - wide layout or wide - small layout for that matter. <br> It will apply: <code class="language-css">width:auto;</code><br> The element with the <code class="language-css">main</code> class will have <code class="language-css">flex:1;</code>applied and take the remaining space. <br></p>',2),Hr=a("hr",null,null,-1),Fr=a("small",null,[a("b",null,"Fun fact:"),P(" the current page uses this aside layout for having a vertical navigation bar. ")],-1),Lr=K('<header><h2>Nav</h2></header><p> When you apply a <code class="language-css">.main-navigation</code> to a <code class="language-html">nav</code> element it will: </p><ul><li>Have a box-shadow</li><li>Have <code class="language-css">max-height: 6rem;</code></li><li>Have <code class="language-css">padding: 1.5rem;</code></li></ul><hr><section><nav class="demo-navbar">Navigation bar</nav></section><p> A navigation bar with links using <code class="language-html">ul</code> and <code class="language-html">li</code>: </p><section><nav class="demo-navbar"><ul><li><a href>Link 1</a></li><li><a href>Link 2</a></li><li><a href>Link 3</a></li></ul></nav></section><hr>',8),Nr=K('<header><h2>Footer</h2></header><p> When you add the class <code class="language-css">.main-footer</code> to a <code class="language-html">&lt;footer&gt;</code> element it will: </p><ul><li>Have a box-shadow</li><li>Have <code class="language-css">min-height: 5rem;</code></li><li>Have <code class="language-css">padding: 1.5rem;</code></li></ul><hr><section><footer class="main-footer"><span>Footer</span></footer></section>',5),jr=K('<header><h2>Nested sections example</h2></header><section class="flex-layout gap-3"><div class="demo-block">Block</div><section class="flex-layout two-columns"><section class="flex-layout five-columns"><div class="demo-block">Block</div><div class="demo-block">Block</div><div class="demo-block">Block</div><div class="demo-block">Block</div><div class="demo-block">Block</div></section><div class="demo-block">Block</div></section><section class="row gap-3"><div class="demo-block w-75">Block</div><div class="demo-block w-25">Block</div></section></section>',2),Dr=a("article",{id:"hr",class:"card column"},[a("header",null,[a("h1",null,"Horizontal rule")]),a("div",{class:"body column"},[a("p",null,[P(" A default "),a("code",{class:"language-html"},"<hr>"),P(" element will render with "),a("code",{class:"language-css"},"width:100%;")]),a("p",null,"Example:"),a("section",null,[a("hr")])])],-1),Ur=K('<header><h1>Table</h1></header><div class="body column no-stretch gap-3"><p>Default table</p><table><thead><tr><th>Column header 1</th><th>Column header 2</th><th>Column header 3</th></tr></thead><tbody><tr><td>A</td><td>B</td><td>C</td></tr><tr><td>D</td><td>E</td><td>F</td></tr><tr><td>G</td><td>H</td><td>I</td></tr></tbody><tfoot><tr><td colspan="3">Table footer with colspan 3</td></tr></tfoot></table><p> A table with the <code class="language-html">table</code> class. </p><table class="table"><thead><tr><th>Column header 1</th><th>Column header 2</th><th>Column header 3</th></tr></thead><tbody><tr><td>A</td><td>B</td><td>C</td></tr><tr><td>D</td><td>E</td><td>F</td></tr><tr><td>G</td><td>H</td><td>I</td></tr></tbody><tfoot><tr><td colspan="3">Table footer with colspan 3</td></tr></tfoot></table><p>With a sticky header:</p><div class="table-container" style="height:12vh;"><table class="table sticky-header"><thead><tr><th>Column 1</th><th>Column 2</th><th>Column 3</th></tr></thead><tbody><tr><td>A</td><td>B</td><td>C</td></tr><tr><td>D</td><td>E</td><td>F</td></tr><tr><td>G</td><td>H</td><td>I</td></tr><tr><td>J</td><td>K</td><td>L</td></tr><tr><td>M</td><td>N</td><td>O</td></tr><tr><td>P</td><td>Q</td><td>R</td></tr><tr><td>S</td><td>T</td><td>U</td></tr><tr><td>V</td><td>W</td><td>X</td></tr><tr><td>Y</td><td>Z</td><td>-</td></tr></tbody><tfoot><tr><td colspan="3">Table footer with colspan 3</td></tr></tfoot></table></div><p> A table with the <code class="language-html">table</code> and <code class="language-html">striped</code> class. </p><table class="table striped"><thead><tr><th>Column header 1</th><th>Column header 2</th><th>Column header 3</th></tr></thead><tbody><tr><td>A</td><td>B</td><td>C</td></tr><tr><td>D</td><td>E</td><td>F</td></tr><tr><td>G</td><td>H</td><td>I</td></tr></tbody></table></div>',2),Wr=`<section class="flex-layout">
+    <div>Block 100% width</div>
+</section>`,Kr=`<section class="flex-layout two-columns">
+    <div>Block 50% width</div>
+    <div>Block 50% width</div>
+</section>`,zr=`<section class="flex-layout three-columns">
+    <div>Block 33.336% width</div>
+    <div>Block 33.336% width</div>
+    <div>Block 33.336% width</div>
+</section>`,Yr=`<section class="flex-layout four-columns">
+    <div>Block 25% width</div>
+</section>`,Xr=`<section class="flex-layout five-columns">
+    <div>Block 20% width</div>
+    <div>Block 20% width</div>
+    <div>Block 20% width</div>
+    <div>Block 20% width</div>
+    <div>Block 20% width</div>
+</section>`,Jr=`<section class="flex-layout six-columns">
+    <div>Block 16.667% width</div>
+    <div>Block 16.667% width</div>
+    <div>Block 16.667% width</div>
+    <div>Block 16.667% width</div>
+    <div>Block 16.667% width</div>
+    <div>Block 16.667% width</div>
+</section>`,Vr=`<section class="flex-layout twelve-columns">
+    <div>Block 8.3333% width</div>
+    <div>Block 8.3333% width</div>
+    <div>Block 8.3333% width</div>
+    <div>Block 8.3333% width</div>
+    <div>Block 8.3333% width</div>
+    <div>Block 8.3333% width</div>
+    <div>Block 8.3333% width</div>
+    <div>Block 8.3333% width</div>
+    <div>Block 8.3333% width</div>
+    <div>Block 8.3333% width</div>
+    <div>Block 8.3333% width</div>
+    <div>Block 8.3333% width</div>
+</section>`,Qr=`<main class="flex-layout">
+    <aside class="sidebar">
+        Element with auto width
+    </aside>
+    <section class="main">
+        Element that takes the remaining space
+    </section>
+</main>`,Gr=`<nav class="main-navigation">
+    <ul>
+      <li><a href>Link 1</a></li>
+      <li><a href>Link 2</a></li>
+      <li><a href>Link 3</a></li>
+    </ul>
+</nav>`,Zr='<footer class="main-footer">Footer</footer>',td=`<section class="flex-layout gap-3">
+  <div class="demo-block">Block</div>
+  <section class="flex-layout two-columns">
+    <section class="flex-layout five-columns">
+      <div class="demo-block">Block</div>
+      <div class="demo-block">Block</div>
+      <div class="demo-block">Block</div>
+      <div class="demo-block">Block</div>
+      <div class="demo-block">Block</div>
+    </section>
+    <div class="demo-block">Block</div>
+  </section>
+  <section class="row gap-3">
+    <div class="demo-block w-75">Block</div>
+    <div class="demo-block w-25">Block</div>
+  </section>
+</section>`,ed=`<table>
+  <thead>
+    <tr>
+      <th>Column header 1</th>
+      <th>Column header 2</th>
+      <th>Column header 3</th>
+    </tr>
+  </thead>
+  <tbody>
+    <tr>
+      <td>A</td>
+      <td>B</td>
+      <td>C</td>
+    </tr>
+    <tr>
+      <td>D</td>
+      <td>E</td>
+      <td>F</td>
+    </tr>
+    <tr>
+      <td>G</td>
+      <td>H</td>
+      <td>I</td>
+    </tr>
+  </tbody>
+  <tfoot>
+    <tr>
+      <td colspan="3">Table footer with colspan 3</td>
+    </tr>
+  </tfoot>
+</table>
+
+<table class="table">
+  <thead>
+    <tr>
+      <th>Column header 1</th>
+      <th>Column header 2</th>
+      <th>Column header 3</th>
+    </tr>
+  </thead>
+  <tbody>
+    <tr>
+      <td>A</td>
+      <td>B</td>
+      <td>C</td>
+    </tr>
+    <tr>
+      <td>D</td>
+      <td>E</td>
+      <td>F</td>
+    </tr>
+    <tr>
+      <td>G</td>
+      <td>H</td>
+      <td>I</td>
+    </tr>
+  </tbody>
+  <tfoot>
+    <tr>
+      <td colspan="3">Table footer with colspan 3</td>
+    </tr>
+  </tfoot>
+</table>
+
+<div class="table-container" style="height: 12vh">
+  <table class="table sticky-header">
+    <thead>
+      <tr>
+        <th>Column 1</th>
+        <th>Column 2</th>
+        <th>Column 3</th>
+      </tr>
+    </thead>
+    <tbody>
+      <tr>
+        <td>A</td>
+        <td>B</td>
+        <td>C</td>
+      </tr>
+      <tr>
+        <td>D</td>
+        <td>E</td>
+        <td>F</td>
+      </tr>
+      <tr>
+        <td>G</td>
+        <td>H</td>
+        <td>I</td>
+      </tr>
+      <tr>
+        <td>J</td>
+        <td>K</td>
+        <td>L</td>
+      </tr>
+      <tr>
+        <td>M</td>
+        <td>N</td>
+        <td>O</td>
+      </tr>
+      <tr>
+        <td>P</td>
+        <td>Q</td>
+        <td>R</td>
+      </tr>
+      <tr>
+        <td>S</td>
+        <td>T</td>
+        <td>U</td>
+      </tr>
+      <tr>
+        <td>V</td>
+        <td>W</td>
+        <td>X</td>
+      </tr>
+      <tr>
+        <td>Y</td>
+        <td>Z</td>
+        <td>-</td>
+      </tr>
+    </tbody>
+    <tfoot>
+      <tr>
+        <td colspan="3">Table footer with colspan 3</td>
+      </tr>
+    </tfoot>
+  </table>
+</div>
+
+<table class="table striped">
+  <thead>
+    <tr>
+      <th>Column header 1</th>
+      <th>Column header 2</th>
+      <th>Column header 3</th>
+    </tr>
+  </thead>
+  <tbody>
+    <tr>
+      <td>A</td>
+      <td>B</td>
+      <td>C</td>
+    </tr>
+    <tr>
+      <td>D</td>
+      <td>E</td>
+      <td>F</td>
+    </tr>
+    <tr>
+      <td>G</td>
+      <td>H</td>
+      <td>I</td>
+    </tr>
+  </tbody>
+</table>`,nd={__name:"Layout",setup(t){const e=_t("scrollTo");return Pe(()=>{Prism.highlightAll()}),(n,s)=>(zt(),Yt("main",Er,[a("aside",Ar,[a("nav",Cr,[a("ul",null,[a("li",null,[a("a",{onClick:s[0]||(s[0]=o=>L(e)("single-column"))},"Single column")]),a("li",null,[a("a",{onClick:s[1]||(s[1]=o=>L(e)("two-columns"))},"Two columns")]),a("li",null,[a("a",{onClick:s[2]||(s[2]=o=>L(e)("three-columns"))},"Three columns")]),a("li",null,[a("a",{onClick:s[3]||(s[3]=o=>L(e)("four-columns"))},"Four columns")]),a("li",null,[a("a",{onClick:s[4]||(s[4]=o=>L(e)("five-columns"))},"Five columns")]),a("li",null,[a("a",{onClick:s[5]||(s[5]=o=>L(e)("six-columns"))},"Six columns")]),a("li",null,[a("a",{onClick:s[6]||(s[6]=o=>L(e)("twelve-columns"))},"Twelve columns")]),a("li",null,[a("a",{onClick:s[7]||(s[7]=o=>L(e)("aside"))},"Aside")]),a("li",null,[a("a",{onClick:s[8]||(s[8]=o=>L(e)("nav"))},"Nav")]),a("li",null,[a("a",{onClick:s[9]||(s[9]=o=>L(e)("footer"))},"Footer")]),a("li",null,[a("a",{onClick:s[10]||(s[10]=o=>L(e)("nested"))},"Nested sections example")]),a("li",null,[a("a",{onClick:s[11]||(s[11]=o=>L(e)("hr"))},"Horizontal rule")]),a("li",null,[a("a",{onClick:s[12]||(s[12]=o=>L(e)("table"))},"Table")]),a("li",null,[a("a",{onClick:s[13]||(s[13]=o=>L(e)("main-nav")),class:"button normal desktop"},"Back to top")])])])]),a("section",{class:"flex-layout main column mr-3"},[Ir,a("article",{id:"single-column",class:"card column"},[Br,Tr,a("pre",{class:"border"},[a("code",{class:"language-html"},U(Wr))])]),a("article",{id:"two-columns",class:"card column"},[Rr,a("pre",{class:"border"},[a("code",{class:"language-html"},U(Kr))])]),a("article",{id:"three-columns",class:"card column"},[Pr,a("pre",{class:"border"},[a("code",{class:"language-html"},U(zr))])]),a("article",{id:"four-columns",class:"card column"},[Sr,a("pre",{class:"border"},[a("code",{class:"language-html"},U(Yr))])]),a("article",{id:"five-columns",class:"card column"},[Or,a("pre",{class:"border"},[a("code",{class:"language-html"},U(Xr))])]),a("article",{id:"six-columns",class:"card column"},[$r,a("pre",{class:"border"},[a("code",{class:"language-html"},U(Jr))])]),a("article",{id:"twelve-columns",class:"card column"},[qr,a("pre",{class:"border"},[a("code",{class:"language-html"},U(Vr))])]),a("article",{id:"aside",class:"card column"},[Mr,a("pre",{class:"border"},[a("code",{class:"language-html"},U(Qr))]),Hr,Fr]),a("article",{id:"nav",class:"card column"},[Lr,a("pre",{class:"border"},[a("code",{class:"language-html"},U(Gr))])]),a("article",{id:"footer",class:"card column"},[Nr,a("pre",{class:"border"},[a("code",{class:"language-html"},U(Zr))])]),a("article",{id:"nested",class:"card column"},[jr,a("pre",{class:"border mt-0"},[a("code",{class:"language-html"},U(td))])]),Dr,a("article",{id:"table",class:"card column"},[Ur,a("pre",{class:"border"},[a("code",{class:"language-html"},U(ed))])])])]))}},sd={class:"flex-layout gap-3"},od={class:"sidebar"},ld={class:"sticky"},ad={class:"flex-layout main mr-3 gap-3"},id=a("article",{class:"p-3 card"},[a("header",null,[a("h1",null,"Elements")]),a("p",{class:"body"},[P(" All the elements on this page are fully controlled by the theme. "),a("br"),P(" Here you will find the elements from the default theme. ")])],-1),cd=K('<header><h2>Buttons</h2></header><section class="body column no-stretch gap-5"><button>A standard button</button><button disabled>A disabled standard button</button><button class="no-border">A button without a border</button><button class="no-border" disabled> A disabled button without a border </button><button class="no-border outline-hover"> A button with only an outline on hover </button><button class="no-border outline-hover" disabled> A disabled button with only an outline on hover </button><button class="link">A link button</button><button class="link" disabled>A disabled link button</button><button class="link accent">An accent colored link button</button><button class="link danger">A danger colored link button</button><button class="link accent" disabled> A disabled colored link button </button><button class="primary">A button with primary color</button><button class="accent">A button with accent color</button><button class="contrast">A button with contrast color</button><button class="danger">A button with danger color</button><button class="danger" disabled>A disabled colored button</button><p class="mb-0">Icon button</p><button class="icon icon-chevron-right"></button><p class="mb-0">Disabled icon button</p><button class="icon icon-chevron-right" disabled></button></section>',2),rd=a("header",null,[a("h2",null,"Links")],-1),dd=K('<section class="column no-stretch gap-5"><a href>A standard link</a><a href class="accent">An accent colored link</a><a href class="danger">A danger colored link</a><a href class="contrast">A contrast colored link</a><a href class="button">A link as a button</a><a href class="button normal">A link as normal button</a><a href class="button primary">A link as primary button</a><a href class="button accent">A link as accent button</a><a href class="button contrast">A link as contrast button</a><a href class="button danger">A link as danger button</a><a href class="button danger disabled">A link as button with the disabled class</a></section>',1),ud={id:"input",class:"card column"},hd=a("header",null,[a("h2",null,"Inputs")],-1),fd={class:"body"},pd={class:"column no-stretch gap-5"},md=a("h3",{id:"input_checkradio"},"Checkbox and Radiobutton inputs",-1),gd=a("hr",null,null,-1),bd=a("section",{class:"row gap-3"},[a("input",{type:"checkbox"}),a("input",{type:"radio"})],-1),vd=a("h3",{id:"input_buttons",class:"mt-1"},"Button inputs",-1),yd=a("hr",null,null,-1),wd=a("input",{type:"button",value:"submit"},null,-1),_d=a("input",{type:"reset"},null,-1),xd=a("input",{type:"submit"},null,-1),kd=K('<h3 id="input_text" class="mt-1">Text inputs</h3><hr><table><tbody><tr><td class="pr-1">Email</td><td><input type="email"></td></tr><tr><td class="pr-1">Password</td><td><input type="password"></td></tr><tr><td class="pr-1">Search</td><td><input type="search"></td></tr><tr><td class="pr-1">Text</td><td><input type="text"></td></tr><tr><td class="pr-1">Url</td><td><input type="url"></td></tr></tbody></table>',3),Ed=K('<h3 id="input_numbers" class="mt-1">Number inputs</h3><hr><table><tbody><tr><td class="pr-1">Number</td><td><input type="number"></td></tr><tr><td class="pr-1">Range</td><td><input type="range" class="w-100"></td></tr><tr><td class="pr-1">Tel (Phonenumber)</td><td><input type="tel"></td></tr></tbody></table>',3),Ad=K('<h3 id="input_datetime" class="mt-1">Date / Datetime inputs</h3><hr><table><tbody><tr><td class="pr-1">Date</td><td><input type="date"></td></tr><tr><td class="pr-1">Datetime local</td><td><input type="datetime-local"></td></tr><tr><td class="pr-1">Time</td><td><input type="time"></td></tr><tr><td class="pr-1">Month</td><td><input type="month"></td></tr><tr><td class="pr-1">Week</td><td><input type="week"></td></tr></tbody></table>',3),Cd=K('<h3 id="input_others" class="mt-1">Other inputs</h3><hr><table><tbody><tr><td class="pr-1">Color</td><td><input type="color"></td></tr><tr><td class="pr-1">File</td><td><input type="file"></td></tr><tr><td class="pr-1">Image</td><td><input type="image"></td></tr><tr><td class="pr-1">Hidden</td><td><input type="hidden"></td></tr></tbody></table>',3),Id=K('<h3 id="input_modifiers" class="mt-1">Input modifiers</h3><hr><p> You can create a very minimalistic input by adding the <code class="language-html">underline</code> class. </p><input type="text" class="underline" placeholder="Input with underline class"><p> The invalid and disabled attributes work on all inputs, even textarea. </p><p> To disable an input add the <code class="language-html">disabled</code> attribute: </p><input type="text" disabled><input type="range" disabled><input type="number" disabled><input type="button" value="submit" disabled><p> To show that an input is invalid add the <code class="language-html">invalid</code> attribute <br> adding <code class="language-html">invalid=&quot;false&quot;</code> attribute will make it normal again (visual only): </p>',11),Bd=a("input",{type:"text",invalid:""},null,-1),Td=a("input",{type:"text",invalid:"false"},null,-1),Rd=a("p",null,[P(" When you want to use the "),a("code",{class:"language-html"},"required"),P(" attribute on an input, you need to add the "),a("code",{class:"language-html"},'invalid="false"'),P(" attribute as well "),a("br"),P(" and alter the state of the field with JavaScript when using client side validation to avoid the box being marked invalid directly when rendered. ")],-1),Pd=a("p",null,[P(" Input with "),a("code",{class:"language-html"},'required invalid="false"'),P(" attributes: ")],-1),Sd=a("input",{type:"text",invalid:"false",required:""},null,-1),Od=a("p",null,[P(" Input with "),a("code",{class:"language-html"},'required invalid="true"'),P(" attributes: ")],-1),$d=a("input",{type:"text",invalid:"true",required:""},null,-1),qd=a("p",null,[P(" Input with only "),a("code",{class:"language-html"},"required"),P(" attribute: ")],-1),Md=a("input",{type:"text",required:""},null,-1),Hd=a("p",null," Or type a letter in the number input to trigger the webbrowsers built-in invalidation: ",-1),Fd=a("input",{type:"number"},null,-1),Ld={id:"select",class:"card column"},Nd=a("header",null,[a("h2",null,"Select")],-1),jd={class:"body no-stretch column"},Dd=K('<p>A default select element:</p><select name="fruits" id="fruit"><option value="apple">Apple</option><option value="banana">Banana</option><option value="pineapple">Pineapple</option><option value="cherry">Cherry</option></select><p>A disabled select element:</p><select name="fruits" id="fruit" disabled><option value="apple">Apple</option><option value="banana">Banana</option><option value="pineapple">Pineapple</option><option value="cherry">Cherry</option></select>',4),Ud={id:"text-area",class:"card column"},Wd=a("header",null,[a("h2",null,"Textarea")],-1),Kd={class:"body column no-stretch"},zd=a("p",null,"A default resizable textarea element:",-1),Yd=a("textarea",{cols:"30",rows:"10"},null,-1),Xd=a("p",null,"A disabled textarea element:",-1),Jd=a("textarea",{cols:"6",rows:"3",disabled:""},null,-1),Vd=a("p",null,"An invalid textarea element:",-1),Qd=a("textarea",{cols:"6",rows:"3",invalid:""},null,-1),Gd=a("header",null,[a("h2",null,"Labels")],-1),Zd=a("p",null,[P(" Labels render with a "),a("code",{class:"language-css"},"font-size: 1.2rem;"),P(" which results in 12px by default. "),a("br"),P(" And have a "),a("code",{class:"language-css"},"font-weigth: 600;")],-1),tu=a("label",null,"A standard label",-1),eu=a("br",null,null,-1),nu=a("p",null," Their presentation can be altered by a some classes or attributes: ",-1),su=a("label",{class:"required"},"Label with required class",-1),ou=a("br",null,null,-1),lu=a("label",{class:"disabled"},"Label with disabled class",-1),au=K('<br><p> If you add <code class="language-html">tabindex=&quot;0&quot;</code> as well as the required class, you get a tooltip which can be focussed: </p><label class="required" tabindex="0">This is a required field with a tooltip</label><br><label data-tooltip="Custom required text" required>Label with custom tooltip text attribute</label><div class="row"><label class="required ml-auto" tabindex="0">This is a required field with a tooltip on the edge which illustrates it does not move around when hovered</label><br></div><p> You can modify the tooltip by providing <code class="language-html">data-tooltip=&quot;Your text&quot;</code></p><label class="required" tabindex="0" data-tooltip="Custom tooltip text.">This is a required field with a custom tooltip</label><br><p> You can also use the attribute <code class="language-html">disabled</code> / <code class="language-html">disabled=&quot;true&quot;</code> or <code class="language-html">required</code> / <code class="language-html">required=&quot;true&quot;</code> for the same effect of adding the class. <br> You can then set these to false to remove the effect. <br><code class="language-html">disabled=&quot;false&quot;</code> and <code class="language-html">required=&quot;false&quot;</code> respectively. </p>',10),iu=`<button>A standard button</button>
+<button disabled>A disabled standard button</button>
+<button class='no-border'>A button without a border</button>
+<button class='no-border' disabled>A disabled button without a border</button>
+<button class='no-border outline-hover'>A button with only an outline on hover</button>
+<button class='no-border outline-hover' disabled>A disabled button with only an outline on hover</button>
+<button class='link'>A link button</button>
+<button class='link' disabled>A disabled link button</button>
+<button class='link accent'>An accent colored link button</button>
+<button class='link danger'>A danger colored link button</button>
+<button class='link accent' disabled>A disabled colored link button</button>
+<button class='primary'>A button with primary color</button>
+<button class='accent'>A button with accent color</button>
+<button class='contrast'>A button with contrast color</button>
+<button class='danger'>A button with danger color</button>
+<button class='danger' disabled>A disabled colored button</button>
+<button class="icon icon-chevron-right"></button>
+<button
+  class="icon icon-chevron-right"
+  disabled
+></button>`,cu=`<a href>A standard link</a>
+<a href class="accent">An accent colored link</a>
+<a href class="danger">A danger colored link</a>
+<a href class="contrast">A contrast colored link</a>
+<a href class="button">A link as a button</a>
+<a href class="button normal">A link as normal button</a>
+<a href class="button primary">A link as primary button</a>
+<a href class="button accent">A link as accent button</a>
+<a href class="button danger">A link as danger button</a>
+<a href class="button danger disabled">A link as button with the disabled class</a>`,ru=`<input type="button" value="submit" />
+<input type="reset" />
+<input type="submit" />`,du=`<input type="checkbox" />
+<input type="radio" />`,uu=`<textarea cols="30" rows="10"></textarea>
+<textarea cols="6" rows="3" disabled></textarea>
+<textarea cols="6" rows="3" invalid></textarea>`,hu=`<select name="fruits" id="fruit">
+      <option value="apple">Apple</option>
+      <option value="banana">Banana</option>
+      <option value="pineapple">Pineapple</option>
+      <option value="cherry">Cherry</option>
+</select>
+<select name="fruits" id="fruit" disabled>
+      <option value="apple">Apple</option>
+      <option value="banana">Banana</option>
+      <option value="pineapple">Pineapple</option>
+      <option value="cherry">Cherry</option>
+</select>`,fu=`<input type="email" />
+<input type="password" />
+<input type="search" />
+<input type="text" />
+<input type="url" />`,pu=`<input type="number" />
+<input type="range" />
+<input type="tel" />`,mu=`<input type="date" />
+<input type="datetime-local" />
+<input type="time" />
+<input type="month" />
+<input type="week" />`,gu=`<input type="color" />
+<input type="file" />
+<input type="image" />
+<input type="hidden" />`,bu=`<input type="text" class="underline" placeholder="Input with underline class" />
+<input type="range" disabled />
+<input type="number" disabled />
+<input type="text" invalid />
+<input type="text" invalid="false" />
+<input type="button" value="submit" disabled>
+<input type="text" invalid="false" required />`,vu=`<label>A standard label</label>
+
+<label class="required">Label with required class</label>
+<label required>Label with required attribute</label>
+<label data-tooltip="Custom required text" required>Label with custom tooltip text attribute</label>
+<label required="false">Not required label</label>
+
+<label class="disabled">Label with disabled class</label>
+<label disabled>Label with disabled attribute</label>
+<label disabled="false">Not disabled label</label>
+
+<label class="required" tabindex="0">This is a required field with a tooltip</label>
+<label required tabindex="0">This is a required field with a tooltip</label>
+<label required tabindex="0" data-tooltip="Custom text">Required label with custom tooltip</label>
+`,yu={__name:"ButtonsAndInputs",setup(t){const e=_t("scrollTo");return Pe(()=>{Prism.highlightAll()}),(n,s)=>{const o=bs("router-link");return zt(),Yt("main",sd,[a("aside",od,[a("nav",ld,[a("ul",null,[a("li",null,[a("a",{onClick:s[0]||(s[0]=l=>L(e)("buttons"))},"Buttons")]),a("li",null,[a("a",{onClick:s[1]||(s[1]=l=>L(e)("links"))},"Links")]),a("li",null,[a("a",{onClick:s[2]||(s[2]=l=>L(e)("input"))},"Inputs")]),a("li",null,[a("a",{onClick:s[3]||(s[3]=l=>L(e)("select"))},"Select")]),a("li",null,[a("a",{onClick:s[4]||(s[4]=l=>L(e)("text-area"))},"Textarea")]),a("li",null,[a("a",{onClick:s[5]||(s[5]=l=>L(e)("label"))},"Labels")]),a("li",null,[a("a",{onClick:s[6]||(s[6]=l=>L(e)("main-nav")),class:"button normal desktop"},"Back to top")])])])]),a("section",ad,[id,a("article",{id:"buttons",class:"flex-layout card"},[cd,a("pre",{class:"border"},[a("code",{class:"language-html"},U(iu))])]),a("article",{id:"links",class:"card column"},[rd,a("div",{class:"body"},[dd,a("pre",{class:"border mt-5"},[a("code",{class:"language-html"},U(cu))])])]),a("article",ud,[hd,a("nav",null,[a("ul",null,[a("li",null,[a("a",{onClick:s[7]||(s[7]=l=>L(e)("input_checkradio"))},"Checkbox and Radiobutton")]),a("li",null,[a("a",{onClick:s[8]||(s[8]=l=>L(e)("input_buttons"))},"Button inputs")]),a("li",null,[a("a",{onClick:s[9]||(s[9]=l=>L(e)("input_text"))},"Text inputs")]),a("li",null,[a("a",{onClick:s[10]||(s[10]=l=>L(e)("input_numbers"))},"Number inputs")]),a("li",null,[a("a",{onClick:s[11]||(s[11]=l=>L(e)("input_datetime"))},"Date, Time and Datetime inputs")]),a("li",null,[a("a",{onClick:s[12]||(s[12]=l=>L(e)("input_others"))},"Other inputs")]),a("li",null,[a("a",{onClick:s[13]||(s[13]=l=>L(e)("input_modifiers"))},"Input modifiers")])])]),a("div",fd,[a("section",pd,[md,gd,bd,Y(o,{to:"/components/custom-inputs"},{default:at(()=>[P("Click here for custom inputs for checkbox and radiobutton.")]),_:1}),a("pre",{class:"border stretch"},[a("code",{class:"language-html"},U(du))]),vd,yd,wd,_d,xd,a("pre",{class:"border stretch"},[a("code",{class:"language-html"},U(ru))]),kd,a("pre",{class:"border stretch"},[a("code",{class:"language-html"},U(fu))]),Ed,a("pre",{class:"border stretch"},[a("code",{class:"language-html"},U(pu))]),Ad,a("pre",{class:"border stretch"},[a("code",{class:"language-html"},U(mu))]),Cd,a("pre",{class:"border stretch"},[a("code",{class:"language-html"},U(gu))]),Id,Bd,Td,Rd,Pd,Sd,Od,$d,qd,Md,Hd,Fd,a("pre",{class:"border stretch"},[a("code",{class:"language-html"},U(bu))])])])]),a("article",Ld,[Nd,a("div",jd,[Dd,a("pre",{class:"border stretch"},[a("code",{class:"language-html"},U(hu))]),a("div",null,[P(" For modifiers checkout the "),a("a",{onClick:s[14]||(s[14]=l=>L(e)("input_modifiers"))},"input modifiers section")])])]),a("article",Ud,[Wd,a("div",Kd,[zd,Yd,Xd,Jd,Vd,Qd,a("pre",{class:"border stretch"},[a("code",{class:"language-html"},U(uu))]),a("div",null,[P(" For modifiers checkout the "),a("a",{onClick:s[15]||(s[15]=l=>L(e)("input_modifiers"))},"input modifiers section")])])]),a("article",{id:"label",class:"card column"},[Gd,a("div",{class:"body column"},[Zd,tu,P(),eu,nu,su,P(),ou,lu,P(),au,a("pre",{class:"border"},[a("code",{class:"language-html"},U(vu))])])])])])}}};function ge(t,e){document.getElementById(t).addEventListener("click",()=>{const n=document.getElementById(e);n.querySelector(".close").addEventListener("click",s=>{s.target.parentNode.removeAttribute("open")}),n.hasAttribute("open")?n.removeAttribute("open"):n.setAttribute("open","")})}const wu={class:"flex-layout gap-3"},_u={class:"sidebar"},xu={class:"sticky"},ku=a("article",{class:"card column"},[a("header",null,[a("h1",null,"Components")]),a("p",null," Here you will find elements that are more customized and can consist of multiple elements combined. ")],-1),Eu=K('<header><h2>Card</h2></header><p> Apply the <code class="language-html">class=&quot;card&quot;</code> to get a nice card effect which has built-in styling for header and footer. <br> Use the <code class="language-html">class=&quot;body&quot;</code> class to apply the body styling. <br><br> If you put a <code class="language-html">nav</code> element above the element with the body it will not have a box-shadow. <br> For the examples I have set a max-width of 25%. </p><article class="card w-25">This is a default card</article><article class="card w-25"><header><h3>Card header</h3></header><section class="body"><p> This is a default card with a <code class="language-html">&lt;header&gt;</code> and a <code class="language-html">&lt;footer&gt;</code> element. </p></section><footer>Card footer</footer></article><hr class="mt-3"><h4>More specialized cards:</h4><section class="flex-layout four-columns no-stretch"><article class="card normal"><header><h3>Header</h3></header><section class="body"><p><code class="language-html">class=&quot;card normal&quot;</code></p></section><footer>Footer</footer></article><article class="card primary"><header><h3>Header</h3></header><section class="body"><p><code class="language-html">class=&quot;card primary&quot;</code></p></section><footer>Footer</footer></article><article class="card accent"><header><h3>Header</h3></header><section class="body"><p><code class="language-html">class=&quot;card accent&quot;</code></p></section><footer>Footer</footer></article><article class="card danger"><header><h3>Header</h3></header><section class="body"><p><code class="language-html">class=&quot;card danger&quot;</code></p><p>Lorum ipsum text</p></section><footer>Footer</footer></article><article class="card primary inverse"><header><h3>Header</h3></header><section class="body"><p><code class="language-html">class=&quot;card primary inverse&quot;</code></p></section><footer>Footer</footer></article><article class="card accent inverse"><header><h3>Header</h3></header><section class="body"><p><code class="language-html">class=&quot;card accent inverse&quot;</code></p></section><footer>Footer</footer></article><article class="card danger inverse"><header><h3>Header</h3></header><section class="body"><p><code class="language-html">class=&quot;card danger inverse&quot;</code></p></section><footer>Footer</footer></article></section>',7),Au=K('<header><h2>Notification</h2></header><div class="body column"><p> A notification is any element that has the <code class="language-html">class=&quot;notification&quot;</code> <br> You can position the notification with the following classes: <br><br><code class="language-html">top</code> <br><code class="language-html">bottom</code> <br><code class="language-html">left</code> <br><code class="language-html">right</code> <br><code class="language-html">center</code> <br><br> When you use top, it will be fixed on the top with the sticky property and will slide in from above. <br> When you use bottom, it will do the same but then on the bottom. <br> You can add or remove the <code class="language-html">open</code> attribute to show / hide the notification. </p><p> If you add a lucide font icon, it will have a contrasting color, matching the notification. </p><section class="row"><button id="basic-trigger">Show basic notification top left</button><button id="basic-center-trigger"> Show basic notification top center </button><button id="basic-center-bottom-trigger"> Show basic notification bottom center </button><button id="warning-trigger"> Show warning notification top right </button><button id="success-trigger"> Show success notification bottom right </button><button id="error-trigger"> Show error notification bottom left </button></section><div id="basic-notification" class="notification top left"><span>Basic notification</span> <button class="close">X</button></div><div id="basic-notification-center" class="notification top center"><span>Basic top center notification</span><button class="close">X</button></div><div id="basic-notification-bottom-center" class="notification bottom center"><span>Basic bottom center notification</span><button class="close">X</button></div><div id="warning-notification" class="notification warning top right"><span>Warning notification</span> <button class="close">X</button></div><div id="success-notification" class="notification success bottom right"><span>Success notification</span> <button class="close">X</button></div><div id="error-notification" class="notification error bottom left"><div class="icon icon-alert-triangle mr-3"></div><span>Error notification</span> <button class="close">X</button></div></div><p>Example:</p>',3),Cu=a("p",null,"Javascript example:",-1),Iu=a("header",null,[a("h2",null,"Modal")],-1),Bu=a("div",{class:"body column no-stretch"},[a("p",null,[P(" A modal is a "),a("code",{class:"language-html"},"<dialog>"),P(" with a "),a("code",{class:"language-html"},'class="modal"'),P(". "),a("br"),P(" It opens when it has a property "),a("code",{class:"language-html"},"open"),P(" or "),a("code",{class:"language-html"},'open="true"'),P(". "),a("br"),P(" It closes when the property is removed or set to false. "),a("br"),P(" You could use a card (article) for the modal as demonstrated below. ")]),a("button",{onclick:"document.getElementById('demo-modal').setAttribute('open', true)"}," Open modal "),a("dialog",{id:"demo-modal",class:"modal",onclick:"this.removeAttribute('open')"},[a("article",{class:"card normal w-25"},[a("header",null,[a("h3",null,"Message")]),a("div",{class:"body"},"I am a modal."),a("footer",null,[a("button",{onclick:"document.getElementById('demo-modal').setAttribute('open', false)"}," Close ")])])]),a("p",null,[P(" If you want to disable scroll on the body add the "),a("code",{class:"language-html"},"no-scroll"),P(" class on the body. ")]),a("button",{onclick:`document.getElementById('demo-modal-noscroll').setAttribute('open', true); \r
+                   document.getElementsByTagName('body')[0].classList.add('no-scroll')`}," Open modal with no-scroll applied. "),a("dialog",{id:"demo-modal-noscroll",class:"modal",onclick:`this.removeAttribute('open'); \r
+                   document.getElementsByTagName('body')[0].classList.remove('no-scroll')`},[a("article",{class:"card normal w-25"},[a("header",null,[a("h3",null,"Message")]),a("div",{class:"body"},"I am a modal."),a("footer",null,[a("button",{onclick:`document.getElementById('demo-modal-noscroll').setAttribute('open', false); \r
+                         document.getElementsByTagName('body')[0].classList.remove('no-scroll')`}," Close ")])])])],-1),Tu=a("small",null,[a("b",null,"N.B."),P(" You can also set a close on the outside of the modal. As done in the example. Try it!")],-1),Ru=a("header",null,[a("h2",null,"Breadcrumb")],-1),Pu=a("p",null,"A simple breadcrumb interface:",-1),Su=a("nav",{"aria-label":"Breadcrumb"},[a("ul",null,[a("li",null,[a("a",{href:"#breadcrumb"},"Home")]),a("li",null,[a("a",{href:"#breadcrumb"},"Workshop")]),a("li",{"aria-current":"page"},"Sawbench")])],-1),Ou=K('<header><h2>Dropdown</h2></header><p> There are two types of dropdown in JESS, the accordion and the dropdown menu. Both use a </p><p> Both use a <code class="language-html">&lt;details&gt;</code> element, with a nested <code class="language-html">&lt;summary&gt;</code> followed by the contents that should be visible on click. <br> The contents inside the <code class="language-html">&lt;summary&gt;</code> tag is the visible part when it is closed. </p><p>Default:</p><details><summary>Menu 1</summary><nav class="dropdown-container"><ul><li><a href>Item one</a></li><li><a href>Item two</a></li><li><a href>Item three</a></li><li><a href>Item four</a></li><li><a href>Item five</a></li></ul></nav></details><p> With the <code class="language-html">align-right</code> class it can be used on the other side of the screen: </p><details class="align-right ml-auto"><summary>Right side menu</summary><nav class="dropdown-container"><ul><li><a href>Item one</a></li><li><a href>Item two</a></li><li><a href>Item three</a></li><li><a href>Item four</a></li><li><a href>Item five</a></li></ul></nav></details><p>Or with an icon:</p><details class="icon-dropdown"><summary class="icon-menu"></summary><nav class="dropdown-container"><ul><li><a href>Item one</a></li><li><a href>Item two</a></li><li><a href>Item three</a></li><li><a href>Item four</a></li><li><a href>Item five</a></li></ul></nav></details><p>With the <code class="language-html">accordion</code> class:</p><details class="accordion"><summary>Lorum ipsum story</summary><p> Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry&#39;s standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to make a type specimen book. It has survived not only five centuries, but also the leap into electronic typesetting, remaining essentially unchanged. </p></details><p> As a button by using the the <code class="language-html">button</code> class: </p><details class="button"><summary>Button dropdown</summary><nav class="dropdown-container"><ul><li><a href>Item one</a></li><li><a href>Item two</a></li><li><a href>Item three</a></li><li><a href>Item four</a></li><li><a href>Item five</a></li></ul></nav></details>',13),$u=K('<header><h2>Custom inputs</h2></header><h3 class="pt-1">Checkbox</h3><hr><label class="custom-input"> Custom checkbox with full color control. <input type="checkbox" checked="checked"><span class="checkmark"></span></label><label class="custom-input"> Unchecked version. <input type="checkbox"><span class="checkmark"></span></label><label class="custom-input"> Disabled version. <input type="checkbox" name="custom" disabled><span class="checkmark"></span></label><label class="custom-input"> Disabled checked version. <input type="checkbox" name="custom" disabled checked><span class="checkmark"></span></label><p> You can use Javascript to add the indeterminate property to make it look indeterminate </p><label class="custom-input"> Indeterminate checkbox <input type="checkbox" id="indeterminateExample"><span class="checkmark"></span></label><label class="custom-input"> Disabled indeterminate checkbox <input type="checkbox" id="indeterminateExampleDisabled" disabled><span class="checkmark"></span></label><br><p>Or with the <code class="language-html">label-left</code> class:</p><label class="custom-input label-left"> Custom checkbox with the label on the left. <input type="checkbox" checked="checked"><span class="checkmark"></span></label>',13),qu=K('<br><h3 class="pt-1">Radiobutton</h3><hr><label class="custom-input"> A checked custom radiobutton with full color control. <input type="radio" name="custom" checked><span class="checkmark"></span></label><label class="custom-input"> Unchecked version. <input type="radio" name="custom"><span class="checkmark"></span></label><label class="custom-input"> Disabled version. <input type="radio" name="custom" disabled><span class="checkmark"></span></label><label class="custom-input"> Disabled checked version. <input type="radio" name="custom-disabled" disabled checked><span class="checkmark"></span></label><p>Or with the <code class="language-html">label-left</code> class:</p><label class="custom-input label-left"> Custom radiobutton with label on the left. <input type="radio" name="custom"><span class="checkmark"></span></label>',9),Mu=K('<br><h3>Switch</h3><hr><p>A switch is a checkbox with some additional styling.</p><p>A normal switch:</p><label class="switch"><input type="checkbox"><span class="toggle"></span><span class="toggle-background"></span></label><p>An unchecked disabled switch:</p><label class="switch"><input type="checkbox" disabled><span class="toggle"></span><span class="toggle-background"></span></label><p>A checked disabled switch:</p><label class="switch"><input type="checkbox" checked disabled><span class="toggle"></span><span class="toggle-background"></span></label><p> A switch with <code class="lang-html">data-label=&quot;Off&quot;</code> and <code class="lang-html">data-label-on=&quot;On&quot;</code>: </p><label class="switch"><input type="checkbox"><span class="toggle"></span><span class="toggle-background" data-label="Off" data-label-on="On"></span></label><p> A switch with <code class="language-html">data-label=&quot;Off&quot;</code> and <code class="language-html">data-label-on=&quot;On&quot;</code> and class <code class="language-html">label-left</code>: </p><label class="switch switch-margin"><input type="checkbox"><span class="toggle"></span><span class="toggle-background label-left" data-label="Off" data-label-on="On"></span></label><p> A switch with only <code class="language-html">data-label</code>, shows only a label when unchecked:<br><small>label-left can also be applied</small></p><label class="switch"><input type="checkbox"><span class="toggle"></span><span class="toggle-background" data-label="Off state only"></span></label><p> A switch with only <code class="language-html">data-label-on</code>, shows only a label when checked:<br><small>label-left can also be applied</small></p><label class="switch"><input type="checkbox" checked><span class="toggle"></span><span class="toggle-background" data-label-on="On state only"></span></label>',18),Hu=a("hr",null,null,-1),Fu=a("small",null,[a("b",null,"N.B."),P(" The data-label attribute is rendered as a pseudo element, make sure it will fit by applying the correct margins around the switch. ")],-1),Lu=K('<header><h2>Tooltips</h2></header><div class="body column"><p> You can add a tooltip by using the property <code class="language-html">data-tooltip</code></p> then you can add the tooltip text to the data-property like so: <code class="language-html">data-tooltip=&quot;I am a tooltip&quot;</code><br><label data-tooltip="Yes! I show on the top-left">Hover me for a top left tooltip</label><label class="tooltip-right self-auto" data-tooltip="Yes! I show on the top-right">Hover me for a top right tooltip</label><label class="tooltip-bottom" data-tooltip="Yes! I show on the bottom-left">Hover me for a bottom left tooltip</label><label class="tooltip-right tooltip-bottom self-auto" data-tooltip="Yes! I show on the bottom-right">Hover me for a bottom right tooltip</label></div>',2),Nu=`<article class="card normal">
+    <header>
+      <h3>Header</h3>
+    </header>
+    <section class="body">
+      <p>
+        <code class="language-html">class="card normal"</code>
+      </p>
+    </section>
+    <footer>Footer</footer>
+</article>`,ju=`<button id="basic-trigger">Show basic notification top left</button>
+
+<button id="basic-center-trigger">Show basic notification top center</button>
+
+<button id="basic-center-bottom-trigger">Show basic notification bottom center</button>
+
+<button id="warning-trigger">
+  Show warning notification top right
+</button>
+
+<button id="success-trigger">
+  Show success notification bottom right
+</button>
+
+<button id="error-trigger">
+  Show error notification bottom left
+</button>
+
+<div id="basic-notification" class="notification top left">
+  <span>Basic notification</span>
+  <button class="close">X</button>
+</div>
+
+<div id="basic-notification-center" class="notification top center">
+  <span>Basic top center notification</span>
+  <button class="close">X</button>
+</div>
+
+<div id="basic-notification-bottom-center" class="notification bottom center">
+  <span>Basic bottom center notification</span>
+  <button class="close">X</button>
+</div>
+
+<div id="warning-notification" class="notification warning top right">
+  <span>Warning notification</span>
+  <button class="close">X</button>
+</div>
+
+<div id="success-notification" class="notification success bottom right">
+  <span>Success notification</span>
+  <button class="close">X</button>
+</div>
+
+<div id="error-notification" class="notification error bottom left">
+  <div class="icon icon-alert-triangle mr-3"></div>
+  <span>Error notification</span>
+  <button class="close">X</button>
+</div>`,Du=`function setupNotificationTrigger(buttonId, notificationId) {
+
+    document.getElementById(buttonId).addEventListener("click", () => {
+        const notificationElement = document.getElementById(notificationId);
+
+        notificationElement.querySelector(".close").addEventListener("click", (e) => {
+            e.target.parentNode.removeAttribute("open");
+        });
+
+        if (notificationElement.hasAttribute("open")) {
+            notificationElement.removeAttribute("open");
+        }
+        else {
+            notificationElement.setAttribute("open", "");
+        }
+    });
+};`,Uu=`<button
+    onclick="document.getElementById("demo-modal").setAttribute("open", true)"
+    >
+    Open modal
+</button>
+
+<dialog
+    id="demo-modal"
+    class="modal"
+    onclick="this.removeAttribute("open")"
+    >
+    <article class="card normal w-25">
+        <header>
+            <h3>Message</h3>
+        </header>
+        <div class="body">I am a modal.</div>
+        <footer>
+            <button
+                onclick="document.getElementById("demo-modal").setAttribute("open", false)"
+            >
+                Close
+            </button>
+        </footer>
+    </article>
+</dialog>
+
+<button
+    onclick="document.getElementById("demo-modal-noscroll").setAttribute("open", true); 
+             document.getElementsByTagName("body")[0].classList.add("no-scroll")"
+    >
+    Open modal with no-scroll applied.
+</button>
+
+<dialog
+    id="demo-modal-noscroll"
+    class="modal"
+    onclick="this.removeAttribute("open"); 
+             document.getElementsByTagName("body")[0].classList.remove("no-scroll")"
+    >
+    <article class="card normal w-25">
+        <header>
+            <h3>Message</h3>
+        </header>
+        <div class="body">I am a modal.</div>
+        <footer>
+            <button
+                onclick="document.getElementById("demo-modal-noscroll").setAttribute("open", false); 
+                         document.getElementsByTagName("body")[0].classList.remove("no-scroll")"
+                >
+                Close
+            </button>
+        </footer>
+    </article>
+</dialog>`,Wu=`<nav aria-label="Breadcrumb">
+  <ul>
+    <li><a href="">Home</a></li>
+    <li><a href="">Workshop</a></li>
+    <li aria-current="page">Sawbench</li>
+  </ul>
+</nav>`,Ku=`<details>
+  <summary>Menu 1</summary>
+  <nav class='dropdown-container'>
+    <ul>
+      <li><a href>Item one</a></li>
+      <li><a href>Item two</a></li>
+      <li><a href>Item three</a></li>
+      <li><a href>Item four</a></li>
+      <li><a href>Item five</a></li>
+    </ul>
+  </nav>
+</details>
+
+<details class='align-right ml-auto'>
+  <summary>Right side menu</summary>
+  <nav class='dropdown-container'>
+    <ul>
+      <li><a href>Item one</a></li>
+      <li><a href>Item two</a></li>
+      <li><a href>Item three</a></li>
+      <li><a href>Item four</a></li>
+      <li><a href>Item five</a></li>
+    </ul>
+  </nav>
+</details>
+
+<details class='icon-dropdown'>
+  <summary class='icon-menu'></summary>
+  <nav class='dropdown-container'>
+    <ul>
+      <li><a href>Item one</a></li>
+      <li><a href>Item two</a></li>
+      <li><a href>Item three</a></li>
+      <li><a href>Item four</a></li>
+      <li><a href>Item five</a></li>
+    </ul>
+  </nav>
+</details>
+
+<p>With the <code class='language-html'>accordion</code> class:</p>
+<details class='accordion'>
+  <summary>Lorum ipsum story</summary>
+  <p>Lorem Ipsum is simply dummy text of the printing and typesetting
+    industry. Lorem Ipsum has been the industry's standard dummy text ever
+    since the 1500s, when an unknown printer took a galley of type and
+    scrambled it to make a type specimen book. It has survived not only
+    five centuries, but also the leap into electronic typesetting,
+    remaining essentially unchanged.</p>
+</details>
+
+<details class='button'>
+  <summary>Button dropdown</summary>
+  <nav class='dropdown-container'>
+    <ul>
+      <li><a href>Item one</a></li>
+      <li><a href>Item two</a></li>
+      <li><a href>Item three</a></li>
+      <li><a href>Item four</a></li>
+      <li><a href>Item five</a></li>
+    </ul>
+  </nav>
+</details>`,zu=`<label class="custom-input">
+  Custom checkbox with full color control.
+  <input type="checkbox" checked="checked" />
+  <span class="checkmark"></span>
+</label>
+
+<label class="custom-input ml-1">
+  Unchecked version.
+  <input type="checkbox" />
+  <span class="checkmark"></span>
+</label>
+
+<label class="custom-input margin">
+  Disabled version.
+  <input type="checkbox" name="custom" disabled />
+  <span class="checkmark"></span>
+</label>
+
+<label class="custom-input label-left">
+  Custom checkbox with the label on the left.
+  <input type="checkbox" checked="checked" />
+  <span class="checkmark"></span>
+</label>`,Yu=`<label class="custom-input">
+  Custom radiobutton with full color control.
+  <input type="radio" checked="checked" name="custom" />
+  <span class="checkmark"></span>
+</label>
+
+<label class="custom-input ml-1">
+  Unchecked version.
+  <input type="radio" name="custom" />
+  <span class="checkmark"></span>
+</label>
+
+<label class="custom-input margin">
+  Disabled version.
+  <input type="radio" name="custom" disabled />
+  <span class="checkmark"></span>
+</label>
+
+<label class="custom-input label-left">
+  Custom radiobutton with label on the left.
+  <input type="radio" name="custom" />
+  <span class="checkmark"></span>
+</label>`,Xu=`<label class="switch">
+    <input type="checkbox" />
+    <span class="toggle"></span>
+    <span class="toggle-background"></span>
+</label>
+
+<label class="switch">
+    <input type="checkbox" disabled />
+    <span class="toggle"></span>
+    <span class="toggle-background"></span>
+</label>
+
+<label class="switch">
+    <input type="checkbox" checked disabled />
+    <span class="toggle"></span>
+    <span class="toggle-background"></span>
+</label>
+
+<label class="switch">
+    <input type="checkbox" />
+    <span class="toggle"></span>
+    <span
+        class="toggle-background"
+        data-label="Off"
+        data-label-on="On"
+    ></span>
+</label>
+
+<label class="switch">
+    <input type="checkbox" />
+    <span class="toggle"></span>
+    <span
+        class="toggle-background label-left"
+        data-label="Off"
+        data-label-on="On"
+    ></span>
+</label>
+
+<label class="switch">
+    <input type="checkbox" />
+    <span class="toggle"></span>
+    <span class="toggle-background" data-label="Off state only"></span>
+</label>
+
+<label class="switch">
+    <input type="checkbox" checked />
+    <span class="toggle"></span>
+    <span class="toggle-background" data-label-on="On state only"></span>
+</label>`,Ju=`<label data-tooltip="Yes! I show on the top-left">
+  Hover me for a top left tooltip
+</label>
+
+<label class="tooltip-right self-auto" data-tooltip="Yes! I show on the top-right">
+  Hover me for a top right tooltip
+</label>
+
+<label class="tooltip-bottom" data-tooltip="Yes! I show on the bottom-left">
+  Hover me for a bottom left tooltip
+</label>
+
+<label class="tooltip-right tooltip-bottom self-auto" data-tooltip="Yes! I show on the bottom-right">
+  Hover me for a right tooltip
+</label>`,ko={__name:"Components",props:["scrollToId"],setup(t){const e=t,n=_t("scrollTo");return Pe(()=>{if(Prism.highlightAll(),document.getElementById("indeterminateExample").indeterminate=!0,document.getElementById("indeterminateExampleDisabled").indeterminate=!0,ge("basic-trigger","basic-notification"),ge("basic-center-trigger","basic-notification-center"),ge("basic-center-bottom-trigger","basic-notification-bottom-center"),ge("warning-trigger","warning-notification"),ge("success-trigger","success-notification"),ge("error-trigger","error-notification"),e.scrollToId){const s=setTimeout(()=>{n(e.scrollToId),clearTimeout(s)},100)}}),(s,o)=>(zt(),Yt("main",wu,[a("aside",_u,[a("nav",xu,[a("ul",null,[a("li",null,[a("a",{onClick:o[0]||(o[0]=l=>L(n)("card"))},"Card")]),a("li",null,[a("a",{onClick:o[1]||(o[1]=l=>L(n)("notification"))},"Notification")]),a("li",null,[a("a",{onClick:o[2]||(o[2]=l=>L(n)("modal"))},"Modal")]),a("li",null,[a("a",{onClick:o[3]||(o[3]=l=>L(n)("breadcrumb"))},"Breadcrumb")]),a("li",null,[a("a",{onClick:o[4]||(o[4]=l=>L(n)("dropdown"))},"Dropdown")]),a("li",null,[a("a",{onClick:o[5]||(o[5]=l=>L(n)("custom-inputs"))},"Custom inputs")]),a("li",null,[a("a",{onClick:o[6]||(o[6]=l=>L(n)("tooltips"))},"Tooltips")]),a("li",null,[a("a",{onClick:o[7]||(o[7]=l=>L(n)("main-nav")),class:"button normal desktop"},"Back to top")])])])]),a("section",{class:"flex-layout column main mr-3"},[ku,a("article",{class:"card column",id:"card"},[Eu,a("pre",{class:"border stretch mt-1"},[a("code",{class:"language-html"},U(Nu))])]),a("article",{id:"notification",class:"card column"},[Au,a("pre",{class:"border stretch mt-1"},[a("code",{class:"language-html"},U(ju))]),Cu,a("pre",{class:"border stretch mt-1"},[a("code",{class:"language-js"},U(Du))])]),a("article",{id:"modal",class:"card column"},[Iu,Bu,a("pre",{class:"border stretch mt-1"},[a("code",{class:"language-html"},U(Uu))]),Tu]),a("article",{id:"breadcrumb",class:"card column"},[Ru,Pu,Su,a("pre",{class:"border stretch mt-1"},[a("code",{class:"language-html"},U(Wu))])]),a("article",{id:"dropdown",class:"card column"},[Ou,a("pre",{class:"border stretch mt-1"},[a("code",{class:"language-html"},U(Ku))])]),a("article",{id:"custom-inputs",class:"card column no-stretch"},[$u,a("pre",{class:"border stretch mt-1"},[a("code",{class:"language-html"},U(zu))]),qu,a("pre",{class:"border stretch mt-1"},[a("code",{class:"language-html"},U(Yu))]),Mu,a("pre",{class:"border stretch mt-1"},[P("          "),a("code",{class:"language-html"},`
+            `+U(Xu)+`
+          `),P(`
+        `)]),Hu,Fu]),a("article",{id:"tooltips",class:"card"},[Lu,a("pre",{class:"border stretch mt-1"},[a("code",{class:"language-html"},U(Ju))])])])]))}},Vu={class:"flex-layout"},Qu=a("header",null,[a("h1",null,"Typography")],-1),Gu=K('<p> JESS uses the following font-size, based on a 62.5% of the default font-size. <br> This translates to 0.1rem equals 1px. </p><section class="flex-layout two-columns"><div><table class="table"><thead><tr><th>Element</th><th>Specified font-size</th><th>Computed pixels</th></tr></thead><tbody><tr><td>h1</td><td>3.2rem</td><td>32px</td></tr><tr><td>h2</td><td>2.4rem</td><td>24px</td></tr><tr><td>h3</td><td>1.8rem</td><td>18px</td></tr><tr><td>h4</td><td>-</td><td>16px</td></tr><tr><td>h5</td><td>-</td><td>13.28px</td></tr><tr><td>label</td><td>1.4rem</td><td>14px</td></tr><tr><td>small</td><td>-</td><td>13.3333px</td></tr><tr><td>p</td><td>-</td><td>16px</td></tr></tbody></table></div><div><h1>Header 1</h1><h2>Header 2</h2><h3>Header 3</h3><h4>Header 4</h4><h5>Header 5</h5><label>label</label> <br><small>small</small><p>Normal text in a paragraph</p></div></section>',2),Zu=K('<header><h2>Blockquote</h2></header><p> Blockquote will have 100% width. To shorten it, use a width utility class. In this case <code class="language-html">w-16</code> is applied: </p><blockquote class="w-16"> A famous quote of someone, spanning multiple lines. <cite>Anonymous</cite></blockquote><p> Blockquote directly inside an article, which has the <code class="language-html">p-0</code> class and a width utility class: </p><article class="p-0 w-16 card"><blockquote> A famous quote of someone, spanning multiple lines. <cite>Anonymous</cite></blockquote></article>',5),th=K('<section><header><h2>Text modifiers</h2></header><section class="flex-layout two-columns pt-3"><table class="border-left"><thead><th>Tag</th><th>Result</th></thead><tbody><tr><td><code class="language-html">abbr</code></td><td><abbr title="Abbreviation">Abbr.</abbr></td></tr><tr><td><code class="language-html">strong</code><code class="language-html">b</code></td><td><strong>Bold</strong></td></tr><tr><td><code class="language-html">i</code><code class="language-html">em</code><code class="language-html">cite</code></td><td><em>Italic</em></td></tr><tr><td><code class="language-html">del</code></td><td><del>Deleted</del></td></tr><tr><td><code class="language-html">ins</code></td><td><ins>Inserted</ins></td></tr><tr><td><code class="language-html">kbd</code><code class="language-html">code</code></td><td><kbd>Ctrl + S</kbd></td></tr></tbody></table><table class="border-left"><thead><th>Tag</th><th>Result</th></thead><tbody><tr><td><code class="language-html">mark</code></td><td><mark>Highlighted</mark></td></tr><tr><td><code class="language-html">strike</code><code class="language-html">s</code></td><td><s>Strikethrough</s></td></tr><tr><td><code class="language-html">small</code></td><td><small>Small </small></td></tr><tr><td><code class="language-html">sub</code></td><td>Text <sub>sub</sub></td></tr><tr><td><code class="language-html">sup</code></td><td>Text <sup>sup</sup></td></tr><tr><td><code class="language-html">u</code></td><td><u>Underline</u></td></tr></tbody></table></section></section>',1),eh=`<blockquote class="w-16">
+  A famous quote of someone, spanning multiple lines.
+  <cite>Anonymous</cite>
+</blockquote>
+
+<article class="p-0 w-16 card">
+  <blockquote>
+    A famous quote of someone, spanning multiple lines.
+    <cite>Anonymous</cite>
+  </blockquote>
+</article>`,nh={__name:"Typography",setup(t){return Pe(()=>{Prism.highlightAll()}),(e,n)=>(zt(),Yt("main",Vu,[a("section",{class:"px-3"},[a("article",{class:"flex-layout px-3 card"},[Qu,a("div",{class:"body column"},[Gu,a("section",{class:"flex-layout gap-3 my-2"},[Zu,a("pre",{class:"border"},[a("code",{class:"language-html"},U(eh))])]),th])])])]))}},sh={},oh=K('<section class="flex-layout mx-3 gap-3"><article class="card"><header><h1>Utility classes</h1></header><div class="body column"><p> Here are the classes you can use to modify or override default behaviour, which are not mentioned in any of the other categories. </p><section class="flex-layout two-columns"><div><h2 class="mb-1">Border utilities</h2><code class="language-html"> border </code> <br><code class="language-html"> border-bottom </code> <br><code class="language-html"> border-top </code> <br><code class="language-html"> border-left </code> <br><code class="language-html"> border-right </code> <br><h3 class="mt-4 mb-2">Border colour classes</h3><code class="language-html"> border-primary </code> <br><code class="language-html"> border-accent </code> <br><code class="language-html"> border-danger </code> <br><code class="language-html"> border-success </code> <br><code class="language-html"> border-black </code> <br><code class="language-html"> border-white </code> <br><code class="language-html"> border-light </code> <br></div><div><p>Examples:</p><div class="border p-3 mt-3">Element with border</div><div class="border-bottom p-3 my-3"> Element with border-bottom </div><div class="border-top p-3 my-3">Element with border-top</div><div class="border border-primary p-3 mb-3"> Element with border and border-primary </div><div class="border border-accent p-3 mb-3"> Element with border and border-accent </div><div class="border border-danger p-3 mb-3"> Element with border and border-danger </div><div class="border border-success p-3 mb-3"> Element with border and border-success </div><div class="border border-black p-3 mb-3"> Element with border and border-black </div><div class="bg-black p-3"><div class="border border-white p-3"> Element with border and border-white </div></div><div class="border border-light p-3 my-3"> Element with border and border-light </div></div></section><section class="flex-layout two-columns"><div><h2 class="mb-3">Background colour utilities</h2><code class="language-html">bg-grey </code> <br><code class="language-html">bg-grey-dark </code> <br><code class="language-html">bg-black </code> <br><code class="language-html">bg-white </code> <br><code class="language-html">bg-primary </code> <br><code class="language-html">bg-accent </code> <br><code class="language-html">bg-danger </code> <br><code class="language-html">bg-error </code> <br><code class="language-html">bg-warning </code> <br><code class="language-html">bg-success </code> <br></div><div><p>Examples:</p><div class="bg-grey p-3 mt-3">grey background</div><div class="bg-grey-dark p-3">dark grey background</div><div class="bg-black p-3"> black background with auto colour adjustment </div><div class="bg-white p-3">white background</div><div class="bg-primary p-3"> primary background with auto colour adjustment </div><div class="bg-accent p-3"> accent background with auto colour adjustment </div><div class="bg-danger p-3"> danger background with auto colour adjustment </div><div class="bg-warning p-3"> warning background with auto colour adjustment </div><div class="bg-error p-3"> error background with auto colour adjustment </div><div class="bg-success p-3"> success background with auto colour adjustment </div></div></section><section class="flex-layout two-columns"><div><h2 class="mb-3">Text colour utilities</h2><code class="language-html">text-black </code> <br><code class="language-html">text-white </code> <br><code class="language-html">text-primary </code> <br><code class="language-html">text-accent </code> <br><code class="language-html">text-danger </code> <br><code class="language-html">text-success </code> <br><code class="language-html">text-grey </code> <br><small>* these can also be applied on an icon</small></div><div><p>Examples:</p><div class="bg-primary text-black p-3">black text</div><div class="bg-black text-white p-3">white text</div><div class="bg-grey-dark text-primary p-3">primary text</div><div class="bg-black text-accent p-3">accent text</div><div class="bg-accent text-danger p-3">danger text</div><div class="text-success p-3">success text</div><div class="text-grey p-3">grey text</div></div></section><section class="flex-layout two-columns"><div><h2 class="mb-3">Text utilities</h2><table><thead><tr><th>Class</th><th>Description</th></tr></thead><tbody><tr><td><code class="language-html">text-normal</code></td><td>Makes the text normal font-weight</td></tr></tbody></table></div></section></div></article><article id="utility-classes" class="card column"><header><h2>Layout utility classes</h2></header><div class="body column"><table><thead><tr><th>Class</th><th>Description</th></tr></thead><tbody><tr><td><code class="language-html">sticky</code></td><td>Makes it stick to the top when scrolling down.</td></tr><tr><td><code class="language-html">no-scroll</code></td><td> Remove all scrollbars, handy to apply on the body when opening a modal. </td></tr><tr><td><code class="language-html">overflow-horizontal</code></td><td> Make an element crollable on the x-axis if the contents is too large. </td></tr><tr><td><code class="language-html">no-stretch</code></td><td> Apply this to a flex layout item like <code class="language-html">section</code>, <code class="language-html">main</code> or <code class="language-html">fieldset</code> <br> to have its children keep their own size. </td></tr><tr><td><code class="language-html">stretch</code></td><td> Apply this to a flex layout item like <code class="language-html">section</code>, <code class="language-html">main</code> or <code class="language-html">fieldset</code> <br> to have its children take up as much space as there is. </td></tr><tr><td><code class="language-html">row</code></td><td> Make an element display flex with the default row direction. </td></tr><tr><td><code class="language-html">column</code></td><td>Make an element display flex with the column direction</td></tr><tr><td><code class="language-html">align-center</code></td><td> Applies: <code class="language-html">align-items: center;</code></td></tr><tr><td><code class="language-html">align-end</code></td><td> Applies: <code class="language-html">align-items: flex-end;</code></td></tr><tr><td><code class="language-html">align-start </code></td><td> Applies: <code class="language-html">align-items: flex-start; </code></td></tr><tr><td><code class="language-html">align-stretch </code></td><td> Applies: <code class="language-html">align-items: stretch; </code></td></tr><tr><td><code class="language-html">justify-center </code></td><td> Applies: <code class="language-html">justify-content: center; </code></td></tr><tr><td><code class="language-html">justify-around </code></td><td> Applies: <code class="language-html">justify-content: space-around; </code></td></tr><tr><td><code class="language-html">justify-evenly </code></td><td> Applies: <code class="language-html">justify-content: space-evenly; </code></td></tr><tr><td><code class="language-html">justify-end </code></td><td> Applies: <code class="language-html">justify-content: flex-end; </code></td></tr><tr><td><code class="language-html">justify-between </code></td><td> Applies: <code class="language-html">justify-content: space-between; </code></td></tr><tr><td><code class="language-html">justify-start </code></td><td> Applies: <code class="language-html">justify-content: flex-start; </code></td></tr><tr><td><code class="language-html">self-auto</code></td><td>Makes an element align-self and justify-self to start.</td></tr><tr><td><code class="language-html">gap-0</code></td><td>Sets the gap of a flex container to 0.</td></tr><tr><td><code class="language-html">gap-1</code></td><td>Sets the gap of a flex container the same as p-1 / m-1.</td></tr><tr><td><code class="language-html">gap-2</code></td><td>Sets the gap of a flex container the same as p-2 / m-2.</td></tr><tr><td><code class="language-html">gap-3</code></td><td>Sets the gap of a flex container the same as p-3 / m-3.</td></tr><tr><td><code class="language-html">gap-4</code></td><td>Sets the gap of a flex container the same as p-4 / m-4.</td></tr><tr><td><code class="language-html">gap-5</code></td><td>Sets the gap of a flex container the same as p-5 / m-5.</td></tr><tr><td><code class="language-html">no-wrap</code></td><td>Makes a text not wrap.</td></tr><tr><td><code class="language-html">inline-block</code></td><td> Applies: <code class="language-html">display: inline-block;</code></td></tr><tr><td><code class="language-html">inline-flex</code></td><td> Applies: <code class="language-html">display: inline-flex;</code></td></tr><tr><td><code class="language-html">full-height</code></td><td> Applies: <code class="language-html">height: 100vh;</code></td></tr></tbody></table><section class="flex-layout overflow-horizontal mt-5"><h3>Width classes</h3><hr><p> Width classes that set the <code class="language-css">width</code> and <code class="language-css">min-width</code>, with safeguards for the padding <br> and out-of-the-box responsive fallback based on viewport max-width: </p><br><table><thead><tr><th>Class</th><th>CSS</th><th>w-100</th><th>w-75</th><th>w-50</th><th>w-33</th><th>w-25</th><th>w-20</th><th>w-16</th><th>w-8</th></tr></thead><tbody><tr><td><code class="language-css">w-100</code></td><td>100%</td><td>-</td><td>-</td><td>-</td><td>-</td><td>-</td><td>-</td><td>-</td></tr><tr><td><code class="language-css">w-75</code></td><td>calc(75% - 0.5rem)</td><td>415px</td><td>576px</td><td>-</td><td>-</td><td>-</td><td>-</td><td>-</td></tr><tr><td><code class="language-css">w-50</code></td><td>calc(50% - 0.75rem)</td><td>415px</td><td>-</td><td>576px</td><td>-</td><td>-</td><td>-</td><td>-</td><td>-</td></tr><tr><td><code class="language-css">w-33</code></td><td>calc(33.3333% - 1rem)</td><td>415px</td><td>-</td><td>576px</td><td>768px</td><td>-</td><td>-</td><td>-</td><td>-</td></tr><tr><td><code class="language-css">w-25</code></td><td>calc(25% - 1.13rem)</td><td>415px</td><td>-</td><td>576px</td><td>768px</td><td>992px</td><td>-</td><td>-</td><td>-</td></tr><tr><td><code class="language-css">w-20</code></td><td>calc(20% - 1.2rem)</td><td>415px</td><td>-</td><td>576px</td><td>768px</td><td>992px</td><td>1200px</td><td>-</td><td>-</td></tr><tr><td><code class="language-css">w-16</code></td><td>calc(16.6667% - 1.25rem)</td><td>415px</td><td>-</td><td>576px</td><td>768px</td><td>992px</td><td>1200px</td><td>1400px</td><td>-</td></tr><tr><td><code class="language-css">w-8</code></td><td>calc(8.3333% - 1.375rem)</td><td>415px</td><td>-</td><td>-</td><td>576px</td><td>768px</td><td>992px</td><td>1200px</td><td>1400px</td></tr><tr><td><code class="language-css">w-auto</code></td><td colspan="9">Sets min-width, width to auto</td></tr><tr></tr></tbody></table></section><section class="column overflow-horizontal mt-5"><h3>Hide / show for Mobile and Desktop layout</h3><hr><table><thead><tr><th>Class</th><th>415px</th><th>576px</th><th>768px</th><th>992px</th><th>1200px</th><th>1400px</th></tr></thead><tbody><tr><td>mobile</td><td>X</td><td>X</td><td>-</td><td>-</td><td>-</td><td>-</td></tr><tr><td>tablet</td><td>-</td><td>-</td><td>X</td><td>X</td><td>-</td><td>-</td></tr><tr><td>desktop</td><td>-</td><td>-</td><td>-</td><td>-</td><td>X</td><td>X</td></tr><tr><td>small-screen</td><td>X</td><td>X</td><td>X</td><td>-</td><td>-</td><td>-</td></tr><tr><td>large-screen</td><td>-</td><td>-</td><td>-</td><td>X</td><td>X</td><td>X</td></tr></tbody></table><div class="column mt-3"><span>Resize the screen to see it in action:</span><div class="mobile bg-danger p-3">Mobile</div><div class="tablet bg-accent p-3">Tablet</div><div class="desktop bg-primary p-3">desktop</div><div class="small-screen bg-grey p-3">small-screen</div><div class="large-screen bg-grey-dark p-3">large-screen</div></div></section><section class="column overflow-horizontal mt-5"><h3>Height classes</h3><hr><p> Height classes that set the <code class="language-css">height</code> property. </p><br><table><thead><tr><th>Class</th><th>CSS</th></tr></thead><tbody><tr><td><code class="language-css">h-100</code></td><td>100%</td></tr><tr><td><code class="language-css">h-75</code></td><td>75%</td></tr><tr><td><code class="language-css">h-50</code></td><td>50%</td></tr><tr><td><code class="language-css">h-33</code></td><td>33.333333333333333333333333333333%</td></tr><tr><td><code class="language-css">h-25</code></td><td>25%</td></tr><tr><td><code class="language-css">h-20</code></td><td>20%</td></tr><tr><td><code class="language-css">h-16</code></td><td>16.666666666666666666666666666667%</td></tr><tr><td><code class="language-css">h-8</code></td><td>8.3333333333333333333333333333333%</td></tr></tbody></table></section><section class="column overflow-horizontal gap-3 mt-5"><h3>Padding / Margin utilities</h3><hr><p> Each version has a number ranging from 0 and 5 and auto. 0 means no padding or margin for that direction. <br></p><ol><li> is the default distance divided by 3 (so 1.5rem / 3 = 0.5rem) </li><li> is the default distance - the distance of one (so 1.5rem - 0.5rem = 1rem) </li><li>is the default distance</li><li> is the default distance times one and a halve. (1.5rem * 1.5 = 2.25rem) </li><li>is the default distance times two. (1.5rem * 2 = 3rem)</li></ol><hr><table class="table"><thead><tr><th>Padding class</th><th>Margin class</th><th>Top</th><th>Right</th><th>Bottom</th><th>Left</th></tr></thead><tbody><tr><td>p-0</td><td>m-0</td><td>0</td><td>0</td><td>0</td><td>0</td></tr><tr><td>p-1</td><td>m-1</td><td>0.5rem</td><td>0.5rem</td><td>0.5rem</td><td>0.5rem</td></tr><tr><td>pt-auto</td><td>mt-auto</td><td>auto</td><td>-</td><td>-</td><td>-</td></tr><tr><td>pt-0</td><td>mt-0</td><td>0</td><td>-</td><td>-</td><td>-</td></tr><tr><td>pt-1</td><td>mt-1</td><td>0.5rem</td><td>-</td><td>-</td><td>-</td></tr><tr><td>pb-auto</td><td>mb-auto</td><td>-</td><td>-</td><td>auto</td><td>-</td></tr><tr><td>pb-0</td><td>mb-0</td><td>-</td><td>-</td><td>0</td><td>-</td></tr><tr><td>pb-1</td><td>mb-1</td><td>-</td><td>-</td><td>0.5rem</td><td>-</td></tr><tr><td>pl-auto</td><td>ml-auto</td><td>-</td><td>-</td><td>-</td><td>auto</td></tr><tr><td>pl-0</td><td>ml-0</td><td>-</td><td>-</td><td>-</td><td>0</td></tr><tr><td>pl-1</td><td>ml-1</td><td>-</td><td>-</td><td>-</td><td>0.5rem</td></tr><tr><td>pr-auto</td><td>mr-auto</td><td>-</td><td>auto</td><td>-</td><td>-</td></tr><tr><td>pr-0</td><td>mr-0</td><td>-</td><td>0</td><td>-</td><td>-</td></tr><tr><td>pr-1</td><td>mr-1</td><td>-</td><td>0.5rem</td><td>-</td><td>-</td></tr><tr><td>px-auto</td><td>mx-auto</td><td>unchanged</td><td>auto</td><td>unchanged</td><td>auto</td></tr><tr><td>px-0</td><td>mx-0</td><td>unchanged</td><td>0</td><td>unchanged</td><td>0</td></tr><tr><td>px-1</td><td>mx-1</td><td>-</td><td>0.5rem</td><td>-</td><td>0.5rem</td></tr><tr><td>py-auto</td><td>my-auto</td><td>auto</td><td>unchanged</td><td>auto</td><td>unchanged</td></tr><tr><td>py-0</td><td>my-0</td><td>0</td><td>unchanged</td><td>0</td><td>unchanged</td></tr><tr><td>py-1</td><td>my-1</td><td>0.5rem</td><td>-</td><td>0.5rem</td><td>-</td></tr></tbody></table></section></div></article></section>',1),lh=[oh];function ah(t,e){return zt(),Yt("main",null,lh)}const ih=Tn(sh,[["render",ah]]),ch=K('<header><h1>Icons</h1></header><div class="body"><p> JESS is tooled to make use of the Lucide icons font. Please check their <a href="https://lucide.dev/guide/packages/lucide-static">website</a> for a guide. </p><p> It has a base font-size of 2.4rem and you can use the following classes to adjust it: </p><table><thead><tr><th>Class</th><th>Font size</th></tr></thead><tbody><tr><td><code class="language-html">icon-small</code></td><td>2rem</td></tr><tr><td><code class="language-html">icon-medium</code></td><td>3.6rem</td></tr><tr><td><code class="language-html">icon-large</code></td><td>4.6rem</td></tr></tbody></table><p><b>Example:</b></p><div class="icon-arrow-big-down icon-small"></div><div class="icon-arrow-big-down text-primary"></div><div class="icon-arrow-big-down icon-medium icon-success"></div><div class="icon-arrow-big-down icon-large text-danger"></div></div>',2),rh=K('<div><h2 class="mb-3">Icon colour utilities</h2><code class="language-html">icon-grey </code> <br><code class="language-html">icon-grey-dark </code> <br><code class="language-html">icon-primary </code><br><code class="language-html">icon-accent </code><br><code class="language-html">icon-danger </code><br><code class="language-html">icon-success </code><br><code class="language-html">icon-black </code><br><code class="language-html">icon-white </code><br><small>* you can also apply any text-color utility class</small></div><div><p>Examples:</p><div class="row align-center mt-2 p-2"><div class="icon-coffee icon-grey mr-1"></div> Grey icon </div><div class="row align-center mt-2 p-2"><div class="icon-coffee icon-grey-dark mr-1"></div> Dark Grey icon </div><div class="row align-center mt-2 p-2"><div class="icon-coffee icon-primary mr-1"></div> Primary icon </div><div class="row align-center mt-2 p-2"><div class="icon-coffee icon-accent mr-1"></div> Accent icon </div><div class="row align-center mt-2 p-2"><div class="icon-coffee icon-danger mr-1"></div> Danger icon </div><div class="row align-center mt-2 p-2"><div class="icon-coffee icon-success mr-1"></div> Success icon </div><div class="row align-center mt-2 p-2"><div class="icon-coffee icon-black mr-1"></div> Black icon </div><div class="bg-black text-white row align-center mt-2 p-2"><div class="icon-coffee icon-white mr-1"></div> White icon </div></div>',2),dh=`<div class="icon-arrow-big-down icon-small"></div>
+<div class="icon-arrow-big-down text-primary"></div>
+<div class="icon-arrow-big-down icon-medium icon-success"></div>
+<div class="icon-arrow-big-down icon-large text-danger"></div>`,uh=`<div class="icon-coffee icon-grey"></div>
+<div class="icon-coffee icon-grey-dark"></div>
+<div class="icon-coffee icon-primary"></div>
+<div class="icon-coffee icon-accent"></div>
+<div class="icon-coffee icon-danger"></div>
+<div class="icon-coffee icon-success"></div>
+<div class="icon-coffee icon-black"></div>
+  
+<div class="bg-black">
+  <div class="icon-coffee icon-white"></div>
+</div>`,hh={__name:"Icons",setup(t){return Pe(()=>{Prism.highlightAll()}),(e,n)=>(zt(),Yt("main",null,[a("section",{class:"mx-3 column"},[a("article",{class:"card column"},[ch,a("pre",{class:"border"},[a("code",{class:"language-html"},U(dh))])]),a("article",{class:"card flex-layout two-columns justify-between gap-3"},[rh,a("pre",{class:"border w-100"},[a("code",{class:"language-html"},U(uh))])])])]))}},fh=[{path:"/",component:kr},{path:"/layout",component:nd},{path:"/buttons-and-inputs",component:yu},{path:"/components/:scrollToId",component:ko,props:!0},{path:"/components",component:ko},{path:"/typography",component:nh},{path:"/utility-classes",component:ih},{path:"/icons",component:hh}],ph=nr({scrollBehavior(){return{top:0}},history:vc("/JESS/"),routes:fh}),As=Gi(br);As.provide("scrollTo",function(e){document.getElementById(e).scrollIntoView()});As.use(ph);As.mount("#app");
