+++ docs/index.html
@@ -8,8 +8,8 @@
     <link rel="stylesheet" href="/JESS/css/documentation.css" />
     <link rel="shortcut icon" type="image/x-icon" href="/JESS/favicon.ico" />
     <script src="/JESS/js/prism.js" type="text/javascript"></script>
-    <script type="module" crossorigin src="/JESS/assets/index-ef4b4970.js"></script>
-    <link rel="stylesheet" href="/JESS/assets/index-c18937e3.css">
+    <script type="module" crossorigin src="/JESS/assets/index-8165e77e.js"></script>
+    <link rel="stylesheet" href="/JESS/assets/index-bda9eafa.css">
   </head>
   <body class="full-height">
     <div id="app" class="h-100 column"></div>
