+++ jess/utilities/layout-utilities.less
@@ -6,16 +6,18 @@
   */
 
 .w-100 {
-  min-width: 100% !important;
-  width: 100% !important;
+  width: 100%;
+  min-width: 100%;
+  max-width: 100%;
 }
 
 .w-75 {
   .w-100();
 
   @media (min-width: @sm) {
-    width: calc(75% - 0.5rem) !important;
-    min-width: calc(75% - 0.5rem) !important;
+    width: calc(75% - 0.5rem);
+    min-width: calc(75% - 0.5rem);
+    max-width: calc(75% - 0.5rem);
   }
 }
 
@@ -23,8 +25,9 @@
   .w-100();
 
   @media (min-width: @sm) {
-    width: calc(50% - 0.75rem) !important;
-    min-width: calc(50% - 0.75rem) !important;
+    width: calc(50% - 0.75rem);
+    min-width: calc(50% - 0.75rem);
+    max-width: calc(50% - 0.75rem);
   }
 }
 
@@ -32,8 +35,9 @@
   .w-50();
 
   @media (min-width: @md) {
-    width: calc(33.333333333333333333333333333333% - 1rem) !important;
-    min-width: calc(33.333333333333333333333333333333% - 1rem) !important;
+    width: calc(33.333333333333333333333333333333% - 1rem);
+    min-width: calc(33.333333333333333333333333333333% - 1rem);
+    max-width: calc(33.333333333333333333333333333333% - 1rem);
   }
 }
 
@@ -41,8 +45,9 @@
   .w-33();
 
   @media (min-width: @lg) {
-    width: calc(25% - 1.13rem) !important;
-    min-width: calc(25% - 1.13rem) !important;
+    width: calc(25% - 1.13rem);
+    min-width: calc(25% - 1.13rem);
+    max-width: calc(25% - 1.13rem);
   }
 }
 
@@ -50,8 +55,9 @@
   .w-25();
 
   @media (min-width: @xl) {
-    width: calc(20% - 1.2rem) !important;
-    min-width: calc(20% - 1.2rem) !important;
+    width: calc(20% - 1.2rem);
+    min-width: calc(20% - 1.2rem);
+    max-width: calc(20% - 1.2rem);
   }
 }
 
@@ -59,8 +65,9 @@
   .w-20();
 
   @media (min-width: @xxl) {
-    width: calc(16.666666666666666666666666666667% - 1.25rem) !important;
-    min-width: calc(16.666666666666666666666666666667% - 1.25rem) !important;
+    width: calc(16.666666666666666666666666666667% - 1.25rem);
+    min-width: calc(16.666666666666666666666666666667% - 1.25rem);
+    max-width: calc(16.666666666666666666666666666667% - 1.25rem);
   }
 }
 
@@ -68,65 +75,40 @@
   .w-100();
 
   @media (min-width: @sm) {
-    width: calc(33.333333333333333333333333333333% - 1rem) !important;
-    min-width: calc(33.333333333333333333333333333333% - 1rem) !important;
+    width: calc(33.333333333333333333333333333333% - 1rem);
+    min-width: calc(33.333333333333333333333333333333% - 1rem);
+    max-width: calc(33.333333333333333333333333333333% - 1rem);
   }
 
   @media (min-width: @md) {
-    width: calc(25% - 1.13rem) !important;
-    min-width: calc(25% - 1.13rem) !important;
+    width: calc(25% - 1.13rem);
+    min-width: calc(25% - 1.13rem);
+    max-width: calc(25% - 1.13rem);
   }
 
   @media (min-width: @lg) {
-    width: calc(20% - 1.2rem) !important;
-    min-width: calc(20% - 1.2rem) !important;
+    width: calc(20% - 1.2rem);
+    min-width: calc(20% - 1.2rem);
+    max-width: calc(20% - 1.2rem);
   }
 
   @media (min-width: @xl) {
-    width: calc(16.666666666666666666666666666667% - 1.25rem) !important;
-    min-width: calc(16.666666666666666666666666666667% - 1.25rem) !important;
+    width: calc(16.666666666666666666666666666667% - 1.25rem);
+    min-width: calc(16.666666666666666666666666666667% - 1.25rem);
+    max-width: calc(16.666666666666666666666666666667% - 1.25rem);
   }
 
   @media (min-width: @xxl) {
-    width: calc(8.3333333333333333333333333333333% - 1.375rem) !important;
-    min-width: calc(8.3333333333333333333333333333333% - 1.375rem) !important;
+    width: calc(8.3333333333333333333333333333333% - 1.375rem);
+    min-width: calc(8.3333333333333333333333333333333% - 1.375rem);
+    max-width: calc(8.3333333333333333333333333333333% - 1.375rem);
   }
 }
 
 .w-auto {
   width: auto !important;
   min-width: auto !important;
-}
-
-.h-100 {
-  height: 100% !important;
-}
-
-.h-75 {
-  height: 75% !important;
-}
-
-.h-50 {
-  height: 50% !important;
-}
-
-.h-33 {
-  height: 33.333333333333333333333333333333% !important;
-}
-.h-25 {
-  height: 25% !important;
-}
-
-.h-20 {
-  height: 20% !important;
-}
-
-.h-16 {
-  height: 16.666666666666666666666666666667% !important;
-}
-
-.h-8 {
-  height: 8.3333333333333333333333333333333% !important;
+  max-width: auto !important;
 }
 
 .mobile {
@@ -640,4 +622,36 @@
 }
 .pb-5 {
   padding-bottom: @distance-5 !important;
+}
+
+/* height */
+.h-100 {
+  height: 100% !important;
+}
+
+.h-75 {
+  height: 75% !important;
+}
+
+.h-50 {
+  height: 50% !important;
+}
+
+.h-33 {
+  height: 33.333333333333333333333333333333% !important;
+}
+.h-25 {
+  height: 25% !important;
+}
+
+.h-20 {
+  height: 20% !important;
+}
+
+.h-16 {
+  height: 16.666666666666666666666666666667% !important;
+}
+
+.h-8 {
+  height: 8.3333333333333333333333333333333% !important;
 }
