+++ src/partials/MenuBar.vue
@@ -1,17 +1,16 @@
 <template>
-  <nav
-    id="main-nav"
-    class="main-navigation"
-  >
+  <nav id="main-nav" class="main-navigation">
     <ul class="desktop">
       <li>
-        <a href="/">
+        <router-link to="/" class="no-deco">
           <h2>Jelmers Easy Style System</h2>
-        </a>
+        </router-link>
       </li>
       <li><router-link to="/">Home</router-link></li>
       <li><router-link to="/layout">Layout</router-link></li>
-      <li><router-link to="/buttons-and-inputs">Buttons and Inputs</router-link></li>
+      <li>
+        <router-link to="/buttons-and-inputs">Buttons and Inputs</router-link>
+      </li>
       <li><router-link to="/components">Components</router-link></li>
       <li><router-link to="/typography">Typography</router-link></li>
       <li><router-link to="/utility-classes">Utility classes</router-link></li>
@@ -19,9 +18,9 @@
     </ul>
     <div class="mobile">
       <div>
-        <a href="/">
+        <router-link to="/" class="no-deco">
           <h2 class="pb-3">Jelmers Easy Style System</h2>
-        </a>
+        </router-link>
       </div>
       <section class="flex-layout three-columns gap-3">
         <router-link to="/">Home</router-link>
@@ -36,8 +35,10 @@
   </nav>
 </template>
 <style scoped>
-a[class$="-active"] {
+a[class$='-active'] {
   text-decoration: underline;
 }
+.no-deco {
+  text-decoration: none;
+}
 </style>
-
